% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{add_rec_shade}
\alias{add_rec_shade}
\title{cSM Eco ggplot2 Theme}
\usage{
add_rec_shade(st_date, ed_date, shade_color = "darkgray")
}
\arguments{
\item{st_date}{Start date of desired data}

\item{ed_date}{End date of desired data}

\item{shade_color}{Color of recession bars}
}
\value{
A geom_rect object
}
\description{
This function adds recession bars to ggplot using daily FRED recession indicator data
}
