% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_suffix.R
\name{add_table_suffix}
\alias{add_table_suffix}
\title{Add suffix to columns names}
\usage{
add_table_suffix(data, suffix, except)
}
\arguments{
\item{data}{(data.frame) a data.frame}

\item{suffix}{(character) a character vector with a single element}

\item{except}{(character) a character vector whose elements represent
the columns names of (data) the suffix will
not be added to.}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble whose columns names
have a suffix equal to
(suffix) but those included
in (except).
}
\description{
Given a data.frame and a string, the functions add to each column's
name of the data the string as suffix.
}
\details{
The functions has been implemented to as helper to create
a unique data.frame with all the forms of a single study.
The functions ensures that all the fields will have a
unique and it avoids to have duplicated names, which can
likely occur for example if more follow-up dates of events
are recorded.
}
\examples{

set.seed(1)

df <- tibble::tibble(
    center = rep(c("center_1", "center_2"), each = 3L),
    id = c(
        "center_1_id_1", "center_1_id_2", "center_1_id_3",
        "center_2_id_1", "center_2_id_2", "center_2_id_3"
    ),
    age = rnorm(6, 60, 15),
    sex = sample(c("female", "male"), 6, TRUE)
)

suffix <- "demo"
except <- c("id", "center")

add_table_suffix(df, suffix, except)

}
