% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Advanced_Distance_function_A_B.R
\name{Advanced_Distance_function_A_B}
\alias{Advanced_Distance_function_A_B}
\title{Calculates distance between two cell types for a given image}
\usage{
Advanced_Distance_function_A_B(DATA, cell_A, cell_B)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell phenotype labels from a single image.}

\item{cell_A}{A character value indicating the cell phenotype label of the Cell of Origin.}

\item{cell_B}{A character value indicating the cell phenotype label of the Target cell.}
}
\value{
A tibble containing the distance matrix. Rows represent Cells of Origin and columns represent Target cells.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}}
}
\keyword{Internal}
