% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Advanced_Distance_function_A_B_C_single.R
\name{Advanced_Distance_function_A_B_C_single}
\alias{Advanced_Distance_function_A_B_C_single}
\title{Calculates distance between two cell types and a cell of origin for a given image selecting a random cell of origin}
\usage{
Advanced_Distance_function_A_B_C_single(DATA, cell_A, cell_B, cell_C)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell phenotype labels from a single image.}

\item{cell_A}{A character value indicating the cell phenotype label of the Cell of Origin.}

\item{cell_B}{A character value indicating the cell phenotype label of the Target cell 1.}

\item{cell_C}{A character value indicating the cell phenotype label of the Target cell 2.}
}
\value{
A list containing two tibbles. Tibble A_to_B contains the distance matrix between the single COO and Target cell 1. Tibble A_to_C contains the distance matrix between the single COO and Target cell 2
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Trio_Random_Distance_matrix_generator]{Trio_Random_Distance_matrix_generator()}}
}
\keyword{Internal}
