% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Advanced_Tumor_Stroma_identifier.R
\name{Advanced_Tumor_Stroma_identifier}
\alias{Advanced_Tumor_Stroma_identifier}
\title{Calculates tumor and stromal compartment based on concave hull}
\usage{
Advanced_Tumor_Stroma_identifier(
  DATA_Phenotypes,
  Index_phenotype,
  Filtering_Method,
  Accuracy = NULL,
  Min_cell_no = NULL,
  Distance_radius = NULL,
  Hull_ratio = NULL,
  Calculate_border = NULL,
  Dist_to_border = NULL,
  Image_preview = NULL,
  N_cores = 1
)
}
\arguments{
\item{DATA_Phenotypes}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Index_phenotype}{A character value indicating the cell phenotype to be used to calculate the tissue compartment.}

\item{Filtering_Method}{A character value indicating the filtering method type. Either "Tiling" or "DBSCAN".}

\item{Accuracy}{If Filtering_Method is 'Tiling': A numeric value indicating the size of the tile. Smaller sizes give more accurate result.}

\item{Min_cell_no}{If Filtering_Method is 'DBSCAN': A integer value indicating the minimum number of cells required to be a cluster.}

\item{Distance_radius}{If Filtering_Method is 'DBSCAN': A numeric value indicating the distance to be sampled.}

\item{Hull_ratio}{A numeric value indicating the hull ratio. Smaller values calculate more precise edge silhouettes at the cost of being more computationally demanding.}

\item{Calculate_border}{A logical value indicating if border to tissue estructure should be computed.}

\item{Dist_to_border}{A numeric value indicating the distance to the border to consider a cell to be in the border}

\item{Image_preview}{A character value indicating the name of the image to be used in the preview.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
Returns a list with a tibble with cell features including a column named 'Compartment' containing cell location and a tibble with tumor compartment area.
}
\description{
The function calculates tissue compartments (usually tumor and stroma) based on concave hull.
The border is calculated based on a user defined distance to the tissue structure. Before the concave hull is calculated a filtering method must be applied to remove non-spatially clustered cells.
The concave hull is calculated using th sf R package.
DBscan filtering is based on the dbscan::dbscan function.
}
\details{
The function results can be used to run the \code{\link[=Compartment_Phenotype_quantifier]{Compartment_Phenotype_quantifier()}} function.
}
\examples{
\dontrun{
Advanced_Tumor_Stroma_identifier(
    DATA_Phenotypes = CSM_Phenotypecell_test,
    Index_phenotype = "TUMOR",

    Filtering_Method = "DBSCAN",
    Min_cell_no = 10,
    Distance_radius = 100,

    Hull_ratio = 0.05,
    Calculate_border = TRUE,
    Dist_to_border = 10
)
}

}
\seealso{
\code{\link[=Compartment_Phenotype_quantifier]{Compartment_Phenotype_quantifier()}}
}
