% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_labels.R
\name{assign_labels}
\alias{assign_labels}
\title{Assign the labels to the elements of a character}
\usage{
assign_labels(x, y)
}
\arguments{
\item{x}{(character) a character vector.}

\item{y}{(character) a named character vector whose names must be
be contained in \code{x}.}
}
\value{
a (character) vector.
}
\description{
Given a character vector \code{x} and a named character vector \code{y} whose
names correspond to the elements contained in \code{x}, the function
switches the elements of \code{x} with the elements of \code{y}.
}
\details{
The function is intended to be used when the report of
data quality check is implemented. When some fields are
shown in the tables, it could be useful to report their
labels for ease of readibility.
}
\examples{

x <- c("a", "b", "c")
y <- c("1", "2", "3") \%>\%
    purrr::set_names(nm = c("a", "b", "c"))

assign_labels(x, y)

}
