% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot_generator.R
\name{Barplot_generator}
\alias{Barplot_generator}
\title{Generate a summary plot of phenotype composition across images.}
\usage{
Barplot_generator(DATA, Phenotypes_included, Ordering_phenotype)
}
\arguments{
\item{DATA}{A dataframe or tibble containing a summary of phenotypes by sample like the one obtained after running \code{\link[=Phenotype_quantifier]{Phenotype_quantifier()}}.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the plot.}

\item{Ordering_phenotype}{A character value indicating the phenotype that will guide barplot ordering.}
}
\value{
Generates a barplot of the phenotypes composition of every image in DATA.
}
\description{
After quantifying cell phenotypes in every image using \code{\link[=Phenotype_quantifier]{Phenotype_quantifier()}} function, the user can plot the results using this function.
}
\examples{
\dontrun{
#Generate cell count summary----------------------------------
Phenotypes_by_Sample <-
Phenotype_quantifier(
    DATA = CSM_Phenotypecell_test,
    Calculate_Density = FALSE
)

#Generate barplot with desired cell phenotype labels-----------
Barplot_generator(
    DATA = Phenotypes_by_Sample,
    Phenotypes_included = c("TUMOR", "CD8_GZMBpos", "CD8_GZMBneg"),
    Ordering_phenotype = "CD8_GZMBpos"
)
}

}
