% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot_Heterogeneity_generator.R
\name{Barplot_Heterogeneity_generator}
\alias{Barplot_Heterogeneity_generator}
\title{Generate a summary plot of global heterogeneity across images.}
\usage{
Barplot_Heterogeneity_generator(DATA, Metric)
}
\arguments{
\item{DATA}{A dataframe or tibble containing a summary of global heterogeneity by sample like the one obtained after running \code{\link[=Global_heterogeneity_calculator]{Global_heterogeneity_calculator()}}.}

\item{Metric}{A character vector indicating the metric that will be displayed in the plot.}
}
\value{
Generates a barplot of the heterogeneity for every image in DATA.
}
\description{
After quantifying heterogeneity for every image using \code{\link[=Global_heterogeneity_calculator]{Global_heterogeneity_calculator()}}, the user can plot the results using this function.
}
\examples{
\dontrun{
#Calculate the global heterogeneity by sample----------------------------
Global_Heterogeneity_by_sample <-
 Global_heterogeneity_calculator(
   DATA = CSM_Phenotypecell_test,
   Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype)
)

#Generate the barplot----------------------------
Barplot_Heterogeneity_generator(
   DATA = Global_Heterogeneity_by_sample,
   Metric = "Gini"
)
}

}
