% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barrier_effect_calculator.R
\name{Barrier_effect_calculator}
\alias{Barrier_effect_calculator}
\title{Calculates which cellular or pixel elements are in the path from COO to target cells}
\usage{
Barrier_effect_calculator(
  N_cores = 1,
  DATA = NULL,
  Cell_Of_Origin = NULL,
  Target_Cell = NULL,
  Barrier_Cells = NULL,
  Image_parameter_list = NULL,
  Pixel_distance_ratio = NULL,
  Distance_sampled = NULL,
  Perform_edge_correction = FALSE,
  Hull_ratio = NULL,
  Polygon_angle = 10,
  Target_cell_Angle_tolerance = 1
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Cell_Of_Origin}{A character value indicating the cell phenotype label of the Cell of Origin.}

\item{Target_Cell}{A character value indicating the cell phenotype label of the Target cell.}

\item{Barrier_Cells}{A character vector containing the names of the cell phenotypes that will be measured.}

\item{Image_parameter_list}{A list of lists containing information of images being used for pixel quantification (see details).}

\item{Pixel_distance_ratio}{(OPTIONAL if pixel quantification is required) A numeric value indicating the ratio between pixel size / cell coordinates. Use this argument when cell coordinates have been transformed to a distance unit like microns.}

\item{Distance_sampled}{The maximum distance allowed to consider a target cell in proximity to a COO.}

\item{Perform_edge_correction}{A logical value indicating if COO close to the edge of the tissue should be removed from the analysis. If TRUE, COO that are less than the Distance_sampled to the edge will be removed.}

\item{Hull_ratio}{If edge correction needs to be performed, a numeric value indicating the hull ratio. Smaller values calculate more precise edge silhouettes at the cost of being more computationally demanding.}

\item{Polygon_angle}{A numeric value indicating the angle used to triangulate COO-target space. Higher values will result in coarser space polygons.}

\item{Target_cell_Angle_tolerance}{(OPTIONAL) If not NULL, a numeric value indicating the amount of angular distance required to keep two target cells. If below the tolerance value, only the closest target cell to the COO will be kept in the analysis.}
}
\value{
A tibble containing information of barrier cells and pixels per COO analyzed.
}
\description{
Given a Cell of Origin (COO) and its closest target cells, the function will compute the space polygon between them. Then cells or pixels inside this polygon are quantified.
}
\details{
If image pixels need to be quantified the Image_parameter_list argument must be provided. Image_parameter_list must be a named list of lists.
Each named element must be a list containing the following items:
\itemize{
\item{Directory: Character specifying the path to the folder where images are present.}
\item{Channel: Either NULL (for single channel images) or a integer value indicating the channel number to be analyzed.}
\item{Image_rotate: Either NULL or a integer value indicating the degrees of rotation of the image. If NULL no rotation is applied.}
\item{Image_x_flip: A logical value indicating if X image flip should be performed.}
\item{Image_y_flip: A logical value indicating if X image flip should be performed.}
}

The function returns the absolute cell and pixel counts within the barrier. The function also computes the ratios normalized by
the total number of target cells studied, the area of the barrier polygon and the total distance of the COO to the target cells studied.
}
\examples{
\dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
          function(Image){
            EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
                               file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
          })

#Generate the Image_parameter_list------------------------------
Image_Parameters <- list(
 PDL1 = list(
   Directory = Input_Dir,
   Channel = 2,
   Image_rotate = NULL,
   Image_x_flip = FALSE,
   Image_y_flip = TRUE),
 GZMB = list(
   Directory = Input_Dir,
   Channel = 3,
   Image_rotate = NULL,
   Image_x_flip = FALSE,
   Image_y_flip = TRUE))

#Compute the cells and pixels in the barrier------------------------------
Barrier_DATA <- Barrier_effect_calculator(
 N_cores = 1,

 DATA = CSM_Phenotypecell_test,


 Cell_Of_Origin = "CD8_GZMBneg",
 Target_Cell = "TUMOR",
 Barrier_Cell = c("OTHER", "CD8_GZMBpos"),
 Image_parameter_list = Image_Parameters,

 Distance_sampled = 30,

 Perform_edge_correction = FALSE,

 Polygon_angle = 30,
 Target_cell_Angle_tolerance = 30
)

#Summarize the results by sample------------------------------
Barrier_effect_analyzer(
 DATA = Barrier_DATA,
 Analysis_type = "Area"
 )

#Plot the results------------------------------
Barrier_graph_maker(
  DATA_Phenotypes = CSM_Phenotypecell_test,
  Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype),

  Barrier_DATA = Barrier_DATA,
  Image_name = "ABCM22001_B14_MiniCrop.tif",

  Color_by = "GZMB_PixelValueSum_Area_Ratio",
  Point_size = 2,

  Image_directory = dir(Input_Dir, full.names = TRUE)[[2]],
  Pixel_distance_ratio = NULL,
  Channel_to_display = 3,
  Image_rotate = 0,
  Image_x_flip = FALSE,
  Image_y_flip = TRUE
)
}

}
\seealso{
\code{\link[=Barrier_effect_analyzer]{Barrier_effect_analyzer()}}, \code{\link[=Barrier_graph_maker]{Barrier_graph_maker()}}
}
