% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_check.R
\name{cat_check}
\alias{cat_check}
\title{Perform Categorical Check}
\usage{
cat_check(data, test, title = "")
}
\arguments{
\item{data}{A data frame with the site name/number (can be string or numeric) in the first column and any categorical variables to check in the following columns.}

\item{test}{'Chisq' or 'Fisher', choice of test method}

\item{title}{This provides the prefix of the title for table. Default is empty.}
}
\value{
table with categorical check outputs
}
\description{
The function is to investigate categorical variables. It compares the proportion of patients in each level in a site to the proportions found in all other sites.
}
\examples{
\dontrun{
cat_check(data,test,title='')
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

