% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_anomal_values.R
\name{categorical_anomal_values}
\alias{categorical_anomal_values}
\title{Check categorical impossible values}
\usage{
categorical_anomal_values(x, values)
}
\arguments{
\item{x}{(character) a character vector.}

\item{values}{(character) a character vector with the allowed values
for \code{x}.}
}
\value{
(character) a character vector with all the elements of \code{x}
that are different from \code{values}. If no values a
\code{NA_character_} is returned.
}
\description{
Check if a categorical variable takes values that are different from
these allowed by the form of Redcap. The function returns a
logical that identifies which elements of the variable assume an
impossible.
}
\examples{

gender <- c("male", "female", "male", "male", "female", "x")
values <- c("male", "female")
csm:::categorical_anomal_values(x = gender, values = values)

}
