% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_phenotyper.R
\name{CELESTA_phenotyper}
\alias{CELESTA_phenotyper}
\title{Assigns phenotype labels using CELESTA algorithm}
\usage{
CELESTA_phenotyper(
  DATA,
  Template_path,
  Alternative_CSV_locale = FALSE,
  N_cores = 1,
  Apply_filters = FALSE,
  high_marker_threshold = NULL,
  low_marker_threshold = NULL,
  max_iteration = 10,
  cell_change_threshold = 0.01
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Template_path}{A character indicating the path to the CSV file containing the fulfilled template.}

\item{Alternative_CSV_locale}{A logical value indicating if the CSV has an alternative locale format (like a European vs USA formatting).}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Apply_filters}{A logical value indicating if cells should be pre-filtered before running the algorithm.}

\item{high_marker_threshold}{A numeric value indicating the max quantile allowed. Cells with expression above this quantile for all markers will not be phenotyped.}

\item{low_marker_threshold}{A numeric value indicating the min quantile allowed. Cells with expression below this quantile for all markers will not be phenotyped.}

\item{max_iteration}{An integer. Number of iterations to run the GMM algorithm used in CELESTA}

\item{cell_change_threshold}{A numeric value that indicates when the iterative cell-type assignment should stop. The default value is 0.01,
which means that if the percentage of additional assigned cells is smaller than 1\% of the unassigned cells, then cell-type assignment will stop.
The recommended range is 0.01 - 0.05. Note that the higher the cell change threshold, the more cells are left unassigned.}
}
\value{
Returns a tibble with cell features and a column named 'Phenotype' containing cell labels.
}
\description{
The function runs the CELESTA algorithm using cell feature data and a template indicating expected cell phenotypes.
}
\examples{
\dontrun{
CELESTA_phenotyper(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Template_path = "Completed_template_path",
  Alternative_CSV_locale = TRUE,
  N_cores = 2,
  Apply_filters = TRUE,
  high_marker_threshold = 0.90,
  low_marker_threshold = 0.10,
  max_iteration = 10,
  cell_change_threshold = 0.01
)
}

}
\seealso{
\code{\link[=CELESTA_template_generator]{CELESTA_template_generator()}}. For more information on how to fulfill the template please visit https://github.com/plevritis-lab/CELESTA.
}
