% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CELESTA_template_generator.R
\name{CELESTA_template_generator}
\alias{CELESTA_template_generator}
\title{Generates a CELESTA cell phenotype template}
\usage{
CELESTA_template_generator(
  DATA = NULL,
  Markers_to_keep = NULL,
  Template_name = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Markers_to_keep}{A character vector indicating the names of the columns to be kept in the template.}

\item{Template_name}{The name of the file that will be written.}
}
\value{
Creates a CSV file containing the template in the working directory.
}
\description{
In order to run CELESTA, a template indicating expected cell phenotypes present in the tissue must be prepared. This function generates the template that can be modified manually by the user.
}
\seealso{
\code{\link[=CELESTA_phenotyper]{CELESTA_phenotyper()}}. For more information on how to fullfill the template please visit https://github.com/plevritis-lab/CELESTA.
}
