% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_functional_assessment.R
\name{Cell_functional_assessment}
\alias{Cell_functional_assessment}
\title{Analyzes the expression of features within a cell category}
\usage{
Cell_functional_assessment(
  DATA = NULL,
  Target_Variable = NULL,
  Targets_Included = NULL,
  DATA_Aside = NULL,
  Threshold_functional_Markers = NULL,
  Levels = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Target_Variable}{A character indicating the name of the column containing cell categories (for example cell phenotypes).}

\item{Targets_Included}{A character vector indicating which cell categories should be analyzed (for example a specific cell phenotype to be included).}

\item{DATA_Aside}{A dataframe or tibble containing the functional features to be analyzed.}

\item{Threshold_functional_Markers}{A logical value indicating if features should be thresholded before analysis.}

\item{Levels}{If thresholding is required, a integer value indicating the number of desired levels to be calculated for multilevel thresholding (see details).}
}
\value{
A list containing cell feature information and a summary of feature expression in target cells. Also generates summary plots.
}
\description{
The expression level of various features that have been set aside can be analyzed within cell categories (either a cell phenotype or neighborhood label).
These features can be set aside with \code{\link[=Data_set_aside]{Data_set_aside()}} function.
}
\details{
Multilevel thresholding is calculated using imagerExtra::ThresholdML function
}
