% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_image_mask_profiler.R
\name{Cell_image_mask_profiler}
\alias{Cell_image_mask_profiler}
\title{Computes cell features based on image and cell segmentation mask}
\usage{
Cell_image_mask_profiler(
  Image = NULL,
  Mask = NULL,
  Ordered_Channels = NULL,
  Compute_basic_features = TRUE,
  quantiles_to_calculate = c(0.25, 0.5, 0.75),
  Compute_texture_features = FALSE,
  Texture_pixel_distance = 1,
  SubCellular_compartment_list = NULL
)
}
\arguments{
\item{Image}{An EBImage object containing the Image.}

\item{Mask}{An EBImage object containing the cell segmentation mask. Each element in the mask must be an integer value.}

\item{Ordered_Channels}{Character vector specifying Image channel names in their exact order.}

\item{Compute_basic_features}{A logical value indicating if basic features (mean, sd, quantiles) should be computed.}

\item{quantiles_to_calculate}{Numeric vector specifying the expression quantiles calculated for every cell and channel.}

\item{Compute_texture_features}{A logical value indicating if texture features should be computed for all the markers measured.}

\item{Texture_pixel_distance}{If Compute_texture_features is TRUE, de pixel distance taken into account to compute the GLCM and texture features.}

\item{SubCellular_compartment_list}{A named list containing the cell masks for subcellular expression computation.}
}
\value{
A tibble containing cell data.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Cell_feature_extractor]{Cell_feature_extractor()}}
}
\keyword{Internal}
