% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_image_plot_generator.R
\name{Cell_image_plot_generator}
\alias{Cell_image_plot_generator}
\title{Plot an image with a cell overlay.}
\usage{
Cell_image_plot_generator(
  Image_directory = NULL,
  Channel_to_display = NULL,
  Image_rotate = NULL,
  Image_x_flip = FALSE,
  Image_y_flip = FALSE,
  Gamma_level = 0,
  Equalize = FALSE,
  Black_level = 0,
  White_level = 100,
  DATA = NULL,
  Image_name = NULL,
  Color_by = NULL,
  Point_size = 1,
  Pixel_distance_ratio = NULL
)
}
\arguments{
\item{Image_directory}{Character string specifying the path to the image file that needs to be plotted.}

\item{Channel_to_display}{A integer value indicating the Channel index to be displayed (1 for single channel images).}

\item{Image_rotate}{(OPTIONAL) A integer value indicating the degrees of rotation of the image.}

\item{Image_x_flip}{A logical value indicating if X image flip should be performed.}

\item{Image_y_flip}{A logical value indicating if X image flip should be performed.}

\item{Gamma_level}{A numeric value between -3 and +3 indicating the gamma level to apply to the image.}

\item{Equalize}{A logical value indicating if the image should be equalized.}

\item{Black_level}{A integer value between 0 and 100 indicating the minimum pixel intensity.}

\item{White_level}{A integer value between 0 and 100 indicating the maximum pixel intensity.}

\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Image_name}{A character value indicating the image to be plotted.}

\item{Color_by}{(OPTIONAL) A character value indicating the column name in DATA that should be used to color cell points.}

\item{Point_size}{A numeric value indicating the size of the points.}

\item{Pixel_distance_ratio}{(OPTIONAL) A numeric value indicating the ratio between pixel size / cell coordinates. Use this argument when cell coordinates have been transformed to a distance unit like microns.}
}
\value{
Generates a plot with the image and the cells overlay.
}
\description{
The function generates a plot with an image in the background and cells overlayed. Cell information must have been obtained from the image in order to match X Y coordinates with image coordinates.
}
\seealso{
\code{\link[=Cell_plotter]{Cell_plotter()}}
}
