% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_plotter.R
\name{Cell_plotter}
\alias{Cell_plotter}
\title{Plot the cells of a single image}
\usage{
Cell_plotter(DATA = NULL, Image_name = NULL, Phenotypes_included = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Image_name}{A character value indicating the image to be plotted.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the plot.}
}
\value{
Generates a plot of the image cells.
}
\description{
The function generates a plot of the cells of a single image. It is similar to \code{\link[=Cell_image_plot_generator]{Cell_image_plot_generator()}} but the original image is not drawn in the background.
}
\examples{
\dontrun{
Cell_plotter(
   DATA = CSM_Phenotypecell_test,
   Image_name = "ABCM22001_B14_MiniCrop.tif",
   Phenotypes_included = c("TUMOR", "CD8_GZMBpos", "CD8_GZMBneg")
)
}


}
\seealso{
\code{\link[=Cell_image_plot_generator]{Cell_image_plot_generator()}}
}
