% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_segmentator_quantificator.R
\name{Cell_segmentator_quantificator}
\alias{Cell_segmentator_quantificator}
\title{Performs cell segmentation and obtains cell feature matrix}
\usage{
Cell_segmentator_quantificator(
  Directory,
  Parameter_list = NULL,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL,
  N_cores = 1,
  quantiles_to_calculate = c(0.25, 0.5, 0.75),
  Compute_texture_features = FALSE,
  Texture_pixel_distance = 1,
  Save_cell_mask = FALSE,
  Output_directory = NULL,
  Nuclear_marker = NULL,
  Cell_body_method = NULL,
  Min_pixel = NULL,
  Smooth_amount = NULL,
  Normalization = NULL,
  Watershed_type = NULL,
  Tolerance_value = NULL,
  Neighborhood_distance = NULL,
  Disc_size = NULL,
  Tissue_mask_markers = NULL,
  Perform_PCA = FALSE,
  Perform_nuclear_channel_processing = NULL,
  Black_level = NULL,
  White_level = NULL,
  Gamma_level = NULL,
  Equalize = NULL,
  Opening_kernel_size = NULL,
  Closing_kernel_size = NULL
)
}
\arguments{
\item{Directory}{Character specifying the path to the folder where images to be segmented are present.}

\item{Parameter_list}{List obtained using the \code{\link[=Watershed_tester_app]{Watershed_tester_app()}} containing segmentation parameters.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{Channels_to_keep}{Character vector specifying image channels to be kept in the analysis (must be a subset of Ordered_Channels).}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{quantiles_to_calculate}{Numeric vector specifying the expression quantiles calculated for every cell and channel. If NULL, no quantiles are computed.}

\item{Compute_texture_features}{A logical value indicating if texture features should be computed for all the markers measured.}

\item{Texture_pixel_distance}{If Compute_texture_features is TRUE, de pixel distance taken into account to compute the GLCM and texture features.}

\item{Save_cell_mask}{A logical value indicating if images containing cell masks should be saved or not.}

\item{Output_directory}{If Save_cell_mask is TRUE, a character value indicating the path to the empty folder where output images should be written.}

\item{Nuclear_marker}{(Used if Parameters_list is NULL) Character vector of markers corresponding to the nuclei.}

\item{Cell_body_method}{(Used if Parameters_list is NULL) Method of cytoplasm identification. Can be 'none', 'dilate' or 'discModel'.}

\item{Min_pixel}{(Used if Parameters_list is NULL) Integer value specifying the minimum pixels for an object to be recognized as a cell and not noise.}

\item{Smooth_amount}{(Used if Parameters_list is NULL) Numeric value specifying the amount of Gaussian smoothing to be applied to the image.}

\item{Normalization}{(Used if Parameters_list is NULL) Single value or vector specifying the transformations from "sqrt", "asinh", "norm99", "maxThresh" and "tissueMask.}

\item{Watershed_type}{(Used if Parameters_list is NULL) Method used to perform watersheding. Accepted values: "intensity", "distance" or "combine".}

\item{Tolerance_value}{(Used if Parameters_list is NULL) Numeric value specifying the minimum height of the object in the units of image intensity between its highest point (seed) and the point where it contacts another object (MAY BE NULL).}

\item{Neighborhood_distance}{(Used if Parameters_list is NULL) Radius of the neighborhood in pixels for the detection of neighboring objects. Higher value smooths out small objects.}

\item{Disc_size}{(Used if Parameters_list is NULL) The size of dilation around nuclei to create cell disc or capture cytoplasm.}

\item{Tissue_mask_markers}{(Used if Parameters_list is NULL) A vector specifying the channels to be used to create the tissue mask if specified in transformation.}

\item{Perform_PCA}{(Used if Parameters_list is NULL) A logical value specifying wether to run PCA on aggregated nucleus markers in order to detect the cellular nucclei.}

\item{Perform_nuclear_channel_processing}{(Used if Parameters_list is NULL) A logical value specifying if nuclear channel pre-processing should be performed.}

\item{Black_level}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Numeric indicating the \% below the max intensity level to be set to absolute black.}

\item{White_level}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Numeric indicating the \% above the max intensity level to be set to absolute white.}

\item{Gamma_level}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Numeric value between -3 and +3 to indicate channel gamma.}

\item{Equalize}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) A logical value specifying if channel should be equalized.}

\item{Opening_kernel_size}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Opening kernel size (set to 1 if no opening is required).}

\item{Closing_kernel_size}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Closing kernel size (set to 1 if no closing is required).}
}
\value{
A tibble containing cell feature data including cell X Y coordinates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{Cell_segmentator_quantificator()} performs cell segmentation according to user defined parameters.
Parameters can then be obtained using \code{\link[=Watershed_tester_app]{Watershed_tester_app()}}.
}
\examples{
\dontrun{
#Create temporary input directory----------------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Check a segmentation parameters list obtained using the dedicated function----------
print(CSM_SegmentParams_test)

#Run the cell segmentation and feature extraction process------------------
Cell_segmentator_quantificator(
  Directory = Input_Dir,
  Parameter_list = CSM_SegmentParams_test,
  N_cores = 1,
  quantiles_to_calculate = c(0.05, 0.25, 0.5, 0.75, 0.95)
)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
}


}
\seealso{
\code{\link[=Watershed_tester_app]{Watershed_tester_app()}}
}
