% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_to_Cell_graph_maker.R
\name{Cell_to_Cell_graph_maker}
\alias{Cell_to_Cell_graph_maker}
\title{Generates a plot of the spatial interaction for a single image}
\usage{
Cell_to_Cell_graph_maker(
  Image_name = NULL,
  DATA_Phenotypes = NULL,
  Strategy = NULL,
  DATA_Distances = NULL,
  DATA_Cumulative = NULL,
  Radius = NULL
)
}
\arguments{
\item{Image_name}{A character value indicating the image to be plotted.}

\item{DATA_Phenotypes}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Strategy}{A character value indicating the plot that will be generated. One of the following: "Min_Distance", "Average_Distance", "Max_Distance" or "Cells_in_Radius".}

\item{DATA_Distances}{If Strategy is "Min_Distance", "Average_Distance" or "Max_Distance",  a distance matrix list generated using \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}} function.}

\item{DATA_Cumulative}{If Strategy is Cells_in_Radius, a cumulative interaction matrix list generated using \code{\link[=Cumulative_Interaction_generator]{Cumulative_Interaction_generator()}} function.}

\item{Radius}{If strategy is Cells_in_Radius, a numeric value indicating the radius to be plotted. It must have been computed during the generation of the cumulative interaction matrix.}
}
\value{
A graph summarizing spatial interactions for the indicated image.
}
\description{
The function generates a summary plot of the spatial interaction between two cell types.
It can graph results for average, min and max distance as well as cell in radius analyses.
}
\examples{
\dontrun{
#Generate distance matrix----------------------------------------------------
DATA_Distances <-
Distance_matrix_generator(
    N_cores = 1,
    DATA = CSM_Phenotypecell_test,
    Cell_Of_Origin = "CD8_GZMBneg",
    Target_Cell = "TUMOR",
    Allow_Cero_Distance = FALSE,
    Perform_edge_correction = FALSE
)

#Generate cumulative interactions (must contain the actual radius distance)--
DATA_Cumulative <-
Cumulative_Interaction_generator(
   N_cores = 1,
   DATA = DATA_Distances,
   Start_from = 25,
   Stop_at = 100,
   Sampling_frequency = 25
)

#Graph min distance to target-------------------------------------------------
Cell_to_Cell_graph_maker(
   Image_name = "ABCM22001_B14_MiniCrop.tif",
   DATA_Phenotypes = CSM_Phenotypecell_test,
   Strategy = "Min_Distance",
   DATA_Distances = DATA_Distances
)
Cell_to_Cell_graph_maker(
   Image_name = "ABCM22001_B14_MiniCrop.tif",
   DATA_Phenotypes = CSM_Phenotypecell_test,
   Strategy = "Cells_in_Radius",
   DATA_Cumulative = DATA_Cumulative,
   Radius = 50
)
}

}
\seealso{
\code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}}, \code{\link[=Cumulative_Interaction_generator]{Cumulative_Interaction_generator()}}, \code{\link[=Distance_analyzer]{Distance_analyzer()}}, \code{\link[=Cells_in_Radius_analyzer]{Cells_in_Radius_analyzer()}}
}
