% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_to_pixel_distance_calculator.R
\name{Cell_to_pixel_distance_calculator}
\alias{Cell_to_pixel_distance_calculator}
\title{Calculate the distance from a cell to the closest positive pixel.}
\usage{
Cell_to_pixel_distance_calculator(
  N_cores = 1,
  Directory,
  Image_rotate = FALSE,
  Image_x_flip = FALSE,
  Image_y_flip = FALSE,
  DATA,
  Phenotypes_included,
  Pixel_distance_ratio = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Directory}{Character string specifying the path to the folder containing thresholded images.}

\item{Image_rotate}{(OPTIONAL) A integer value indicating the degrees of rotation of the image.}

\item{Image_x_flip}{A logical value indicating if X image flip should be performed.}

\item{Image_y_flip}{A logical value indicating if X image flip should be performed.}

\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}

\item{Pixel_distance_ratio}{(OPTIONAL) A numeric value indicating the ratio between pixel size / cell coordinates. Use this argument when cell coordinates have been transformed to a distance unit like microns.}
}
\value{
A tibble containing Cell feature information along with the distance to the closest positive pixel.
}
\description{
Given a thresholded image (either binary or multilevel) and the cell feature information matrix, the function will compute the distance from cell to the closest non-zero value pixels.
This can be used to analyze the spatial interaction of cells with extra-cellular elements in the image. Images can be thresholded using the \code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}} function.
}
\examples{
\dontrun{
Cell_to_pixel_distance_calculator(
    N_cores = 1,
    Directory = "Path/To/Thresholded_images",
    Image_rotate = 90,
    Image_x_flip = FALSE,
    Image_y_flip = TRUE,
    DATA = CSM_Phenotypecell_test,
    Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype),
    Pixel_distance_ratio = NULL
 )
}

}
\seealso{
\code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}}
}
