% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells_in_Radius_analyzer.R
\name{Cells_in_Radius_analyzer}
\alias{Cells_in_Radius_analyzer}
\title{Calculates the average target cells within a distance from Cell of Origin for every image}
\usage{
Cells_in_Radius_analyzer(
  DATA = NULL,
  DATA_RANDOM = NULL,
  Radius = NULL,
  Include_Random = NULL,
  By_Sample_Random = NULL
)
}
\arguments{
\item{DATA}{A list containing cumulative interaction data. It can be computed using \code{\link[=Cumulative_Interaction_generator]{Cumulative_Interaction_generator()}} function.}

\item{DATA_RANDOM}{(OPTIONAL) A list containing random cumulative interaction data. It can be computed using \code{\link[=Cumulative_Interaction_generator]{Cumulative_Interaction_generator()}} function.}

\item{Radius}{A numeric value indicating the size of the radius to perform analysis. It must have been computed during the generation of the cumulative interaction matrix.}

\item{Include_Random}{A logical value indicating if the Random background should be used.}

\item{By_Sample_Random}{If a random background needs to be included, a logical value indicating if the random background should be calculated by sample or by experiment (see details).}
}
\value{
A tibble containing a summary by sample of spatial interactions.
}
\description{
The function calculates the average number of target cells within a user defined distance for every image in a dataset.
}
\details{
If By_Sample_Random is TRUE, for every sample, a specific random background will be calculated to compute expected spatial interaction metrics. If By_Sample_Random is FALSE, the random background will be calculated using all image information. The same random background will be used for all images.
}
\seealso{
\code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}}, \code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}}, \code{\link[=Cumulative_Interaction_generator]{Cumulative_Interaction_generator()}}
}
