% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Channel_deconvolution_function.R
\name{Channel_deconvolution_function}
\alias{Channel_deconvolution_function}
\title{Thresholds RGB channels of a color image and performs pre and post-processing steps}
\usage{
Channel_deconvolution_function(Image = NULL, Parameters = NULL)
}
\arguments{
\item{Image}{A EBImage object.}

\item{Parameters}{A list of image pre and post-processing}
}
\value{
The function returns a greyscale EBImage object
}
\description{
Intended for internal use only
}
\keyword{Internal}
