% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_data.R
\name{check_missing}
\alias{check_missing}
\title{Returns the fields with missing information for each patient}
\usage{
check_missing(
  nested_tables,
  meta_data,
  tab,
  fields_names,
  branching_logic,
  redcap_info = c("record_id", "center")
)
}
\arguments{
\item{nested_tables}{(tibble) the output from \link{nest_tables}}

\item{meta_data}{(tibble) the output from \link{tidy_extract} when
\code{type = 'meta'}}

\item{tab}{(character) the name of the sheet on which missing data
check must be performed.}

\item{fields_names}{(character) a character with the name of the
column of \code{meta_data} that contains
the names of fields.}

\item{branching_logic}{(character)) a character with the name of the
column of \code{meta_data} that the
branching logic associated to
each field.}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled.}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble that identifies the
missing fields in a given
form for each patient
}
\description{
The function returns the fields that are missing for each patient
in a given form.
}
\examples{
\dontrun{
  nested <- readr::read_rds(db_update_from_server())

  get_incomplete_sheets(nested, redcap_info = c("id", "center"))
}

}
