% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_data.R
\name{check_missing_data}
\alias{check_missing_data}
\title{Returns the fields with missing information for each patient}
\usage{
check_missing_data(
  nested_tables,
  meta_data,
  tab,
  fields_names,
  branching_logic,
  redcap_info = c("record_id", "center")
)
}
\arguments{
\item{nested_tables}{(tibble) the output from \link{nest_tables}}

\item{meta_data}{(tibble) the output from \link{tidy_extract} when
\code{type = 'meta'}}

\item{tab}{(character) the name of the sheet on which missing data
check must be performed.}

\item{fields_names}{(character) a character with the name of the
column of \code{meta_data} that contains
the names of fields.}

\item{branching_logic}{(character) a character with the name of the
column of \code{meta_data} that the
branching logic associated to
each field.}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled.}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble that identifies the
missing fields in a given
form for each patient
}
\description{
The function returns the fields that are missing for each patient
in a given form.
}
\examples{
nested_tab <- tibble::tibble(
    fields = c(rep("demo_clinical", 3L), rep("follow_up", 3L)),
    sheets = c(
        "demo", "risk_factors", "clinical", "discharge",
        "month_1", "year_1"
    ),
    tables = purrr::map(
        .x = seq_len(6),
        ~ tibble::tibble(
            id = rep(glue::glue("id_{1:5}"), 2L),
            center = rep(c("center_1", "center_2"), each = 5L),
            x = c(runif(8), rep(NA, 2L)),
            y = c(runif(9), NA),
            z = c(
                "no", "no", "yes", NA_character_, NA_character_,
                "no", "yes", "yes", "no", NA_character_
            )
        )
    )
)

meta_data <- tibble::tibble(
    field_name = rep(c("x", "y", "z"), 6L),
    sheet = rep(
        c(
            "demo", "risk_factors", "clinical", "discharge",
            "month_1", "year_1"
        ), each = 3L
    ),
    branching_logic = rep(
        c(
            NA_character_, NA_character_, NA_character_,
            NA_character_, "x", NA_character_,
            "z", NA_character_, NA_character_
        ), 2L
    )
)

check_missing_data(
  nested_tab, meta_data, "demo", "field_name",
  "branching_logic", redcap_info = c("id", "center")
)


}
