% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinical_Data_arrange_function.R
\name{Clinical_Data_arrange_function}
\alias{Clinical_Data_arrange_function}
\title{Formats image metadata}
\usage{
Clinical_Data_arrange_function(DATA, Subject_Names, Outcomes_to_keep)
}
\arguments{
\item{DATA}{A dataframe or tibble containing image metadata.}

\item{Subject_Names}{A character indicating the column name containing image name.}

\item{Outcomes_to_keep}{A character vector indicating the names of the columns to be kept in the analysis.}
}
\value{
Returns a tibble with image features.
}
\description{
\code{Clinical_Data_arrange_function()}must be run in order to format image metadata. These can be either numeric or strings. Time to event data must be coded in two separate features (one for event and other for time).
This function is similar to \code{\link[=Data_arrange_function]{Data_arrange_function()}}, without adequate formatting, image metadata will no be able to work with \code{\link[=Clinical_Data_analyzer]{Clinical_Data_analyzer()}}
}
\examples{
\dontrun{
Clinical_Data_arrange_function(
     DATA = CSM_ClinicalTMA_test,
     Subject_Names = "Sample",
     Outcomes_to_keep = c("AGE", "MMRP_status", "DEATH", "OS_m")
)
}

}
