% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustered_Tiled_Images_analyzer.R
\name{Clustered_Tiled_Images_analyzer}
\alias{Clustered_Tiled_Images_analyzer}
\title{Generates a summary of neighborhood analysis based on clustering of tiles.}
\usage{
Clustered_Tiled_Images_analyzer(
  Tiled_images = NULL,
  Perform_heterogeneity_analysis = NULL,
  Graph_Modularity_resolution = NULL
)
}
\arguments{
\item{Tiled_images}{A list containing neighborhood information generated using \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}} function.}

\item{Perform_heterogeneity_analysis}{A logical value indicating if heterogeneity analysis should be performed.}

\item{Graph_Modularity_resolution}{The resolution parameter for graph modularity analysis.}
}
\value{
Generates a tibble containing neighborhood summary by sample and optionally heterogeneity metrics.
}
\description{
The function returns neighborhood counts and percentages. Optionally the function may compute neighborhood heterogeneity metrics.
}
\details{
Shannon, Simpson,Inverse Simpson and Renyi entropy indexes are calculated using the vegan package.

Rao index is calculated using the picante::raoD function.

Gini index is calculated using the DescTools::Gini function.

Graph modularity is computed using the igraph::modularity function.
}
