% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustered_Tiled_Images_graphicator.R
\name{Clustered_Tiled_Images_graphicator}
\alias{Clustered_Tiled_Images_graphicator}
\title{Generates plots of tiled neighborhood identity}
\usage{
Clustered_Tiled_Images_graphicator(Tiled_images, Image_name)
}
\arguments{
\item{Tiled_images}{A list containing tiled images obtained using \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}.}

\item{Image_name}{A character value indicating the image to be plotted.}
}
\value{
A plot of the neighborhood identity of each tile.
}
\description{
The function generates plots of the neighborhood labels of each tile after running the \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}.
}
\examples{
\dontrun{
#Tile images with cell phenotype information---------------------------------
Tiled_Images <-
 Image_tiling_processing_function(
   N_cores = 1,
   DATA = CSM_Phenotypecell_test,
   Tile_width = 125,
   Tile_height = 125,
   Variables_to_keep = "Phenotype"
)

#Cluster cell composition by tile to find neighborhoods---------------------
Clustered_Tiled_Images <-
Tiled_Image_Clustering_function(
    Tiled_images = Tiled_Images,
    Minimum_cell_no_per_tile = 4,
    Minimum_valid_tiles_per_image = 4,
    Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype),

    Cluster_Data = "Cell_Density",

    Perform_Dimension_reduction = FALSE,
    Cluster_on_Reduced = FALSE,

   Strategy = "Consensus_Clustering",
   Max_N_Clusters = 5,
   Consensus_reps = 2,
   Consensus_p_Items = 1,
   Consensus_Cluster_Alg = "pam",
   Consensus_Distance = "euclidean",
   Consensus_Name = "Consensus_clustering_test"
)

#Graph the results---------------------------------------------------------
Clustered_Tiled_Images_graphicator(
    Tiled_images = Clustered_Tiled_Images,
    Image_name = "ABCM22001_B14_MiniCrop.tif"
)
}


}
