% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustered_Tiled_Images_renamer.R
\name{Clustered_Tiled_Images_renamer}
\alias{Clustered_Tiled_Images_renamer}
\title{Modifies neighborhood labels}
\usage{
Clustered_Tiled_Images_renamer(Tiled_images = NULL, New_names = NULL)
}
\arguments{
\item{New_names}{A character vector indicating the new names. The length must be equal to the number of unique neighborhood labels.}

\item{DATA}{An object containing neighborhood labels generated using \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}} function.}
}
\value{
Returns a list with tiles per image with the modified labels.
}
\description{
The function allows the user to modify neighborhood labels. The function works with any object created using the \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}} function.
}
\seealso{
\code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}
}
