% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_deconvolution_App_launcher.R
\name{Color_deconvolution_App_launcher}
\alias{Color_deconvolution_App_launcher}
\title{Launches a shiny APP to explore color deconvolution parameters}
\usage{
Color_deconvolution_App_launcher(Directory, Max_Gb_cache = 10)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where RGB color images are present.}

\item{Max_Gb_cache}{A single numeric value indicating the memory size of the image cache (see details). 10 Gb by default.}
}
\description{
\code{Color_deconvolution_App_launcher()} launches an APP to interactively explore color deconvolution parameters.
Parameters can then be used to feed the \code{\link[=Image_deconvolution_function]{Image_deconvolution_function()}}.
}
\details{
In order to deal with large images and speed up image toggling, the shiny APP works with a memoised version of the \code{\link[=Smart_image_importer]{Smart_image_importer()}} function. Loaded images will
be stored in a temporary cache until the APP is closed or the cache reaches it's limit.

The control panel in the right controls the image display settings. Channel name box allows the user to specify the name of the channel to be extracted.
The "+Add" button adds the channel extraction parameters to the parameter list.
The "-Remove" button removes any channel parameters that match the "Channel name" box.

Initial Pre-processing allows the initial modification of the image to improve results. The "colors" option allows the user to limit the maximum amount of different colors in the image.

Color targeting allows different options:
\itemize{
\item{Color: The user can select the target color by it's HEX code.}
\item{RED GREEN BLUE: The user can select the target color by providing 8-bit RGB values.}
\item{Upper right panel: The user can select the target color by clicking on the desired target pixel color value.}
}

Tolerance values are key to extract the color adequately:
\itemize{
\item{Individual tolerance: Allows the user to specify the flexibility of individual color channel extraction.}
\item{End tolerance: After the individual RGB channels have been extracted and merged, individual tolerance regulates flexibility of pixel values with respect to the color target.}
}

Final processing stablishes how the image is going to be modified after color channel extraction.

"Download Parameters" button saves parameters in the Global Environment.

Image Panels:
\itemize{
\item{Upper left: Image being color processed. Use it to zoom in.}
\item{Upper right: Pixel selector. Initial Pre-processing steps will be applied. Use it to select target color.}
\item{Lower left: Initial color extracted according to active parameters.}
\item{Lower right: Final channel result after Final processing steps.}
}
}
\examples{
\dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniHE_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniHE_test)[Image]))
})

#Launch the app------------------------------------------------------------
Color_deconvolution_App_launcher(Directory = Input_Dir)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)

}


}
