% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colored_print.R
\name{Colored_print}
\alias{Colored_print}
\title{Prints text in color based in the cat function}
\usage{
Colored_print(text, color = "green")
}
\arguments{
\item{text}{The text being printed.}

\item{color}{One of the following: "black", "red", "green", "yellow", "blue", "magenta", "cyan", "white".}
}
\value{
the text with the formatted color.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}}, \code{\link[=Cell_in_edge_remover]{Cell_in_edge_remover()}}, \code{\link[=Cell_to_pixel_distance_calculator]{Cell_to_pixel_distance_calculator()}}, \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}}, \code{\link[=Gcross_calculator]{Gcross_calculator()}}, \code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}},
\code{\link[=Structure_morphometry_calculator]{Structure_morphometry_calculator()}}, \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}, \code{\link[=Trio_Distance_matrix_generator]{Trio_Distance_matrix_generator()}}, \code{\link[=Trio_Random_Distance_matrix_generator]{Trio_Random_Distance_matrix_generator()}}

This code is based on Joseph Crispell's answer to the stack overflow forum question
https://stackoverflow.com/questions/10802806/is-there-a-way-to-output-text-to-the-r-console-in-color
}
\keyword{Internal}
