% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartment_Phenotype_quantifier.R
\name{Compartment_Phenotype_quantifier}
\alias{Compartment_Phenotype_quantifier}
\title{Quantifies cell proportions and cell densities by compartment}
\usage{
Compartment_Phenotype_quantifier(
  DATA = NULL,
  Calculate_Density = NULL,
  DATA_Area = NULL
)
}
\arguments{
\item{DATA}{A list containing cell compartment location and compartment area generated with \code{\link[=Advanced_Tumor_Stroma_identifier]{Advanced_Tumor_Stroma_identifier()}}.}

\item{Calculate_Density}{A logical value indicating if density should be calculated. Currently density in border computation is not supported.}

\item{DATA_Area}{If density needs to be calculated, a tibble or dataframe containing total tissue size information. It may be generated using \code{\link[=Image_size_calculator]{Image_size_calculator()}}.}
}
\value{
Returns tibble with information of cell percentages and (optionally) densities by image and compartment.
}
\description{
The function calculates cell proportions and optionally cell densities in tumor and stromal compartments.
Compartment areas must be first computed using \code{\link[=Advanced_Tumor_Stroma_identifier]{Advanced_Tumor_Stroma_identifier()}} function.
}
\seealso{
\code{\link[=Compartment_Phenotype_quantifier]{Compartment_Phenotype_quantifier()}}, \code{\link[=Image_size_calculator]{Image_size_calculator()}}.
}
