% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conflicted_image_mask_modifier.R
\name{Conflicted_image_mask_modifier}
\alias{Conflicted_image_mask_modifier}
\title{Resolves conflicts between cell masks and generates erosion masks that are saved in an output directory}
\usage{
Conflicted_image_mask_modifier(
  Image_param_list = NULL,
  Conflict_type = NULL,
  First_Column = NULL,
  First_Row = NULL,
  Last_Column = NULL,
  Last_Row = NULL,
  Overlap_value = NULL,
  Resolution_approach = NULL,
  Priority_type = NULL,
  Min_cell_size = NULL,
  Mask_refinement_path = NULL
)
}
\arguments{
\item{Image_param_list}{A tibble containing information regarding images involved.}

\item{Conflict_type}{A character value indicating the type of conflict.}

\item{First_Column}{A logical value indicating if conflict area is located in the left-side column.}

\item{First_Row}{A logical value indicating if the conflict area is located in the top-side row.}

\item{Last_Column}{A logical value indicating if conflict area is located in the right-side column.}

\item{Last_Row}{A logical value indicating if the conflict area is located in the bottom-side row.}

\item{Overlap_value}{A integer value indicating the overlap between tiles.}

\item{Resolution_approach}{A character value indicating how conflict between overlapping tiles should be solved. One of the following: Remove_whole_cells, Erode_pixels (see details).}

\item{Priority_type}{The type of metric to decide which cells or pixels should prevail after conflict identification. Either Dist_to_edge, Cell_size or both (see details). If both criteria are provided, the first one
will be used first, and if there is still a draw, the second one will be used.}

\item{Min_cell_size}{If Resolution_approach is Erode_pixels, an integer value that indicates the minimum pixels a cell must have to be kept in the analysis. Cells below threshold will be removed from the cell mask image.}

\item{Mask_refinement_path}{(Optional) If NULL, a temporary directory will be created to temporarily save erosion masks (see details). Alternatively, the user can provided the path to an empty folder were erosion masks should be writen.}
}
\value{
It will save erosion masks in Mask_refinement_path directory. No other outputs are generated
}
\description{
Intended for internal use only.
}
\details{
Used in \code{\link[=Mask_conflict_resolver]{Mask_conflict_resolver()}}
}
\keyword{Internal}
