% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consensus_phenotype_assigner.R
\name{Consensus_phenotype_assigner}
\alias{Consensus_phenotype_assigner}
\title{Assigns phenotype labels by reaching a consensus between various methods}
\usage{
Consensus_phenotype_assigner(
  ...,
  Win_strategy = NULL,
  Arbitrary_threshold = NULL,
  Weights = NULL,
  No_consensus_value = NA_character_,
  Tie_break_method = NULL,
  N_cores = NULL
)
}
\arguments{
\item{...}{Two or more dataframes or tibbles containing a column with cell labels. A name must be provided for every tibble.}

\item{Win_strategy}{One of the following: 'Majority', 'Arbitrary', 'Difference' (see details).}

\item{Arbitrary_threshold}{If Win_strategy is Arbitrary or Difference, a numeric threshold between 0 and 1.}

\item{Weights}{(OPTIONAL) A numeric vector of length equal to the number of datasets provided.}

\item{No_consensus_value}{A character value indicating which label to use if no consensus is reached (default is NA_character_).}

\item{Tie_break_method}{If Win_strategy is Majority, the method to break potential ties. Either 'Random' or 'No_consensus' (see details).}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
Returns a tibble containing consensus cell labels.
}
\description{
The function allows various cell phenotyping methods to be combined into a final consensus label.
}
\details{
Majority voting assigns the label with the highest votes. If there is a tie between two or more methods, the tie can be resolved by randomly selecting a label, or by using the No_consensus_value.

Arbitrary voting assigns the label with the highest votes as long as it has received enough arbitrary votes. Arbitrary threshold must be between 0.5 and 1.

Difference voting assigns the label with the highest votes as long as the difference between the selected label and the second most voted option is above an arbitrary threshold.
}
\examples{
\dontrun{
#Generate datasets with random cell phenotype labels----------------------
CSM_Phenotypecell_test_1 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))
CSM_Phenotypecell_test_2 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))
CSM_Phenotypecell_test_3 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))

#Find consensus phenotype labels between the three results--------------
DATA_Phenotypes_consensus <-
Consensus_phenotype_assigner(
 Random_1 = CSM_Phenotypecell_test_1,
 Random_2 = CSM_Phenotypecell_test_2,
 Random_3 = CSM_Phenotypecell_test_3,

 Win_strategy = "Majority",
 Weights = c(2,1,1),
 No_consensus_value = NA_character_,
 Tie_break_method = "Random",
 N_cores = 2
 )
}

}
