% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_anomal_values.R
\name{continuous_anomal_values}
\alias{continuous_anomal_values}
\title{Check continuous out-of-range values}
\usage{
continuous_anomal_values(x = NULL, range = NULL)
}
\arguments{
\item{x}{(numeric) a numeric vector.}

\item{range}{(numeric) a numeric vector with 2 values, i.e. the
minimum and the maximum allowed values for
\code{x}.}
}
\value{
(numeric) a numeric vector with all the elements of \code{x} that
are not within the \code{range}. If no value is
identified a \code{NA} is returned.
}
\description{
Check if a continuous variable takes values that are outside the
range defined by two numerical values. The function returns a
numeric vector with the values of the variable that lie outside of
the range.
}
\examples{

x <- rnorm(n = 10L, mean = 10, sd = 2)
range   <- c(3, 12)
csm:::continuous_anomal_values(x = x, range = range)

}
