% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_check.R
\name{correlation_check}
\alias{correlation_check}
\title{Perform Correlation Check}
\usage{
correlation_check(data, title = "", min.obs)
}
\arguments{
\item{data}{This should be a data frame the first column should contain the site number the remaining columns should have the variables to be tested.}

\item{title}{This provides the prefix of the title for each plot. Default is empty.}

\item{min.obs}{This is the minimum number of observations a site can have and still be tested.}
}
\value{
figures with correlation check outputs
}
\description{
The function is to compare correlations for multiple continuous variables for different sites.
}
\examples{
\dontrun{
correlation_check(lb_mout1,title='',min.obs=3)
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

