% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_Arrangedcellfeaturedata_test.R
\docType{data}
\name{CSM_Arrangedcellfeaturedata_test}
\alias{CSM_Arrangedcellfeaturedata_test}
\title{Cell feature data formatted with Data_arrange_function.}
\format{
A data frame with 1047 rows and 7 variables
}
\usage{
data(CSM_Arrangedcellfeaturedata_test)
}
\description{
A dataset containing breast carcinoma cell feature that has been formatted. The original data
source is \code{\link[=CSM_RAWcellfeaturedata_test]{CSM_RAWcellfeaturedata_test()}}.
}
\keyword{datasets}
