% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_ClinicalTMA_test.R
\docType{data}
\name{CSM_ClinicalTMA_test}
\alias{CSM_ClinicalTMA_test}
\title{Image metadata information}
\format{
A data frame with 49 rows and 5 variables
}
\usage{
data(CSM_ClinicalTMA_test)
}
\description{
A dataset containing endometrial carcinoma image metadata. The metadata contains
information regarding tumor expression of Mismatch Repair Proteins, Age of the
patient, and survival status
}
\keyword{datasets}
