% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_Clustering_function.R
\name{CSM_Clustering_function}
\alias{CSM_Clustering_function}
\title{Computes clusters based on MARKERS}
\usage{
CSM_Clustering_function(
  Original_data,
  MARKERS,
  Strategy,
  Force_N_Clusters,
  Max_N_clusters_Consensus = NULL,
  Consensus_reps = NULL,
  Consensus_p_Items = NULL,
  Consensus_Cluster_Alg = NULL,
  Consensus_Distance = NULL,
  Consensus_Name = NULL,
  Max_SOM_clusters = NULL,
  Graph_type = NULL,
  Graph_Distance_method = NULL,
  Nearest_neighbors_for_graph = NULL,
  Graph_Method = NULL,
  Graph_Resolution = NULL,
  N_steps = NULL,
  N_K_centroids = NULL,
  Max_N_clusters_Meta = NULL,
  Consensus_reps_Meta = NULL,
  Consensus_p_Items_Meta = NULL,
  Consensus_Name_Meta = NULL,
  Batch_size = NULL,
  Max_N_clusters_Batch = NULL,
  N_initiations = NULL,
  Max_iterations = NULL,
  Quality_metric = NULL,
  Max_N_clusters_GMM = NULL,
  Max_iterations_km = NULL,
  Max_iterations_em = NULL,
  GMM_Distance = NULL,
  Samples_CLARA = NULL,
  Sample_per_CLARA = NULL,
  Max_N_clusters_CLARA = NULL,
  Distance_CLARA = NULL,
  N_cores = NULL
)
}
\value{
A tibble containing cell data and a column named \code{Cluster} containing clustering result.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Clustering_Phenotyper]{Clustering_Phenotyper()}}, \code{\link[=Neighborhood_discovery_function]{Neighborhood_discovery_function()}}, \code{\link[=ReClustering_function]{ReClustering_function()}}, \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}, \code{\link[=UTAG_Neighborhood_identifier]{UTAG_Neighborhood_identifier()}}
}
\keyword{Internal}
