% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_colordeconv_test.R
\docType{data}
\name{CSM_colordeconv_test}
\alias{CSM_colordeconv_test}
\title{Color extraction parameters}
\format{
A list containig parameters
}
\usage{
data(CSM_colordeconv_test)
}
\description{
A list containing color extraction parameters to obtain Hematoxylin and eosin
channels
}
\keyword{datasets}
