% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_Denoising_function.R
\name{CSM_Denoising_function}
\alias{CSM_Denoising_function}
\title{Identifies noisy cells}
\usage{
CSM_Denoising_function(
  Original_data = NULL,
  Denoising_strategy = NULL,
  Percentile = NULL,
  N_Standard_Deviations = NULL,
  Selected_threshold = NULL,
  Perform_Dimension_reduction = NULL,
  DATA_Reduction = NULL,
  Min_cell_no = NULL,
  Distance_radius = NULL
)
}
\value{
A list containing four items: NOISE_VECTOR containing logical values indicating if cells are noise or not, DATA_NOISE a tibble containing noisy cells,
DATA a tibble containing non-noisy cells, MARKERS the markers of non-noisy cells.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Clustering_Phenotyper]{Clustering_Phenotyper()}}, \code{\link[=UTAG_Neighborhood_identifier]{UTAG_Neighborhood_identifier()}}
}
\keyword{Internal}
