% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_Dimension_reduction_function.R
\name{CSM_Dimension_reduction_function}
\alias{CSM_Dimension_reduction_function}
\title{Performs dimension reduction (2 dimensions computed)}
\usage{
CSM_Dimension_reduction_function(
  Original_data = NULL,
  Dimension_reduction_strategy = NULL,
  Dimension_reduction_prop = NULL
)
}
\value{
A dataframe containing three columns Cell_no, DIMENSION_1 and DIMENSION_2.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Clustering_Phenotyper]{Clustering_Phenotyper()}}, \code{\link[=Neighborhood_discovery_function]{Neighborhood_discovery_function()}}, \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}, \code{\link[=UTAG_Neighborhood_identifier]{UTAG_Neighborhood_identifier()}}
}
\keyword{Internal}
