% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_DistanceToPixelcell_test.R
\docType{data}
\name{CSM_DistanceToPixelcell_test}
\alias{CSM_DistanceToPixelcell_test}
\title{Cell feature data containing cell distances to positive pixels}
\format{
A data frame with 1047 rows and 9 variables
}
\usage{
data(CSM_DistanceToPixelcell_test)
}
\description{
A dataset containing breast carcinoma cell feature that has been phenotyped.
The distance from each cell to the closest positive CK-Epcam pixel has been calculated.
The original data source is \code{\link[=CSM_RAWcellfeaturedata_test]{CSM_RAWcellfeaturedata_test()}}.
}
\keyword{datasets}
