% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_MiniHE_test.R
\docType{data}
\name{CSM_MiniHE_test}
\alias{CSM_MiniHE_test}
\title{Images of tissue stained with hematoxylin and eosin}
\format{
A list containig EBImage type image objects
}
\usage{
data(CSM_MiniHE_test)
}
\description{
A list containing two 1000x1000 pixel images of cardiac muscle stained with HE.
}
\keyword{datasets}
