% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_MiniMultiTiff_test.R
\docType{data}
\name{CSM_MiniMultiTiff_test}
\alias{CSM_MiniMultiTiff_test}
\title{Multiplexed images of breast carcinoma samples}
\format{
A list containig EBImage type image objects
}
\usage{
data(CSM_MiniMultiTiff_test)
}
\description{
A list containing two 500x500 pixel images of breast carcinoma samples.
Channels: "DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3"
}
\keyword{datasets}
