% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_ModelParam_test.R
\docType{data}
\name{CSM_ModelParam_test}
\alias{CSM_ModelParam_test}
\title{List containing model and parameters to predict cell labels}
\format{
A list containig model and parameters
}
\usage{
data(CSM_ModelParam_test)
}
\description{
A list containing the model and the parameters required to predict cell phenotype
labels from feature information.
}
\keyword{datasets}
