% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_Neighborhoods_test.R
\docType{data}
\name{CSM_Neighborhoods_test}
\alias{CSM_Neighborhoods_test}
\title{Closest neighbor matrix.}
\format{
A data frame with 798 rows and 15 variables
}
\usage{
data(CSM_Neighborhoods_test)
}
\description{
A dataset containing the closest neighbors for every cell carcinoma cell in the
dataset (OTHER cells cannot be COO). A distance of 50 pixels has been used to
define neighbors. The original data source is \code{\link[=CSM_RAWcellfeaturedata_test]{CSM_RAWcellfeaturedata_test()}}.
}
\keyword{datasets}
