% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_Phenotypecell_test.R
\docType{data}
\name{CSM_Phenotypecell_test}
\alias{CSM_Phenotypecell_test}
\title{Cell feature data containing cell phenotype labels.}
\format{
A data frame with 1047 rows and 8 variables
}
\usage{
data(CSM_Phenotypecell_test)
}
\description{
A dataset containing breast carcinoma cell feature that has been thresholded and
cell phenotype labels have been assigned. The original data source is
\code{\link[=CSM_RAWcellfeaturedata_test]{CSM_RAWcellfeaturedata_test()}}.
}
\keyword{datasets}
