% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_PhenotypeTMA_test.R
\docType{data}
\name{CSM_PhenotypeTMA_test}
\alias{CSM_PhenotypeTMA_test}
\title{Cell feature data containing cell phenotype labels}
\format{
A data frame with 504,958 rows and 5 variables
}
\usage{
data(CSM_PhenotypeTMA_test)
}
\description{
A dataset containing endometrial carcinoma cell feature that has been
phenotyped. The dataset contains information of more than 500K cells from
49 Tissue Micro Array images.
}
\keyword{datasets}
