% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_RAWcellfeaturedata_test.R
\docType{data}
\name{CSM_RAWcellfeaturedata_test}
\alias{CSM_RAWcellfeaturedata_test}
\title{Raw cell feature data, without format.}
\format{
A data frame with 1047 rows and 28 variables
}
\usage{
data(CSM_RAWcellfeaturedata_test)
}
\description{
A dataset containing cell feature data without specific format obtained from
two 500x500 breast carcinoma images. Original images were obtained from
immunoatlas.org. The data contains morphologic features and marker expression
data for DAPI, CD8a, GZMB and CK-EPCAM
}
\keyword{datasets}
