% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_SegmentParams_test.R
\docType{data}
\name{CSM_SegmentParams_test}
\alias{CSM_SegmentParams_test}
\title{List containing cell segmentation parameters}
\format{
A list containig parameters
}
\usage{
data(CSM_SegmentParams_test)
}
\description{
A list containing the parameters to perform cell segmentation and feature extraction.
}
\keyword{datasets}
