% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulative_Interaction_generator.R
\name{Cumulative_Interaction_generator}
\alias{Cumulative_Interaction_generator}
\title{Calculates the cumulative interaction matrix}
\usage{
Cumulative_Interaction_generator(
  N_cores = NULL,
  DATA = NULL,
  Start_from = NULL,
  Stop_at = NULL,
  Sampling_frequency = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A list containing cell distance matrix. It can be computed using \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}} or \code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}} functions.}

\item{Start_from}{A numeric value indicating the minimum distance to be analyzed.}

\item{Stop_at}{A numeric value indicating the maximum distance to be analyzed.}

\item{Sampling_frequency}{A numeric value indicating the distance sampling frequency.}
}
\value{
A list containing the cumulative interaction matrix for every image.
}
\description{
The cumulative interaction matrix contains information regarding the total potential interactions between COO and target cells. For a a given set of increasing distances,
the number of target cells present from every COO is calculated.
}
\seealso{
\code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}}, \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}}, \code{\link[=Distance_analyzer]{Distance_analyzer()}}, \code{\link[=Cells_in_Radius_analyzer]{Cells_in_Radius_analyzer()}}.
}
