% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_arrange_function.R
\name{Data_arrange_function}
\alias{Data_arrange_function}
\title{Formats cell feature data}
\usage{
Data_arrange_function(DATA, X, Y, Subject_Names, Markers_to_keep)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{X}{A character indicating the column name containing X coordinates.}

\item{Y}{A character indicating the column name containing Y coordinates.}

\item{Subject_Names}{A character indicating the column name containing image name.}

\item{Markers_to_keep}{A character vector indicating the names of the columns to be kept in the analysis.}
}
\value{
Returns a tibble with cell features. The function adds a unique Cell_no to every cell in the dataset.
}
\description{
\code{Data_arrange_function()}must be run in order to format cell feature matrix. Many CSM functions won't work if the data has not been adequately formatted
}
