% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_neighborhoods_renamer.R
\name{DATA_neighborhoods_renamer}
\alias{DATA_neighborhoods_renamer}
\title{Modifies neighborhood labels}
\usage{
DATA_neighborhoods_renamer(DATA = NULL, New_names = NULL, Old_names = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing neighborhood data, including a column named 'Neighborhood_assignment'.}

\item{New_names}{A character vector indicating the new names. The length must be equal to the number of unique neighborhood labels.}

\item{Old_names}{(OPTIONAL) A character vector indicating the neighborhood labels that need to be modified. The length of this vector must be equal to the length of New_names.}
}
\value{
Returns a tibble with cell features and the Neighborhood_assignment column with the modified labels.
}
\description{
The function allows the user to modify neighborhood labels. The function works with any dataframe or tibble containing a column named 'Neighborhood_assignment'.
}
\examples{
\dontrun{
print(unique(CSM_Neighborhoods_test$Neighborhood_assignment))

New_names_DATA <-
 DATA_neighborhoods_renamer(
    DATA = CSM_Neighborhoods_test,
    New_names = c("Tumor_low_density", "Tumor_high_density"),
    Old_names = c("Tumor_Other", "Tumor_rich")
)

print(unique(New_names_DATA$Neighborhood_assignment))
}

}
