% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_Phenotype_renamer.R
\name{DATA_Phenotype_renamer}
\alias{DATA_Phenotype_renamer}
\title{Modifies cell phenotype labels}
\usage{
DATA_Phenotype_renamer(DATA = NULL, New_names = NULL, Old_names = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data, including a column named 'Phenotype'.}

\item{New_names}{A character vector indicating the new names. The length must be equal to the number of unique cell phenotype labels.}

\item{Old_names}{(OPTIONAL) A character vector indicating the phenotype labels that need to be modified. The length of this vector must be equal to the length of New_names.}
}
\value{
Returns a tibble with cell features and the phenotype column with the modified labels.
}
\description{
The function allows the user to modify cell phenotype labels. The function works with any dataframe or tibble containing a column named 'Phenotype'.
}
\examples{
\dontrun{
print(unique(CSM_Phenotypecell_test$Phenotype))

New_names_DATA <-
 DATA_Phenotype_renamer(
    DATA = CSM_Phenotypecell_test,
    New_names = c("CD8_GZMB_NEGATIVE", "CD8_GZMB_POSITIVE"),
    Old_names = c("CD8_GZMBneg", "CD8_GZMBpos")
)

print(unique(New_names_DATA$Phenotype))
}

}
