% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_from_server.R
\name{db_update_from_server}
\alias{db_update_from_server}
\title{Get the data of the study from the REDCap server}
\usage{
db_update_from_server(
  token,
  path_data = data_path(),
  redcap_info = c("record_id", "redcap_data_access_group"),
  file_name = "tidy_data.rds"
)
}
\arguments{
\item{token}{(chr) a character string that specifies the token to
get access to the data of the study}

\item{path_data}{(chr) path for data folder (default is \verb{data/}
under the current project or the \link[here:here]{here} path. If the
folder does not exist, it will be created.)}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled.}

\item{file_name}{(chr) file name in which data are stored.}
}
\value{
(lgl) \code{TRUE} if success, \code{FALSE} otherwise
}
\description{
Get the data of the study from the REDCap server
}
\examples{
\dontrun{
  db_update_from_server()
}
}
