% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBSCAN_Tumor_Stroma_identifier.R
\name{DBSCAN_Tumor_Stroma_identifier}
\alias{DBSCAN_Tumor_Stroma_identifier}
\title{Calculates tumor and stromal compartment based on density based clustering}
\usage{
DBSCAN_Tumor_Stroma_identifier(
  DATA_Phenotypes,
  Index_phenotype,
  Image_preview = NULL,
  Min_cells,
  Distance_radius,
  N_cores = 1
)
}
\arguments{
\item{DATA_Phenotypes}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Index_phenotype}{A character value indicating the cell phenotype to be used to calculate the tissue compartment.}

\item{Image_preview}{A character value indicating the name of the image to be used in the preview.}

\item{Min_cells}{A integer value indicating the minimum number of cells required to be a cluster.}

\item{Distance_radius}{A numeric value indicating the distance to be sampled.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
Returns a tibble with cell features and a column named 'Compartment' containing cell location.
}
\description{
The function calculates tissue compartments (usually tumor and stroma) based on density based clustering. Cells are considered to be in the tissue compartment if they are clustered in proximity to index cell types.
The function is based on the dbscan::dbscan function.
}
\examples{
\dontrun{
DBSCAN_Tumor_Stroma_identifier(
    DATA_Phenotypes = CSM_Phenotypecell_test,
    Index_phenotype = "TUMOR",
    Min_cells = 3,
    Distance_radius = 50,
    N_cores = 1
)
}

}
