% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Describe_my_computing_unit.R
\name{Describe_my_computing_unit}
\alias{Describe_my_computing_unit}
\title{Obtain CPU information and available RAM}
\usage{
Describe_my_computing_unit()
}
\value{
A list containing information regarding CPU, RAM, R version being used and Operating system
}
\description{
Obtain CPU information and available RAM
}
