% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_data.R
\name{diff_til_stat}
\alias{diff_til_stat}
\title{Go To FRED-MD/FRED-QD Website}
\usage{
diff_til_stat(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
Output vector
}
\description{
This function differences time series until KPSS Test deems stationary at 95\% confidence level.
}
