% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_pref.R
\name{digit_preferece}
\alias{digit_preferece}
\title{Perform Digit Preference Check}
\usage{
digit_preferece(data, digit, benford, title = "")
}
\arguments{
\item{data}{This should be a dataframe with the site number in the first column followed by the continuous measurements to check}

\item{digit}{leading or tailing. This tells the program whether to look at the leading (first) or tailing (last) digit of each value. Note: Benford's law only works with leading digits so if Benford is set as TRUE the value of digit will be overridden and taken as leading}

\item{benford}{TRUE or FALSE. If this is set as TRUE, then the frequency of the leading digits in each site will be compared to Benford's distribution. If this is set as FALSE then the frequency of the leading or tailing digits in each site will be compared to the frequency in all of the other sites put together}

\item{title}{This provides the prefix of the title for each plot. Default is empty.}
}
\value{
Table with digit preference check result along with calculated p-values
}
\description{
The function is to check the existence of digit preference in each site for continuous variables.
}
\examples{
\dontrun{
digit_preferece(data,"tailing",FALSE,'title')
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

