% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_analyzer.R
\name{Distance_analyzer}
\alias{Distance_analyzer}
\title{Analyzes the distance pattern between two cell types for all images}
\usage{
Distance_analyzer(
  N_cores = NULL,
  DATA = NULL,
  DATA_RANDOM = NULL,
  Metric = NULL,
  Include_Random = NULL,
  By_Sample_Random = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A list containing cell distance matrix. It can be computed using \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}} function.}

\item{DATA_RANDOM}{(OPTIONAL) A list containing random cell distance matrix. It can be computed using \code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}} function.}

\item{Metric}{A character value indicating the type of analysis strategy. One of the following: Average_Distance, Min_Distance, Max_Distance (see details).}

\item{Include_Random}{A logical value indicating if the Random background should be used.}

\item{By_Sample_Random}{If a random background needs to be included, a logical value indicating if the random background should be calculated by sample or by experiment (see details).}
}
\value{
A tibble containing a summary by sample of spatial interactions.
}
\description{
The function summarizes the spatial interaction pattern between two cell types given a distance matrix.
}
\details{
Average distance computes the average distance by image between Cells of origin (COO) and target cells. Min distance averages the minimum distance from every COO to target cells. Max ditance averages
the maximum distance between COO and target cells.

If By_Sample_Random is TRUE, for every sample, a specific random background will be calculated to compute expected spatial interaction metrics. If By_Sample_Random is FALSE, the random background will be calculated using all image information. The same random background will be used for all images.
}
\seealso{
\code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}}, \code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}}, \code{\link[=Cell_to_Cell_graph_maker]{Cell_to_Cell_graph_maker()}}
}
