% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_matrix_generator.R
\name{Distance_matrix_generator}
\alias{Distance_matrix_generator}
\title{Calculates distance between two cell types}
\usage{
Distance_matrix_generator(
  N_cores = 1,
  DATA,
  Cell_Of_Origin,
  Target_Cell,
  Allow_Cero_Distance = FALSE,
  Perform_edge_correction = FALSE,
  Hull_ratio = NULL,
  Distance_to_edge = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Cell_Of_Origin}{A character value indicating the cell phenotype label of the Cell of Origin.}

\item{Target_Cell}{A character value indicating the cell phenotype label of the Target cell.}

\item{Allow_Cero_Distance}{A logical value indicating if zero distances should be allowed. This may be relevant when COO and Target cell labels are the same.}

\item{Perform_edge_correction}{A logical value indicating if COO close to the edge of the tissue should be removed from the analysis.}

\item{Hull_ratio}{If edge correction needs to be performed, a numeric value indicating the hull ratio. Smaller values calculate more precise edge silhouettes at the cost of being more computationally demanding.}

\item{Distance_to_edge}{If edge correction needs to be performed, the distance the maximum distance to the edge allowed. Cells closer to the edge will be removed.}
}
\value{
A list containing the distance matrix for each image. Rows represent Cells of Origin and columns represent Target cells.
}
\description{
The function calculates the Euclidean distance between two cell phenotypes. One is considered the Cell of Origin (COO) and another is considered the target cell.
The COO and target cell can be the same cell type.
}
\examples{
\dontrun{
Distance_matrix_generator(
    N_cores = 2,
    DATA = CSM_Phenotypecell_test,
    Cell_Of_Origin = "CD8_GZMBneg",
    Target_Cell = "TUMOR",
    Allow_Cero_Distance = FALSE,
    Perform_edge_correction = TRUE,
    Hull_ratio = 1,
    Distance_to_edge = 10
)
}

}
\seealso{
\code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}}, \code{\link[=Cumulative_Interaction_generator]{Cumulative_Interaction_generator()}}, \code{\link[=Distance_analyzer]{Distance_analyzer()}}, \code{\link[=Cells_in_Radius_analyzer]{Cells_in_Radius_analyzer()}}, \code{\link[=Cell_to_Cell_graph_maker]{Cell_to_Cell_graph_maker()}}
}
