% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM_start_message.R
\name{.onAttach}
\alias{.onAttach}
\title{This regulates the package start-up message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Used to obtain library name}

\item{pkgname}{Used to obtain package name}
}
\description{
This regulates the package start-up message
}
\keyword{Internal}
