% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBImage_color_thresholder.R
\name{EBImage_color_thresholder}
\alias{EBImage_color_thresholder}
\title{Thresholds RGB channels of a color image}
\usage{
EBImage_color_thresholder(
  Image = NULL,
  Target = c(1, 1, 1),
  Color_Tolerance = NULL,
  Final_Tolerance = NULL
)
}
\arguments{
\item{Image}{A EBImage object.}

\item{Target}{A numeric vector of length = 3. Represents the target RGB channels.}

\item{Color_Tolerance}{A numeric vector of length = 3. Represents the individual tolerances of target RGB channels.}

\item{Final_Tolerance}{A numeric value specifying the final tolerance.}
}
\value{
The function returns a greyscale EBImage object
}
\description{
Intended for internal use only
}
\keyword{Internal}
