% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_absent.R
\name{exclude_absent}
\alias{exclude_absent}
\title{Exclude absent elements}
\usage{
exclude_absent(target, reference)
}
\arguments{
\item{target}{(chr) vector with elements whose inclusion must be
checked}

\item{reference}{(named chr) vector with reference names}
}
\value{
(chr) vector with the remaining elements (subset of the
target) after the checking has been performed
}
\description{
Given a target vector with elements and a reference vector with
named elements, it returns all the elements from the target
vector that are not present in the reference vector or that, if
they appear in the reference vector, the associated names are
themselves included in the target vector.
}
\details{
the function (after it checks the elements from the
reference vector are not included in the target vector) implements
the subsetting of a character vector throughout the names of its
elements (https://adv-r.hadley.nz/subsetting.html#subset-multiple).
Moreover, \code{\link{\%in\%}} takes as \link{FALSE} the presence of a \link{NA}
(see \emph{Details} section of \code{\link{\%in\%}})
}
\examples{

\dontrun{

#' matteo <- c("c", "d", "f", "g", "h")
paolo <- c("a", "b", "c", "d", "f", "g", "h")
stefano <- c("a", "c", "d", "f", "g", "h")
marco <- c("f", "g")
davide <- purrr::set_names(x = c("f", "g"), nm = c("a", "b"))

exclude_absent(matteo, davide) # c("c", "d", "h")
exclude_absent(paolo, davide)  # paolo
exclude_absent(stefano, davide) # c("a", "c", "d", "f", "h")
exclude_absent(marco, davide) # NA_character_

}

}
