% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_categorical_labels.R
\name{extract_categorical_labels}
\alias{extract_categorical_labels}
\title{Return the labels of categorical variables}
\usage{
extract_categorical_labels(meta_data, fields_names, labels)
}
\arguments{
\item{meta_data}{(tibble) the output from \link{tidy_extract} when
\code{type = 'meta'}}

\item{fields_names}{(character) a character with the name of the
column of \code{meta_data} that contains
the names of fields.}

\item{labels}{(character) a character with the name of the column of
\code{meta_data} that contains the labels
associated to the variables.}
}
\value{
(named list) a named list. Each element corresponds to the
labels associated to the categorical variables.
The names of the elements are the names of the
variables
}
\description{
The functions returns the labels associated to each level of the
categorical variables.
}
\examples{
md <- tibble::tibble(
    field_name = c("age", "sex", "hypertension", "diabetes"),
    sheet = c("demo", "demo", "clinical", "clinica"),
    fct_level = list(
        character(0), c("female", "male"), c("yes", "no"),
        c("yes", "no")
    )
)

extract_categorical_labels(md, "field_name", "fct_level")

}
