% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_range.R
\name{extract_range}
\alias{extract_range}
\title{Extract the range of continuous variables}
\usage{
extract_range(
  meta_data,
  fields_names,
  range_low = "text_validation_min",
  range_upp = "text_validation_max"
)
}
\arguments{
\item{meta_data}{(tibble) the output from \link{tidy_extract} when
\code{type = 'meta'}}

\item{fields_names}{(character) a character with the name of the
column of \code{meta_data} that contains
the names of fields.}

\item{range_low}{(character) a character with the name of the column of
\code{meta_data} that contains the lower
bound of the range associated to each
continuous variable.}

\item{range_upp}{(character) a character with the name of the column of
\code{meta_data} that contains the upper
bound of the range associated to each
continuous variable.}
}
\value{
(named list) a named list. Each element corresponds to the
range associated to the continuous variables.
The names of the elements are the names of the
variables
}
\description{
The functions returns range associated to each continuous variable.
}
\examples{
md <- tibble::tibble(
    field_name = c("age", "sex", "hypertension", "diabetes"),
    sheet = c("demo", "demo", "clinical", "clinica"),
    min = c(18, NA, NA, NA),
    max = c(75, NA, NA, NA)
)

extract_range(md, "field_name", "min", "max")

}
