% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_data.R
\name{fred_md_site}
\alias{fred_md_site}
\title{Go To FRED-MD/FRED-QD Website}
\usage{
fred_md_site()
}
\value{
Website in default browser
}
\description{
This function opens the website with latest versions, codes, and documentation.
}
