% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gcross_calculator.R
\name{Gcross_calculator}
\alias{Gcross_calculator}
\title{Analyzes the spatial interaction between two cell types across images by using the Gcross function}
\usage{
Gcross_calculator(
  DATA,
  Cell_Of_Origin,
  Target_Cell,
  Stop_at,
  Sampling_frequency,
  Use_Clinical = FALSE,
  DATA_Clinical = NULL,
  Clinical_var = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Cell_Of_Origin}{A character value indicating the cell phenotype label of the Cell of Origin.}

\item{Target_Cell}{A character value indicating the cell phenotype label of the Target cell.}

\item{Stop_at}{A numeric value indicating the maximum distance to be analyzed.}

\item{Sampling_frequency}{A numeric value indicating the distance sampling frequency.}

\item{Use_Clinical}{A logical value indicating if clinical data should be used in the plot generation process.}

\item{DATA_Clinical}{A dataframe or tibble containing image metadata.}

\item{Clinical_var}{A character value indicating the column name used in the clinical data analysis.}
}
\value{
A list containing two outputs: Raw_DATA contains the expected interaction probability for every distance point. Simplified_DATA cotains the AUC data by image.
}
\description{
The function calculates the probability of spatial encounter between two cell types as distance sampling increases using the Gcross functions.
Steep rise in probability as distance increases means high chances of spatial interaction. The function is calculated using the spatstat R metapackage
(specifically the spatstat.explore::Gcross function). The Area Under the Curve (AUC) of the resulting function can be used as a metric of spatial association.
}
\examples{
\dontrun{
Gcross_calculator(
   DATA = CSM_Phenotypecell_test,
   Cell_Of_Origin = "CD8_GZMBneg",
   Target_Cell = "TUMOR",
   Stop_at = 200,
   Sampling_frequency = 10,
   Use_Clinical = FALSE
)
}

}
