% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dm.R
\name{gen_dm}
\alias{gen_dm}
\title{Generate (#tx) by (#co) distance matrix}
\usage{
gen_dm(df, scaling = 1, metric = c("maximum", "euclidean", "manhattan"))
}
\arguments{
\item{df}{dataframe}

\item{scaling}{vector of scaling constants for covariates}

\item{metric}{distance metric to use}
}
\value{
(#tx) by (#co) distance matrix
}
\description{
Generate (#tx) by (#co) distance matrix
}
