% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branching_logic.R
\name{get_branching_logic}
\alias{get_branching_logic}
\title{Get the branching logic conditions}
\usage{
get_branching_logic(meta_data, fields_names, branching_logic)
}
\arguments{
\item{meta_data}{(tibble) a tibble containing the meta_data
information of the study.}

\item{fields_names}{(character) a character with the name of the
column of \code{meta_data} that contains
the names of fields.}

\item{branching_logic}{(character)) a character with the name of the
column of \code{meta_data} that the
branching logics associated to
each field.}
}
\value{
(tibble) a tibble with two columns: one column must contains
the names of the fields. The other column must
contains the associated condition for missing data
checking.
}
\description{
Get the conditions for missing data checking. If a variable has no
branching logic and it is NA, then the information is missing. If a
variable is missing and it has a branching logic, then it is missing
only if it is missing and the branching logic is not satisfied.
}
\examples{
\dontrun{

 meta_data <- tibble::tibble(
   field_name = c("id", "sex", "diabetes", "iddm"),
   sheet = c(rep("demo", 2L), rep("clinical", 2L)),
   branching_logic = c(rep(NA_character_, 3L), "diabetes")
 )

 get_branching_logic(
   meta_data = meta_data,
   fields_names = "field_name",
   branching_logic = "branching_logic"
 )

}
}
