% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incomplete_sheets.R
\name{get_incomplete_sheets}
\alias{get_incomplete_sheets}
\title{Returns the incomplete sheets for each patients}
\usage{
get_incomplete_sheets(nested_tables, redcap_info = c("record_id", "center"))
}
\arguments{
\item{nested_tables}{(tibble) the output from \link{nest_tables}}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled.}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble that identifies the
incomplete or unverified
sheets for all the patients
}
\description{
The function returns the sheets that are marked as incomplete or
unverified for each patients enrolled in the study.
}
\examples{
\dontrun{
  nested <- readr::read_rds(db_update_from_server())

  get_incomplete_sheets(nested, redcap_info = c("id", "center"))
}

}
