% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_vars.R
\name{get_missing_vars}
\alias{get_missing_vars}
\title{Returns the fields with missing information for a patient in a form}
\usage{
get_missing_vars(data)
}
\arguments{
\item{data}{(tibble) a tibble with one row and as many columns as the
fields in a given form}
}
\value{
a (character) with the names of the missing fields for the
patient.
}
\description{
The function returns the fields that are missing for a patient
in a given form.
}
\examples{

df <- tibble::tibble(
  age = 63,
  sex = "female",
  diabetes = NA_character_,
  hypertension = "yes",
  nyha = NA_character_
)

get_missing_vars(df) # c("diabetes", "nyha")

}
