% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_vars_by.R
\name{get_missing_vars_by}
\alias{get_missing_vars_by}
\title{Returns the fields with missing information for a grouping
variable(s) in a form}
\usage{
get_missing_vars_by(data, group)
}
\arguments{
\item{data}{(tibble) a tibble with one row and as many columns as the
fields in a given form}

\item{group}{(character) the name(s) of the variable(s) by which the
data.frame should be grouped}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble with one column
with the values of the
grouping variable and one
column with the missing
fields associated to each
value of the grouping
variable
}
\description{
The function returns the fields that are missing in a given form for
one or more grouping variables.
}
\examples{

data <- tibble::tibble(
    id = c("id_1", "id_2", "id_3"),
    age = c(63, 65, 66),
    sex = c("female", NA_character_, "male"),
    diabetes = c(NA_character_, "yes", "no"),
    hypertension = c("yes", NA_character_, NA_character_),
    nyha = rep(NA_character_, 3L)
)

get_missing_vars_by(data, "id")

}
