% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheet.R
\name{get_sheet}
\alias{get_sheet}
\title{Get a sheet table}
\usage{
get_sheet(x, sheet, field = NULL)
}
\arguments{
\item{x}{(tbl_df) output from \link{db_update_from_server}}

\item{sheet}{(chr) name of sheet to get the data from}

\item{field}{(chr, default = NULL) if not NULL, provide the name of
field containing the required sheet. If NULL, only one field
containing the required sheet is supposed to be present; that
field is used. If more than one field with the required sheet
exists, and field is NULL, an error is thrown, suggesting the
possible fields to consider.}
}
\value{
the required sheet \link[tibble:tibble-package]{tibble}
}
\description{
The function retrieves a data.frame with the data of a sheet
associated to a given field.
}
\examples{
\dontrun{
  tipnet <- read_rds(db_update_from_server())
  tipnet \%>\% get_sheet("anagrafica")
  tipnet \%>\% get_sheet("anagrafica", field = "anagrafica")

}
}
