% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheet_labels.R
\name{get_sheet_labels}
\alias{get_sheet_labels}
\title{Get a the labels associated to each field of a given sheet}
\usage{
get_sheet_labels(
  data,
  meta_data,
  redcap_info = c("record_id", "center"),
  fields_names = "field_name",
  fields_labels = "field_label"
)
}
\arguments{
\item{data}{(data.frame) output from \link{get_sheet}.}

\item{meta_data}{(data.frame) the data.frame the contains the
meta-data of the study.}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled.}

\item{fields_names}{(character) a character with the name of the
column of \code{meta_data} that contains
the names of fields.}

\item{fields_labels}{(character) a character with the name of the
column of \code{meta_data} that contains
the labels of fields.}
}
\value{
a (character) vector whose elements are the labels associated
to the fields of the sheet. The vector must be named and the
names correspond to the names of the fields.
}
\description{
The function returns the labels associated to each field of a given
sheet.
}
\examples{
df <- tibble::tibble(
    id = rep(glue::glue("id_{1:5}"), 2L),
    center = rep(c("center_1", "center_2"), each = 5L),
    x = c(0.1, 0.2, 0.15, 0.4, 0.5, 0.6, 0.3, 0.2, 0.2, 0.1),
    y = c(0.1, 0.11, 0.12, 0.3, 0.3, 0.3, 0.4, 0.5, 0.6, 0.55),
    z = c(0.1, 0.2, 0.15, 0.4, 0.5, 0.6, 0.3, 0.2, 0.2, 0.1),
    complete = rep("complete", 10L)
)

meta_data <- tibble::tibble(
    field_name = c("age", "sex", "diabetes", "x", "y", "z"),
    sheet = c("demo", "demo", "clinical", "random", "random", "random"),
    field_label = c(
        "age", "sex", "diabetes", "first tt", "second tt", "third tt"
    )
)

redcap_info <- c("id", "center")
fields_labels <- "field_label"
fields_names <- "field_name"

get_sheet_labels(
    data = df, meta_data = meta_data,
    redcap_info = redcap_info, fields_names = fields_names,
    fields_labels = fields_labels
)

}
