% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcomplfup.R
\name{ggcomplfup}
\alias{ggcomplfup}
\title{Plot the completeness of the follow-up}
\usage{
ggcomplfup(data, outcome, total, type, percentage)
}
\arguments{
\item{data}{(data.frame) a data.frame that contains the data
of the completeness.}

\item{outcome}{(character) a string with the name of the variable
in \link{data} that contains the name of the endpoints.}

\item{total}{(character) a string with the name of the variable
in \link{data} that contains the number of subject on which
the completeness is evaluated.}

\item{type}{(character) a string with the name of the variable
in \link{data} that contains the type of percentage, i.e.
completeness or missingness.}

\item{percentage}{(character) a string with the name of the variable
in \link{data} that contains the percentage of completeness
and missingness of the follow-up.}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
The function aims to provide a graphical representation of the
completeness of the follow-up.
}
\details{
The function is intended to be used when the report of
data quality check is implemented. It can be useful for
clinicians to check the completeness of follow-up in their
data. The input must be a data.frame with the same structure
as described above.
}
\examples{

data <- tibble::tibble(
    outcome = c("death", "death", "mi", "mi"),
    total = rep(75, 4L),
    type = c(
        "completeness", "missingness", "completeness", "missingness"
    ),
    percentage = c(65.5, 34.5, 98, 2)
)

ggcomplfup(data, "outcome", "total", "type", "percentage")

}
