% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_heterogeneity_calculator.R
\name{Global_heterogeneity_calculator}
\alias{Global_heterogeneity_calculator}
\title{Calculates cell composition heterogeneity.}
\usage{
Global_heterogeneity_calculator(DATA, Phenotypes_included)
}
\arguments{
\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}
}
\value{
Generates a tibble containing heterogeneity metrics by image.
}
\description{
The function calculates various heterogeneity metrics based on cell composition distribution. It does not take into account spatial information.
}
\details{
Shannon, Simpson,Inverse Simpson and Renyi entropy indexes are calculated using the vegan package.

Rao index is calculated using the picante::raoD function.

Gini index is calculated using the DescTools::Gini function.

Kullback_Leibler and Jensen_Shannon indexes are calculated using the philentropy package. They are computed in two different ways:
First '_Sample' indexes compare the observed cell composition against an homogeneous distribution where all cells have the same prevalence in the sample.
Second '_Experiment' indexes compare the observed cell composition against the cell distribution observed in all the cells from DATA.
}
\examples{
\dontrun{
Global_heterogeneity_calculator(
    DATA = CSM_Phenotypecell_test,
    Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype)
)
}

}
