% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_channel_color_merger.R
\name{Image_channel_color_merger}
\alias{Image_channel_color_merger}
\title{Merges image channels using different colors}
\usage{
Image_channel_color_merger(Image, Parameter_list)
}
\arguments{
\item{Image}{A magick image object generated using magick::image_read().}

\item{Parameter_list}{A list containing merging parameters (see details and examples below).}
}
\value{
A magick image object with the channels merged.
}
\description{
The function generates and RGB image resulting from merging various image channels.
}
\details{
The parameter list must be a list of lists. Every element in parameter list must be named. Individual list elements within the Parameter_list must contain
the parameters used to process that specific channel before merging. This include indicating the channel being processed (a numeric value indicating the channel position within the image),
the color to be applied to the channel (indicated by name or HEX code), the gamma correction (a numeric value between 0.001 and 1000), the min and max pixel value (a numeric value between 0 and 100 for both of them) and
equalize (a logical value).
}
\examples{
\dontrun{
#Read the image to be processed----------------------------------------------
Image <- magick::image_read(CSM_MiniMultiTiff_test[[2]])

#Generate the merging parameters (in this case 4 channels will be merged)----
Merge_parameter_list <- 
  list(
      DAPI = list(channel_index = 1,
                  color = "blue",
                  gamma = 1,
                  min = 0,
                  max = 80,
                  equalize = FALSE),

      CK = list(channel_index = 5,
                color = "white",
                gamma = 1.1,
                min = 5,
                max = 20,
                equalize = FALSE),
 
      CD8 = list(channel_index = 6,
                 color = "#f0703a",
                 gamma = 1,
                 min = 10,
                 max = 30,
                 equalize = FALSE),

      GZMB = list(channel_index = 3,
                  color = "cyan",
                  gamma = 1,
                  min = 5,
                  max = 20,
                  equalize = FALSE)

)

#Perform the merging
Image_channel_color_merger(Image = Image,
                          Parameter_list = Merge_parameter_list)

}
}
