% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_deconvolution_function.R
\name{Image_deconvolution_function}
\alias{Image_deconvolution_function}
\title{Separates RGB color images into color channels according to user preferences}
\usage{
Image_deconvolution_function(
  Directory = NULL,
  Output_directory = NULL,
  Deconvolution_parameters = NULL,
  N_cores = NULL
)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where RGB images are present.}

\item{Output_directory}{Character string specifying the path to the folder where  output images are written. It must be an empty folder.}

\item{Deconvolution_parameters}{A list containing parameters. Created using \code{\link[=Color_deconvolution_App_launcher]{Color_deconvolution_App_launcher()}}.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
The function writes the images in the ouput directory.
}
\description{
\code{Image_deconvolution_function()} generates multi-channel tiff images according to color deconvolution parameters.
}
\seealso{
\code{\link[=Color_deconvolution_App_launcher]{Color_deconvolution_App_launcher()}}
}
