% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_from_tile_rebuilder.R
\name{Image_from_tile_rebuilder}
\alias{Image_from_tile_rebuilder}
\title{Build images from tile image chunks}
\usage{
Image_from_tile_rebuilder(
  Directory,
  Output_directory,
  RGB_Color_images = FALSE,
  N_cores = 1
)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where image tiles are present.}

\item{Output_directory}{Character string specifying the path to the folder where  output images are written. It must be an empty folder.}

\item{RGB_Color_images}{Logical. Is the image a RGB color image?}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
The function writes the images in the ouput directory.
}
\description{
\code{Image_from_tile_rebuilder()} builds a complete image from it's tiles.
Warning: Image reconstruction is limited to ≈ 2-3Gb.
Image reconstruction is designed to rebuild complete images from tiles once these have undergone pixel thresholding using \code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}}.
}
\details{
Another warning: Image re-build for overlapping tiles is not currently supported.

Another warning: The re-building process is based on the file names. If file names are modified, the re-building process can yield unexpected results.
}
\examples{
\dontrun{
Input_Dir <- tempfile(pattern = "tempdir1_Input")
Output_Dir_tiles <- tempfile(pattern = "tempdir2_Output")
Output_Dir_reconstruction <- tempfile(pattern = "tempdir3_Output")
dir.create(Input_Dir, recursive = TRUE)
dir.create(Output_Dir_tiles, recursive = TRUE)
dir.create(Output_Dir_reconstruction, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Divide images into tiles------------------------------------------------------
Image_tile_deconstruction_function(
   Directory = Input_Dir,
   Output_directory = Output_Dir_tiles,
   Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3"),
   Channels_to_keep = c("DAPI", "GZMB", "CK-EPCAM", "CD8a"),
   RGB_Color_images = FALSE,
   Tile_pixel_size = 250,
   Tile_Overlap = 0,
   N_cores = 1
)

#Rebuild images from tiles-----------------------
Image_from_tile_rebuilder(
   Directory = Output_Dir_tiles,
   Output_directory = Output_Dir_reconstruction,
   RGB_Color_images = FALSE,
   N_cores = 1
)

#Check the files created-------------------------------------------------
list.files(Output_Dir_reconstruction)

#Remove directories---------------------------------------------------------
unlink(c(Input_Dir, Output_Dir_tiles, Output_Dir_reconstruction), recursive = TRUE)

}
}
\seealso{
\code{\link[=Image_tile_deconstruction_function]{Image_tile_deconstruction_function()}}
}
