% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_length_calculator.R
\name{Image_length_calculator}
\alias{Image_length_calculator}
\title{Generate a summary of the width, length and surface of images in a dataset}
\usage{
Image_length_calculator(DATA)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}
}
\value{
A tibble containing image dimension information.
}
\description{
The function can be used  to plan the image tiling strategy.
}
\examples{
\dontrun{
Image_length_calculator(DATA = CSM_Phenotypecell_test)
}

}
\seealso{
\code{\link[=Suggested_Tile_Size_Calculator]{Suggested_Tile_Size_Calculator()}}, \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}
}
