% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_size_calculator.R
\name{Image_size_calculator}
\alias{Image_size_calculator}
\title{Approximates tissue size}
\usage{
Image_size_calculator(
  DATA = NULL,
  Strategy = NULL,
  Image_to_plot = NULL,
  Tile_accuracy = NULL,
  Hull_ratio = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Strategy}{Either 'Tiling' or 'Concave_hull'}

\item{Image_to_plot}{(Optional) A character value indicating the image to be used in the preview. If NULL the smallest tissue will be used.}

\item{Tile_accuracy}{Lower values calculate the area in a more precise manner, with higher computational times. If the value is too low, it can classify areas of stroma as not being tissue.}

\item{Hull_ratio}{A numeric value indicating the hull ratio. Smaller values calculate more precise edge silhouettes at the cost of being more computationally demanding.}
}
\value{
Returns a tibble with tissue size for each image
}
\description{
The function approximates tissue size either by tiling the tissue and analyzing the amount of tiles that present cells or by calculating the silouette of the tissue by calculating the concave hull.
The area can then be used by other functions to calculate densities like \code{\link[=Phenotype_quantifier]{Phenotype_quantifier()}} or \code{\link[=Neighborhood_Quantifier]{Neighborhood_Quantifier()}}.
}
