% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_tile_deconstruction_function.R
\name{Image_tile_deconstruction_function}
\alias{Image_tile_deconstruction_function}
\title{Divide images into tile chunks}
\usage{
Image_tile_deconstruction_function(
  Directory,
  Output_directory,
  RGB_Color_images = FALSE,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL,
  Tile_pixel_size,
  Tile_Overlap = 0,
  N_cores = 1
)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where images to be chopped into tiles are present.}

\item{Output_directory}{Character string specifying the path to the folder where  output images are written. It must be an empty folder.}

\item{RGB_Color_images}{Logical. Are we dealing with RGB color images?}

\item{Ordered_Channels}{Character vector specifying image channels. If RGB_Color_images is TRUE it will be set automatically to c("R", "G", "B").}

\item{Channels_to_keep}{Character vector specifying image channels to be kept after image chopping.}

\item{Tile_pixel_size}{Integer specifying the size of the tiles in pixels.}

\item{Tile_Overlap}{Integer specifying the overlap amount between tiles in pixels.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
The function writes the tiles in the ouput directory.
}
\description{
\code{Image_tile_deconstruction_function()} divides into smaller chunks all the images in a directory.
This allows faster computations when using CSM apps, performing image thresholding and cell segmentation.
The names of the resulting images should not be modified. If modified, this can
result in unexpected results when using companion functions \code{\link[=Image_from_tile_rebuilder]{Image_from_tile_rebuilder()}}, \code{\link[=Tile_to_image_cell_arrange_function]{Tile_to_image_cell_arrange_function()}}.
Tile overlap may be useful to avoid segmentation errors in cells close to the tile edge.
}
\details{
The function has been developed using tiff and ome.tiff image formats. The ome.tiff files have been less extensively tested than tiff files. Other formats (like png) may work but have not been tested.

Warning!! The function uses the RBioFormats Bioconductor package that parses the Bioformats Java library to R. Sometimes even if the user
has installed the RBioFormats package from bioconductor the function will still return an error. If this happens, the issue can be solved
checking the installed Java version in the computer and manually installing the CRAN rJava package.

If the resulting tiles are large images (due to tile size, or number of channels), the function may hit the maximum RAM usage capacity of any JAVA-based functions. In order to avoid this,
the user may need to open a fresh R session and before running any other code run \code{options(java.parameters = "-Xmx200g")} or \code{rJava::.jinit(parameters="-Xmx200g")} to increase max RAM
use to an absurdly high memory use (200Gb in this case).
}
\examples{
\dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
Output_Dir <- tempfile(pattern = "tempdir2_Output")
dir.create(Input_Dir, recursive = TRUE)
dir.create(Output_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Perform tiling---------------------------------------------------
Image_tile_deconstruction_function(
   Directory = Input_Dir,
   Output_directory = Output_Dir,
   Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3"),
   Channels_to_keep = c("DAPI", "GZMB", "CK-EPCAM", "CD8a"),
   RGB_Color_images = FALSE,
   Tile_pixel_size = 250,
   Tile_Overlap = 0,
   N_cores = 1
)

#Check that files have been created-----------------------------------------
list.files(Output_Dir)

#Remove directories---------------------------------------------------------
unlink(c(Input_Dir, Output_Dir), recursive = TRUE)
}

}
\seealso{
\code{\link[=Mask_conflict_resolver]{Mask_conflict_resolver()}}, \code{\link[=Image_from_tile_rebuilder]{Image_from_tile_rebuilder()}}
}
