% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap.R
\name{import_redcap}
\alias{import_redcap}
\title{Import the data from REDCap}
\usage{
import_redcap(token, url)
}
\arguments{
\item{token}{(character) a character string the specifies the token.}

\item{url}{(character) a character string the specifies the url.}
}
\value{
(list) a list with 2 data.frame. One data.frame contains the
data of the study, whereas the other contains three
columns with the name of each field, i.e. the names
of the variables in the first data.frame, the names of
the form associated to each field and the labels
associated to each field
}
\description{
Import the data from REDCap repository into a data.frame format.
API token and url are needed to retrieve the data.
}
\examples{

url     <- "https://bbmc.ouhsc.edu/redcap/api/"
token   <- "9A81268476645C4E5F03428B8AC3AA7B"
csm:::import_redcap(token = token, url = url)

}
