% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomplete_sheet.R
\name{incomplete_sheet}
\alias{incomplete_sheet}
\title{Return the patients with the incomplete (or unverified) sheet}
\usage{
incomplete_sheet(
  nested_tables,
  tab,
  field,
  redcap_info = c("record_id", "center")
)
}
\arguments{
\item{nested_tables}{(tibble) the output from \link{nest_tables}}

\item{tab}{(chr) the name of the sheet targeted by the check}

\item{field}{(chr) the name of the field which the targeted sheet
belongs to.}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled.}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble that identifies the
patients with incomplete
or unverified sheet
}
\description{
The function returns the patients whose form was marked as
incomplete or unverified.
}
\examples{
\dontrun{
  nested <- readr::read_rds(db_update_from_server())

  incomplete_sheet(
    nested, "demographics", "demographics and clinical",
    redcap_info = c("id", "center")
  )
}

}
