% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlier_check.R
\name{inlier_check}
\alias{inlier_check}
\title{Perform Inlier Check}
\usage{
inlier_check(data, min.part = 3, n, pervar = TRUE, title = "")
}
\arguments{
\item{data}{This should be a data frame with the id in the first column,    the site name/number in the second column, followed by the variables to check in the remaining columns.}

\item{min.part}{This is the minimum number of participants a site can have and still be tested. This MUST be at least 2 (though 5 or more is suggested). Default is 3.}

\item{n}{Number of SDs from the mean a point should lie before being counted as an inlier.}

\item{pervar}{TRUE or FALSE. Default is TRUE.Pervar is logical and should be set as TRUE if you wish to calculate the average distance from the mean (i.e. the sum of the distances divided by the number of variables used to calculate it) or FALSE if not.}

\item{title}{This provides the prefix of the title for each plot. Default is empty.}
}
\value{
Inlier check plot with inlier identified
}
\description{
The function is to detect inlier(s) for continuous variables.
}
\examples{
\dontrun{
inlier_check(DM_outlier, min.part=4, n=3, pervar=TRUE)
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

