% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_check.R
\name{integers_check}
\alias{integers_check}
\title{Perform Integer Check}
\usage{
integers_check(data, min_num = 3, title = "")
}
\arguments{
\item{data}{This should be in the form of a data frame with the site number in the first column (must be numeric-if a string is given please recode as numeric), the date that the measurement was taken in the second column (in the form dd/mm/yyyy) and the measurements to check in the following columns (i.e. columns 3+). If measurements were taken at different dates please split the data into more than one data frame and run the function on each data frame.}

\item{title}{This provides the prefix of the title for each plot. Default is empty.}

\item{min.part}{This is the minimum number of observations a site can have and still be tested. Default is 3.}
}
\value{
Plot with number of cumulative integer pattern
}
\description{
The function is to detect outlier(s) for continuous variables.
}
\examples{
\dontrun{
integers_check(data, min_num=3, title='')
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

