% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interaction_analyzer.R
\name{Interaction_analyzer}
\alias{Interaction_analyzer}
\title{Summarize images according to their cell-to-cell pairwise spatial interaction patterns}
\usage{
Interaction_analyzer(
  DATA = NULL,
  Exclude_non_significant = NULL,
  Cluster = TRUE,
  Max_N_Clusters = NULL,
  Consensus_reps = NULL,
  Consensus_p_Items = NULL,
  Consensus_Cluster_Alg = NULL,
  Consensus_Distance = NULL,
  Consensus_Name = NULL
)
}
\arguments{
\item{DATA}{A list of pairwise cell-to-cell spatial interactions computed using \code{\link[=Interaction_counter]{Interaction_counter()}}.}

\item{Exclude_non_significant}{A logical value indicating if non-significant spatial interactions should be removed from the analysis.}

\item{Cluster}{A logical value indicating if clustering should be performed.}

\item{Max_N_Clusters}{Number of maximum clusters that can be identified.}

\item{Consensus_reps}{Number of iterations to converge.}

\item{Consensus_p_Items}{Percentage of features that you desire to sample in each iteration.}

\item{Consensus_Cluster_Alg}{Clustering algorithm to be used (’hc’ hierarchical (hclust), ’pam’ for paritioning around medoids, ’km’ for k-means).}

\item{Consensus_Distance}{Distance metric to be used (pearson(1 - Pearson correlation), spearman(1 - Spearman correlation), euclidean, binary, maximum, canberra, minkowski.}

\item{Consensus_Name}{Name of the folder that is going to be created in order to place the resulting graphs.}
}
\value{
A list containing a summary of the pairwise cell-to-cell interactions by sample.
}
\description{
The function summarizes the results of the the pairwise cell-to-cell spatial interaction. In addition, the function identifies recurrent patterns of cell-to-cell spatial interactions in images. Before running the function pairwise spatial interactions need to be computed using \code{\link[=Interaction_counter]{Interaction_counter()}}.
}
\seealso{
\code{\link[=Interaction_counter]{Interaction_counter()}}
}
