% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interaction_counter.R
\name{Interaction_counter}
\alias{Interaction_counter}
\title{Analyzes the spatial interaction pattern between all cell types of all the images in the dataset}
\usage{
Interaction_counter(
  DATA,
  Phenotypes_included,
  N_cores = 1,
  Graph_type,
  K_number = NULL,
  Dist_threshold = NULL,
  Method,
  patch_size = NULL,
  Perform_significance_testing = TRUE,
  N_iterations = 10,
  p_threshold = 0.05
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Graph_type}{A character value indicating how to build the cell graph. One of the following: "expansion", "knn" or "delaunay".}

\item{K_number}{If Graph_type is knn, an integer value indicating the number of K nearest neighbors to build the graph.}

\item{Dist_threshold}{If Graph_type is expansion, a numeric value indicating the distance threshold.}

\item{Method}{A character value indicating the approach to spatial interaction quantification. One of the following: "classic", "histocat" or "patch" (see details).}

\item{patch_size}{If Method is "patch", a integer value indicating the size of the patch.}

\item{Perform_significance_testing}{A logical value indicating if significance testing should be performed.}

\item{N_iterations}{If significance testing needs to be performed, the number of iterations to calculate p values.}

\item{p_threshold}{If significance testing needs to be performed, the p value threshold for significance.}
}
\value{
A list containing pairwise cell spatial interactions by sample.
}
\description{
The function calculates the cell to cell spatial interaction pattern between cell type pairs. It is based on the imcRtools::countInteractions function.
Spatial interactions are quantified on graph representation of datasets and can be computed in different ways. Afterwards, recurrent spatial interaction patterns
can be identified in the dataset by running the \code{\link[=Interaction_analyzer]{Interaction_analyzer()}} function.
}
\details{
For additional information please visit imcRtools Vignette: https://www.bioconductor.org/packages/release/bioc/vignettes/imcRtools/inst/doc/imcRtools.html#5_Spatial_analysis.
}
\examples{
\dontrun{
Interaction_counter(
    DATA = CSM_Phenotypecell_test,
    Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype),
    N_cores = 1,

    Graph_type = "knn",
    K_number = 20

    Method = "classic",
    patch_size = 125,

    Perform_significance_testing = TRUE,
    N_iterations = 10,
    p_threshold = 0.05
)
}

}
\seealso{
\code{\link[=Interaction_analyzer]{Interaction_analyzer()}}
}
