% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_categorical_anomal.R
\name{is_categorical_anomal}
\alias{is_categorical_anomal}
\title{Check categorical impossible values}
\usage{
is_categorical_anomal(x, values)
}
\arguments{
\item{x}{(character) a character vector.}

\item{values}{(character) a character vector with the allowed values
for \code{x}.}
}
\value{
(logical) a logical vector that identifies which elements
of \code{x} has a value different from \code{values}.
}
\description{
Check if a categorical variable takes values that are different from
these allowed by the form of Redcap. The function returns a
logical that identifies which elements of the variable assume an
impossible.
}
\examples{

gender <- c("male", "female", "male", "male", "female", "x")
values <- c("male", "female")
is_categorical_anomal(x = gender, values = values)

}
