% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_continuous_anomal.R
\name{is_continuous_anomal}
\alias{is_continuous_anomal}
\title{Check continuous out-of-range values}
\usage{
is_continuous_anomal(x, range)
}
\arguments{
\item{x}{(numeric) a numeric vector.}

\item{range}{(numeric) a numeric vector with 2 values, i.e. the
minimum and the maximum allowed values for
\code{x}.}
}
\value{
(logical) a logical vector that identifies which element is
out of range.
}
\description{
Check if a continuous variable takes values that are outside the
range defined by two numerical values. The function returns logical
denoting which element of the vector is out of range.
}
\examples{

x <- rnorm(n = 10L, mean = 10, sd = 2)
range   <- c(3, 12)
is_continuous_anomal(x = x, range = range)

}
