% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LISA_Tumor_Stroma_identifier.R
\name{LISA_Tumor_Stroma_identifier}
\alias{LISA_Tumor_Stroma_identifier}
\title{Calculates tumor and stromal compartment based on Local Indicators of Spatial Association}
\usage{
LISA_Tumor_Stroma_identifier(
  DATA_Phenotypes,
  Index_phenotype,
  Image_preview = NULL,
  Association_Dist_min,
  Association_Dist_max,
  Type_of_LISA_function,
  Window_type,
  N_cores = 1
)
}
\arguments{
\item{DATA_Phenotypes}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Index_phenotype}{A character value indicating the cell phenotype to be used to calculate the tissue compartment.}

\item{Image_preview}{A character value indicating the name of the image to be used in the preview.}

\item{Association_Dist_min}{A numeric value indicating the minimum distance to test cell-cell spatial interaction.}

\item{Association_Dist_max}{A numeric value indicating the maximum distance to test cell-cell spatial interaction.}

\item{Type_of_LISA_function}{A character value indicating the type of LISA function used: Either 'L' or 'K'.}

\item{Window_type}{A character value indicating the interaction window shape: ’square’, ’convex’ or ’concave’.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
Returns a tibble with cell features and a column named 'Compartment' containing cell location.
}
\description{
The function calculates tissue compartments (usually tumor and stroma) based on Local Indicators of Spatial Association.
The function is based on the lisaClust::lisaClust function.
}
\examples{
\dontrun{
LISA_Tumor_Stroma_identifier(
   DATA_Phenotypes = CSM_Phenotypecell_test,
   Index_phenotype = "TUMOR",
   Association_Dist_min = 5,
   Association_Dist_max = 50,
   Type_of_LISA_function = "L",
   Window_type = "convex",
   N_cores = 1
)
}

}
