% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_factors.R
\name{make_factors}
\alias{make_factors}
\title{Make factors}
\usage{
make_factors(x, lev)
}
\arguments{
\item{x}{(data frame)}

\item{lev}{(named list) of levels. All the names must be included in
variables' names of \code{x}}
}
\value{
a new version of x, of the same class, with the required
variables mutated to factors
}
\description{
Variables of a data frame are transformed into factors. For each
variable that must be transformed as factor, a vector of the
corresponding levels labels is provided.
}
\examples{
library(tibble)

df <- tibble(
  foo = c("a", "a", "b"),
  bar = 1:3,
  baz = c("one", "two", "three"),
  qux = c("random", "sample", "text")
)

fct_levels <- list(
  foo = c("a", "b", "c", "d"),
  baz = c("one", "two", "three")
)

make_factors(df, fct_levels)
}
