% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker_combinator_generator.R
\name{Marker_combinator_generator}
\alias{Marker_combinator_generator}
\title{Generates all thresholded feature combinations present in a thresholded cell feature dataset}
\usage{
Marker_combinator_generator(DATA, Markers)
}
\arguments{
\item{DATA}{A dataframe or tibble containing thresholded cell feature data.}

\item{Markers}{A character vector indicating which features should be used.}
}
\value{
Returns a tibble where each row are unique positive feature combinations found in DATA.
}
\description{
Given a thresholded cell feature dataset, the function will find all unique combinations of positive features present. These combinations are actually cell phenotypes.
The result can then be used to feed the \code{\link[=Phenotype_assigner_function]{Phenotype_assigner_function()}} to obtain cell phenotypes.
}
\examples{
\dontrun{
#Threshold data-------------------------------------
DATA_thresholded <- Thresholding_function(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Strategy = "EBI_Otsu",
  Local_thresholding = FALSE,
  Method_autothreshold = "Otsu",
  number_iterations_TriClass = 20,
  Percentile = 0.5,
  Defined_threshold = 0.1,
  Levels = 3
)
#Find unique feature positivity combinations----------
Phenotype_possibilities <- Marker_combinator_generator(
  DATA = DATA_thresholded,
  Markers = names(DATA_thresholded)[-c(1:4)]
)
#Assign a phenotype to each combination------------
Phenotype_possibilities$Phenotype <-
  c('TUMOR', 'OTHER', 'CD8_GZMBneg', 'CD8_GZMBneg', 'OTHER', 'CD8_GZMBpos', 'CD8_GZMBpos')

#Perform phenotyping-------------------------------
Phenotype_assigner_function(
  DATA = DATA_thresholded,
  Phenotype_possibilities = Phenotype_possibilities
)
}

}
\seealso{
\code{\link[=Thresholding_function]{Thresholding_function()}}, \code{\link[=Thresholding_function_tailored]{Thresholding_function_tailored()}} \code{\link[=Phenotype_assigner_function]{Phenotype_assigner_function()}}
}
