% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker_segmentator.R
\name{Marker_segmentator}
\alias{Marker_segmentator}
\title{Segments any cell feature according to user provided cut-off values}
\usage{
Marker_segmentator(
  DATA,
  DATA_variable,
  DATA_cutoff,
  New_labels,
  Merge = FALSE,
  Var_to_Merge = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{DATA_variable}{A character indicating the feature name to be segmented.}

\item{DATA_cutoff}{A numeric vector specifying the cut-off points.}

\item{New_labels}{Character vector specifying the new labels (must be N of cut off - 1).}

\item{Merge}{A logical value indicating if the new variable should be merged with another variable.}

\item{Var_to_Merge}{(OPTIONAL) if Merge is TRUE, a character indicating the feature name to be combined with the segmented feature.}
}
\value{
Returns a tibble with the new segmented feature.
}
\description{
Given a numeric cell feature and user specified cut-off values, a character vector is generated to summarize the feature.
}
\examples{
\dontrun{
Marker_segmentator(
    DATA = CSM_DistanceToPixelcell_test,
    DATA_variable = "CK_DIST",
    DATA_cutoff = c(0, 1, 5, 500),
    New_labels = c("Inside", "Border", "Outside"),
    Merge = TRUE,
    Var_to_Merge = "Phenotype"
)
}

}
