% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mask_conflict_resolver.R
\name{Mask_conflict_resolver}
\alias{Mask_conflict_resolver}
\title{Solves conflicts between overlapping image tiles}
\usage{
Mask_conflict_resolver(
  N_cores = 1,
  Directory,
  Output_directory,
  Resolution_approach = "Remove_whole_cells",
  Priority_type = c("Dist_to_edge", "Cell_size"),
  Min_cell_size = NULL,
  Mask_refinement_path = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Directory}{Character specifying the path to the folder where cell segmentation masks are present. They must be tiled images with overlap.}

\item{Output_directory}{A character value indicating the path to the empty folder where the refined cell mask images should be written.}

\item{Resolution_approach}{A character value indicating how conflict between overlapping tiles should be solved. One of the following: Remove_whole_cells, Erode_pixels (see details).}

\item{Priority_type}{The type of metric to decide which cells or pixels should prevail after conflict identification. Either Dist_to_edge, Cell_size or both (see details). If both criteria are provided, the first one
will be used first, and if there is still a draw, the second one will be used.}

\item{Min_cell_size}{If Resolution_approach is Erode_pixels, an integer value that indicates the minimum pixels a cell must have to be kept in the analysis. Cells below threshold will be removed from the cell mask image.}

\item{Mask_refinement_path}{(Optional) If NULL, a temporary directory will be created to temporarily save erosion masks (see details). Alternatively, the user can provided the path to an empty folder were erosion masks should be writen.}
}
\value{
Writes the refined cell mask images in the output directory
}
\description{
The function is able to resolve conflicts between tiled images that have overlapping edges, reducing edge bias.
}
\details{
The function first identifies the amount of tile overlap. Afterwards identifies the conflict areas between tiles and solves them using to possible priority scores:
\itemize{
\item{Distance to edge (Dist_to_edge): If two cells or pixels collide, the pixel or cell that is closest to it's image edge is removed.}
\item{Cell size (Cell_size): If two cells or pixels collide, the one belonging to the smallest cell is removed.}
}

Once the losing cell or pixel has been identified then one of the following actions will be taken
\itemize{
\item{Removing the whole cell (Remove_whole_cells): The loser cell is removed.}
\item{Remove the pixel (Erode_pixels): The loser pixel is removed. Afterwards if the cell size is below the Min_cell_size threshold, the whole cell is removed.}
}
}
\examples{
\dontrun{

Mask_conflict_resolver(
  Directory = Path/to/Cell_mask/folder,
  Output_directory = Path/to/output/folder
)
}

}
\seealso{
\code{\link[=Image_tile_deconstruction_function]{Image_tile_deconstruction_function()}}
}
