% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_check.R
\name{mean_check}
\alias{mean_check}
\title{Perform Mean Check}
\usage{
mean_check(data, min)
}
\arguments{
\item{data}{A data frame with the subject ID in the first column; site ID in the second column; VISIT information in the third column; any continuous variables to check in the following columns.}

\item{min}{minimum number of observations a site can have and still be tested.}
}
\value{
figures chernoff face plots
}
\description{
The function is to investigate multiple continuous variables. It compares mean value for each site and visit using face plots.
}
\examples{
\dontrun{
mean_check(data,min=2)
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

