% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFI_calculator.R
\name{MFI_calculator}
\alias{MFI_calculator}
\title{Calculates the mean fluorescence intensity of the foreground pixels of an image}
\usage{
MFI_calculator(Target = NULL, Tissue_mask = NULL)
}
\arguments{
\item{Target}{An EBImage type object with a single channel to be thresholded}

\item{Tissue_mask}{An EBImage type object of the binary tissue mask}
}
\value{
A list containing the MFI and the total foreground pixels
}
\description{
Internal use only
}
\keyword{Internal}
