% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_mask_generator.R
\name{Multi_mask_generator}
\alias{Multi_mask_generator}
\title{Calculates a combined tissue mask combining different binary thresholded images}
\usage{
Multi_mask_generator(...)
}
\arguments{
\item{...}{EBImage type objects containing binary thresholded tissue masks}
}
\value{
The result of the intersection of all images provided
}
\description{
Internal use only
}
\keyword{Internal}
