% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighborhood_Quantifier.R
\name{Neighborhood_Quantifier}
\alias{Neighborhood_Quantifier}
\title{Summarizes cell neighborhoods for every image in the dataset}
\usage{
Neighborhood_Quantifier(DATA, Calculate_Density = FALSE, DATA_Area = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data and a column named 'Neighborhood_assignment'.}

\item{Calculate_Density}{A logical value indicating if density should be calculated.}

\item{DATA_Area}{If Calculate_Density is TRUE, a dataframe or tibble containing image names and tissue area information. It can be calculated using \code{\link[=Image_size_calculator]{Image_size_calculator()}}.}
}
\value{
Returns a tibble with neighborhood counts, percentages and optionally densities per image.
}
\description{
The function generates neighborhood counts, percentages and optionally neighborhood densities.
}
\examples{
\dontrun{
Neighborhood_Quantifier(
    DATA = CSM_Neighborhoods_test,
    Calculate_Density = FALSE
)
}

}
