% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighborhood_voting_function.R
\name{Neighborhood_voting_function}
\alias{Neighborhood_voting_function}
\title{Performs elections within tiled images with neighborhood information}
\usage{
Neighborhood_voting_function(
  N_cores = NULL,
  Tiled_Images = NULL,
  Minimum_cell_no_per_tile = NULL,
  Neighborhoods_included = NULL
)
}
\arguments{
\item{Minimum_cell_no_per_tile}{An integer indicating the minimum number of cells that a tile must contain. Tiles below the limit will not be included in the analysis.}

\item{Tiled_images}{A list containing tiled images obtained using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}} with neighborhood information.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}
}
\value{
A list containing image information with by-tile neighborhood counts and a tibble with a summary of the number of winner per image.
}
\description{
Given a tiled image, the function calculates neighborhood prevalence by tile. It then assigns neighborhood identity to each tile based on results.
}
