% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tables.R
\name{nest_tables}
\alias{nest_tables}
\title{Nest the tables of REDCap db}
\usage{
nest_tables(data, redcap_info)
}
\arguments{
\item{data}{(list) the output of \link{read_redcap}.}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled.}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble that contains a
column that identifies the
fields, a column that
identifies the sheets and
a column with list of
tibbles associated to each
field and sheet.
}
\description{
Create a tibble with nested tables for each field and its associated
sheets.
}
\examples{
library(tibble)

dd <- tibble(
  id = c(rep("1", 2L), rep("2", 2L), rep("3", 2L)),
  site = c(rep("AA", 4L), rep("BB", 2L)),
  fields = rep(c("demo and clinical", "discharge"), 3L),
  sex = c(
    "female", NA_character_,
    "male", NA_character_,
    "female", NA_character_
  ),
  age = c(
    44, NA_real_,
    50, NA_real_,
    71, NA_real_
  ),
  demo_complete = c(
    "complete", NA_character_,
    "complete", NA_character_,
    "complete", NA_character_
  ),
  hp = c(
    "yes", NA_character_,
    "no", NA_character_,
    "yes", NA_character_
  ),
  diabetes = c(
    "no", NA_character_,
    "no", NA_character_,
    "yes", NA_character_
  ),
  clinical_complete = c(
    "complete", NA_character_,
    "complete", NA_character_,
    "complete", NA_character_
  ),
  fe = c(
    NA_real_, 35,
    NA_real_, 56,
    NA_real_, 44
  ),
  death = c(
    NA_character_, "no",
    NA_character_, "no",
    NA_character_, "yes"
  ),
  discharge_complete = c(
    NA_character_, "complete",
    NA_character_, "complete",
    NA_character_, "complete"
  )
)

nest_tables(data = dd, redcap_info = c("id", "site"))
}
