% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_function_parallel.R
\name{Normalization_function_parallel}
\alias{Normalization_function_parallel}
\title{Normalizes cell feature expression data using multicore}
\usage{
Normalization_function_parallel(DATA, Strategy, Parameters, N_cores = 1)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Strategy}{'mxnorm' or 'simpleSeg'.}

\item{Parameters}{List of parameters to be used in the normalization process (see details).}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
Returns a tibble containing normalized cell feature data.
}
\description{
Normalization of cell feature data following two potential strategies included in mxnorm and simpleSeg packages.
Parallelizing the computation can speed up the process in huge datasets (in the order of 10^7 cells). In smaller datasets there is no improvement or it can be counterproductive.
}
\details{
For mxnorm normalization, the list must contain the following elements:
slide_id (The column name of the DATA that contains the slide ID. this can be the Subject_Names or a user supplied vector containing the slide identity of these cells),
image_id (The column name of the DATA that contains the image ID. Can be the same as slide_id),
marker_cols (The name of the columns to be normalized),
transform ("None", "log10", "mean_divide","log10_mean_divide"),
method ("None", "ComBat","Registration"),
method_override (OPTIONAL) ("specify a custom normalization method. See mxnorm package documentation for more details),
method_override_name (OPTIONAL) (specify the name of the user defined method).

For simpleSeg approach, the list must contain the following elements:
cells (The tibble containing the cell feature data),
markers (The name of the columns to be normalized),
imageID (The column name of the DATA that contains the image ID),
transformation (Choose between:  NULL, "asinh", "sqrt"),
method ("trim99", "minMax"),
cores (Irrelevant argument. If parallelization is required use N_cores argument and \code{\link[=Normalization_function_parallel]{Normalization_function_parallel()}} function)
}
\seealso{
\code{\link[=Normalization_function]{Normalization_function()}}, \code{\link[=Normalization_diagnostics]{Normalization_diagnostics()}}
}
