% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_aedata.R
\name{outlier_aedata}
\alias{outlier_aedata}
\title{Perform Outlier Check for AE}
\usage{
outlier_aedata(data0, data1, n, method, var.name = "AE_Rate", title = "")
}
\arguments{
\item{data0}{This should be an AE dataframe with the site number in the first column; Subject ID in the second column; Each row represents a different AE.}

\item{data1}{This should be an subject level dataframe with the site number in the first column; Subject ID in the second folumn; patient's enrollment/first dose date in \%m/\%d/\%Y format in the third column; patient's last date of AE collection in \%m/\%d/\%Y format in the forth column; Each row represents a different subject.}

\item{n}{the number of SD (or IQR) from the mean that values can fall and be considered extreme (outliers) OR (for the Mahalanobis distance only) used to find the critical value on the Chi squared distribution. i.e. for the top 2.5\% n should be set as 0.975. For percentage method, it shows outliers fall outside of top and lower certain percentage values,  i.e. for the top 2.5\% n should be set as 0.975.}

\item{method}{should be one of 5 options: Select from c("sd", "grubbs","euclid","mahal" ,"IQR","PCT"). "sd" - the sd method chooses value which lies more than n SDs from the mean as extreme； "grubbs" - very similar to the SD method but here, only the most extreme value is chosen, it is then removed and the mean and sd recalculated. The process continues until all extreme values have been found. The idea is that the most extreme values won't mask smaller extreme values by inflating the mean and sd. "euclid" - finds multivariate outliers using the Euclidian distance. "mahal" - finds multivariate outliers using the Mahalanobis distance. "IQR" - to be used with non-normally distributed data. Outliers are points which lie more than n times the IQR from the median. PCT - Percentage method shows outliers fall outside of top and lower certain percentage values.}

\item{var.name}{should be a variable name with no space or special character in between. Default is 'AE_Rate'}

\item{title}{This provides the prefix of the title for each plot. Default is empty.}
}
\value{
outliers for event incidence data
}
\description{
The function is to check the existence of outliers accounting for time (/month) and number of subjects (/person) in each site for continuous variables.
}
\examples{
\dontrun{
outlier_aedata(data0,data1,n=.95, method='PCT',var.name='AE_Rate',title='')
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

