% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier_check}
\alias{outlier_check}
\title{Perform Outlier Check}
\usage{
outlier_check(data, n, method, normal.plot, title = "")
}
\arguments{
\item{data}{This should be a data frame with the id in the first column,    the site name/number in the second column, followed by the variables to check in the remaining columns.}

\item{n}{the number of SD (or IQR) from the mean that values can fall and be considered extreme (outliers) OR (for the Mahalanobis distance only) used to find the critical value on the Chi squared distribution. i.e. for the top 2.5\% n should be set as 0.975. For percentage method, it shows outliers fall outside of top and lower certain percentage values,  i.e. for the top 2.5\% n should be set as 0.975.}

\item{method}{should be one of 5 options: Select from c("sd", "grubbs","euclid","mahal" ,"IQR","PCT"). "sd" - the sd method chooses value which lies more than n SDs from the mean as extreme； "grubbs" - very similar to the SD method but here, only the most extreme value is chosen, it is then removed and the mean and sd recalculated. The process continues until all extreme values have been found. The idea is that the most extreme values won't mask smaller extreme values by inflating the mean and sd. "euclid" - finds multivariate outliers using the Euclidian distance. "mahal" - finds multivariate outliers using the Mahalanobis distance. "IQR" - to be used with non-normally distributed data. Outliers are points which lie more than n times the IQR from the median. PCT - Percentage method shows outliers fall outside of top and lower certain percentage values.}

\item{normal.plot}{This declares whether a normal probability plot should be drawn for each variable. It should be set as TRUE if plots are required and FALSE if not.}

\item{title}{This provides the prefix of the title for each plot. Default is empty.}
}
\value{
outlier for continuous data
}
\description{
The function is to detect outlier(s) for continuous variables.
}
\examples{
\dontrun{
outlier_check(data, n = 2, method = "sd",  normal.plot = FALSE,"Plot 1. ")
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

