% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Own_grid_metrics.R
\name{Own_grid_metrics}
\alias{Own_grid_metrics}
\title{Interal function to RUN SPIAT safely and without bugs}
\usage{
Own_grid_metrics(spe_object, FUN, n_split, ...)
}
\arguments{
\item{spe_object}{Argument}

\item{FUN}{Argument}

\item{n_split}{Argument}

\item{...}{Argument}
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=SPIAT_Heterogeneity_Analyzer]{SPIAT_Heterogeneity_Analyzer()}}
}
\keyword{Internal}
