% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype_quantifier.R
\name{Phenotype_quantifier}
\alias{Phenotype_quantifier}
\title{Summarizes cell phenotypes for every image in a dataset}
\usage{
Phenotype_quantifier(DATA = NULL, Calculate_Density = NULL, DATA_Area = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data that and a column named 'Phenotype' containing cell labels.}

\item{Calculate_Density}{A logical value indicating if density should be calculated.}

\item{DATA_Area}{If Calculate_Density is TRUE, a dataframe or tibble containing image names and tissue area information.}
}
\value{
Returns a tibble with cell counts, pertentages and optionally cell densities per image.
}
\description{
The function generates positive cell counts, percentages and optionally cell densities according to cell phenotypes.
}
