% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping_evaluator_shiny_app_launcher.R
\name{Phenotyping_evaluator_shiny_app_launcher}
\alias{Phenotyping_evaluator_shiny_app_launcher}
\title{Launches a shiny App to explore cell phenotyping results}
\usage{
Phenotyping_evaluator_shiny_app_launcher(
  DATA,
  Directory,
  Ordered_Channels,
  Max_Gb_cache = 10
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data and a column named 'Phenotype' containing cell phenotype labels.}

\item{Directory}{Character specifying the path to the folder where images are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{Max_Gb_cache}{A single numeric value indicating the memory size of the image cache (see details). 10 Gb by default.}
}
\description{
After phenotyping, cell phenotypes can be explored interactively proyected on original images.
}
\details{
In order to deal with large images and speed up image toggling, the shiny APP works with a memoised version of the \code{\link[=Smart_image_importer]{Smart_image_importer()}} function. Loaded images will
be stored in a temporary cache until the APP is closed or the cache reaches it's limit.

Image parameters control the image and channel to be displayed.

User can use the 'Color', 'Add channel', 'Remove channel' and 'Reset' buttons to merge different channels into a single image. The add channel button
will color the current channel according to the color provided and will be merged to the current image.

Sample summary table is depicted at the bottom of the control panel.

Image Panels:
\itemize{
\item{Upper left: Image being explored. Use it to zoom in.}
\item{Upper right: Cell phenotype labels overlayed. Cells of interest can be selected by clicking or by the area selection tools.}
\item{Lower left: Cell selection summary.}
\item{Lower right: Heatmap. If any cells are selected, a heatmap of the relative feature expression will be generated.}
}
}
\examples{
\dontrun{
#Create temporary directory------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Deploy app----------------------------------------------
Phenotyping_evaluator_shiny_app_launcher(
    DATA = CSM_Phenotypecell_test,
    Directory = Input_Dir,
    Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3")
)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
}


}
