% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pixel_Multilevel_thresholder.R
\name{Pixel_Multilevel_thresholder}
\alias{Pixel_Multilevel_thresholder}
\title{Thresholds the foreground pixels of an image according to user parameters into several levels}
\usage{
Pixel_Multilevel_thresholder(
  Target = NULL,
  Tissue_mask = NULL,
  Threshold_values = NULL,
  Blurr = NULL,
  Sigma = NULL
)
}
\arguments{
\item{Target}{An EBImage type object with a single channel to be thresholded}

\item{Tissue_mask}{An EBImage type object of the binary tissue mask}

\item{Threshold_values}{The value applied to threshold the image}

\item{Blurr}{A logical value indicating if blurring should be applied before thresholding}

\item{Sigma}{Sigma applied to the Gaussian kernel for blurring}

\item{Threshold_type}{Type of threshold method to be performed: 'Arbitrary' or 'Otsu'}

\item{Threshold_value}{If Threshold_type = 'Arbitrary', the value applied to threshold the image}
}
\value{
An EBImage object containing the binary thresholded image

An EBImage object containing the multi-level thresholded image
}
\description{
Internal use only
}
\keyword{Internal}
\keyword{Thresholds}
\keyword{according}
\keyword{an}
\keyword{foreground}
\keyword{image}
\keyword{into}
\keyword{levels}
\keyword{of}
\keyword{only}
\keyword{parameters}
\keyword{pixels}
\keyword{several}
\keyword{the}
\keyword{to}
\keyword{use}
\keyword{user}
