% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_correlation_matrix.R
\name{Plot_correlation_matrix}
\alias{Plot_correlation_matrix}
\title{Draws a correlation matrix}
\usage{
Plot_correlation_matrix(
  DATA = NULL,
  Variables_included = NULL,
  Correlation_method = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Variables_included}{A character vector indicating the names of the features to be used.}

\item{Correlation_method}{Either "pearson" or "spearman".}
}
\value{
Generates a correlation matrix plot
}
\description{
The function calculates correlation between variables and draws a correlation matrix. Highly correlated or anti-correlated variables can cause trouble in subsequent analyses.
Therefore, checking for correlation is a good practice before proceeding with analyses.
}
