% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_Tumor_Stroma_identifier.R
\name{Quick_Tumor_Stroma_identifier}
\alias{Quick_Tumor_Stroma_identifier}
\title{Calculates tumor and stromal compartment based on tiles}
\usage{
Quick_Tumor_Stroma_identifier(
  DATA_Phenotypes = NULL,
  Index_phenotype = NULL,
  Accuracy = NULL,
  Min_cell_no = NULL,
  Image_preview = NULL,
  N_cores = NULL
)
}
\arguments{
\item{Index_phenotype}{A character value indicating the cell phenotype to be used to calculate the tissue compartment.}

\item{Accuracy}{A numeric value indicating the size of the tile. Smaller sizes give more accurate result.}

\item{Min_cell_no}{A integer value indicating the minimum number of cells within a tile to consider the tile positive for a compartment.}

\item{Image_preview}{A character value indicating the name of the image to be used in the preview.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}
}
\value{
Returns a tibble with cell features and a column named 'Compartment' containing cell location.
}
\description{
The function calculates tissue compartments (usually tumor and stroma) based on tile occupation by tumor cells.
}
