% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet_basic.R
\name{quiet_basic}
\alias{quiet_basic}
\title{Interal function to RUN SPIAT safely and without bugs}
\usage{
quiet_basic(x)
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Own_grid_metrics]{Own_grid_metrics()}}, \code{\link[=SPIAT_Heterogeneity_Analyzer]{SPIAT_Heterogeneity_Analyzer()}}
}
\keyword{Internal}
