% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_image_tiling_processing_function.R
\name{Random_image_tiling_processing_function}
\alias{Random_image_tiling_processing_function}
\title{Performs image tiling according to randomly iterating parameters}
\usage{
Random_image_tiling_processing_function(
  N_cores = 1,
  DATA,
  Tile_width,
  Tile_height,
  Variables_to_keep,
  N_iterations = 10,
  Max_tile_size_variation = 1.5,
  Force_squared_tiles = FALSE
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Tile_width}{A numeric value indicating the reference width of the tiles.}

\item{Tile_height}{A numeric value indicating the reference height of the tiles.}

\item{Variables_to_keep}{A character vector indicating the names of the columns to be kept when tiling the image.}

\item{N_iterations}{A integer value indicating the number of random parameter combinations that will generate tiled images.}

\item{Max_tile_size_variation}{A numeric value >= 1, indicating the max ratio of tile size that will be used in the random parameters.}

\item{Force_squared_tiles}{A logical value indicating if tiles width and height should be forced to be equal in every random parameter iteration.}
}
\value{
A list containing the parameters used to generate the tiled images and the actual results.
}
\description{
Performs image tiling according to randomly iterating parameters
}
\details{
The function is aimed to estimate impact of the Modifiable areal unit problem (MAUP) (please see the 'Calculating tissue heterogeneity' Vignette for more details).
The function will generate tiled information for every image randomly iterating certain image parameters (tile dimension, cell coordinate angle rotating and position flipping).
These tiled images can be used to compute the influence of the MAU on spatial heterogeneity analysis (using the \code{\link[=Random_tiled_image_heterogeneity_calculator]{Random_tiled_image_heterogeneity_calculator()}} and
\code{\link[=Random_texture_features_calculator]{Random_texture_features_calculator()}} functions).

The first two random iterations will always contain tiling strategies in the extremes of the user defined tile size variation factor.
}
\examples{
\dontrun{
#Generate tiled image data using 10 randomly generated parameter set---------
Random_tile_DATA <-
  Random_image_tiling_processing_function(
    N_cores =  1,
    DATA = CSM::CSM_Phenotypecell_test,
    Tile_width = 100,
    Tile_height = 100,
    Variables_to_keep = "Phenotype",

    N_iterations = 10,
    Max_tile_size_variation = 1.1,
    Force_squared_tiles = TRUE
  )

#Compute the heterogeneity by tile using this random tile dataset------------
Tiled_heterogeneity_DATA <-
  Random_tiled_image_heterogeneity_calculator(
    N_cores = 1,
    Random_tiled_images = Random_tile_DATA,
    Minimum_cell_no_per_tile = 1,
    Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype)
  )

#Analyze the data-----------------------------------------------------------
Random_tiled_image_heterogeneity_analyzer(
    Random_tiled_heterogeneity_DATA = Tiled_heterogeneity_DATA,
    Strategy = "Overall_Summary",
    Metric = "Shannon"
  )
}


}
\seealso{
\code{\link[=Random_tiled_image_heterogeneity_calculator]{Random_tiled_image_heterogeneity_calculator()}}, \code{\link[=Random_tiled_image_heterogeneity_analyzer]{Random_tiled_image_heterogeneity_analyzer()}}, \code{\link[=Random_texture_features_calculator]{Random_texture_features_calculator()}}
}
