% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_texture_features_calculator.R
\name{Random_texture_features_calculator}
\alias{Random_texture_features_calculator}
\title{Calculates the average texture features of a single cell type using tiled images generated with random parameters}
\usage{
Random_texture_features_calculator(
  N_cores = 1,
  Random_tiled_images,
  Phenotype_included
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Random_tiled_images}{A list containing tiled images obtained using \code{\link[=Random_image_tiling_processing_function]{Random_image_tiling_processing_function()}}.}

\item{Phenotype_included}{A character value indicating which cell phenotype to analyze.}
}
\value{
A list containing the parameters used to generate the tiled images and the actual results.
}
\description{
The function is aimed to estimate impact of the Modifiable areal unit problem (MAUP). The function calculates the average texture features in a similar
way as it's homologue function \code{\link[=Texture_features_calculator]{Texture_features_calculator()}}.
}
\examples{
\dontrun{

#Generate tiled image data using 10 randomly generated parameter set
Random_tile_DATA <-
  Random_image_tiling_processing_function(
    N_cores =  1,
    DATA = CSM::CSM_Phenotypecell_test,
    Tile_width = 100,
    Tile_height = 100,
    Variables_to_keep = "Phenotype",

    N_iterations = 10,
    Max_tile_size_variation = 1.1,
    Force_squared_tiles = TRUE
  )

#Compute texture features in the dataset
Random_texture_features_calculator(
          N_cores = 1,
          Random_tiled_images = Random_tile_DATA,
          Phenotype_included = "CD8_GZMBneg"
  )
}

}
\seealso{
\code{\link[=Random_image_tiling_processing_function]{Random_image_tiling_processing_function()}}, \code{\link[=Random_tiled_image_heterogeneity_calculator]{Random_tiled_image_heterogeneity_calculator()}}, \code{\link[=Random_tiled_image_heterogeneity_analyzer]{Random_tiled_image_heterogeneity_analyzer()}},
}
