% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_tiled_image_heterogeneity_calculator.R
\name{Random_tiled_image_heterogeneity_calculator}
\alias{Random_tiled_image_heterogeneity_calculator}
\title{Calculates heterogeneity by tile using tiled images generated with random parameters}
\usage{
Random_tiled_image_heterogeneity_calculator(
  N_cores = 1,
  Random_tiled_images,
  Minimum_cell_no_per_tile = 1,
  Phenotypes_included
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Random_tiled_images}{A list containing tiled images obtained using \code{\link[=Random_tiled_image_heterogeneity_calculator]{Random_tiled_image_heterogeneity_calculator()}}.}

\item{Minimum_cell_no_per_tile}{An integer indicating the minimum number of cells that a tile must contain. Tiles below the limit will not be included in the analysis.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}
}
\value{
A list containing the parameters used to generate the tiled images and the actual results.
}
\description{
Calculates heterogeneity by tile using tiled images generated with random parameters
}
\details{
The function is aimed to estimate impact of the Modifiable areal unit problem (MAUP). For a given set of tiled images generated using random parameters, the function will compute by-tile heterogeneity metrics in a similar
way as it's homologue function \code{\link[=Tiled_image_heterogeneity_calculator]{Tiled_image_heterogeneity_calculator()}}.
}
\examples{
\dontrun{
#Generate tiled image data using 10 randomly generated parameter set---------
Random_tile_DATA <-
  Random_image_tiling_processing_function(
    N_cores =  1,
    DATA = CSM::CSM_Phenotypecell_test,
    Tile_width = 100,
    Tile_height = 100,
    Variables_to_keep = "Phenotype",

    N_iterations = 10,
    Max_tile_size_variation = 1.1,
    Force_squared_tiles = TRUE
  )

#Compute the heterogeneity by tile using this random tile dataset------------
Tiled_heterogeneity_DATA <-
  Random_tiled_image_heterogeneity_calculator(
    N_cores = 1,
    Random_tiled_images = Random_tile_DATA,
    Minimum_cell_no_per_tile = 1,
    Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype)
  )

#Analyze the data-----------------------------------------------------------
Random_tiled_image_heterogeneity_analyzer(
    Random_tiled_heterogeneity_DATA = Tiled_heterogeneity_DATA,
    Strategy = "Overall_Summary",
    Metric = "Shannon"
  )
}


}
\seealso{
\code{\link[=Random_image_tiling_processing_function]{Random_image_tiling_processing_function()}}, \code{\link[=Random_tiled_image_heterogeneity_analyzer]{Random_tiled_image_heterogeneity_analyzer()}}, \code{\link[=Random_texture_features_calculator]{Random_texture_features_calculator()}}
}
