% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_last_numbers.R
\name{remove_last_number}
\alias{remove_last_number}
\title{Remove last numbers}
\usage{
remove_last_number(x)
}
\arguments{
\item{x}{(chr) a character vector}
}
\value{
(chr) a character vector
}
\description{
The function takes a character vector and it removes the last two
positions of the element, if the an element ends with a "_" and a
number.
}
\examples{

a <- c("aa", "bb", "cc", "dd_1", "dd_2", "ff")

remove_last_number(x = a)

}
