% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render a report for each center}
\usage{
render_report(
  token,
  path_data = data_path(),
  redcap_info = c("record_id", "center"),
  file_name = "tidy_data.rds",
  center_names,
  center_labels,
  path_template
)
}
\arguments{
\item{token}{(chr) a character string that specifies the token to
get access to the data of the study. This will
be passed to \link{db_update_from_server}.}

\item{path_data}{(chr) path for data folder (default is \verb{data/}
under the current project or the
\link[here:here]{here} path. If the folder does
not exist, it will be created). This will
be passed to \link{db_update_from_server}.}

\item{redcap_info}{(character) the names of the variables that
identify the subject and the site
where the subject was enrolled. This
will be passed to
\link{db_update_from_server}.}

\item{file_name}{(chr) file name in which data are stored. This will
be passed to \link{db_update_from_server}.}

\item{center_names}{(chr) a character vector whose elements are the
names of the center as labelled in the
data.}

\item{center_labels}{(chr) a character vector whose elements are the
names that will be given to the html
file containing the report.}

\item{path_template}{(chr) the path and the name of the files that
contains the template for the report.}
}
\description{
The function renders a report for the data quality monitoring for
each center. A template (which must be a .Rmd file) and a list of
centers must be provided.
}
\details{
The function aims to automate the implementation of
data quality report for each center involved in the study.
It works as follows:
- An updated version of the dataset is retrieved from the
server
- A list of nested_tables (see \link{nest_tables}) containing
the data for each center is created
- A report file is rendered for each element of the list
created in the previous step. One html file will be
created for each element of the list.
\strong{Note}: in the template, the object that contains the
nested table of the center must be named as \code{db}. Moreover,
the object that ontains the meta-data of the study must
be names as \code{meta_data} in the template.
}
\examples{

\dontrun{
render_report()
}

}
