% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ripley_function_calculator.R
\name{Ripley_function_calculator}
\alias{Ripley_function_calculator}
\title{Analyzes the spatial interaction using Ripleys K and L functions}
\usage{
Ripley_function_calculator(
  DATA = NULL,
  Cell_type = NULL,
  Max_distance = NULL,
  Strategy = NULL,
  N_simulations = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Cell_type}{A character value indicating the cell phenotype to be analyzed.}

\item{Max_distance}{A numeric value indicating the maximum distance to be analyzed.}

\item{Strategy}{A character value indicating the function type: either "Ripleys_K" or "Ripleys_L"}

\item{N_simulations}{A integer value indicating the number of simulations to calculate the random background.}
}
\value{
A tibble containing results by image.
}
\description{
The function calculates the spatial aggregation of a single cell type using Ripleys K and L functions. Function results are integrated to obtain an AUC,
that can serve as a metric of spatial interaction. The function is calculated using the spatstat R metapackage.
(specifically the spatstat.explore::envelope). If integration fails, the user may try different Max_distance thresholds to make it work.
}
\examples{
\dontrun{
Ripley_function_calculator(
   DATA = CSM_Phenotypecell_test,
   Cell_type = "TUMOR",
   Max_distance = 15,
   Strategy = "Ripleys_K",
   N_simulations = 10
)
}

}
