% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_center.R
\name{select_center}
\alias{select_center}
\title{Select the data of a center}
\usage{
select_center(nested_tables, center_id = "center", center_label)
}
\arguments{
\item{nested_tables}{(data.frame) the output of the function
\link{nest_tables}}

\item{center_id}{(character) the name of the column that identifies
the centers.}

\item{center_label}{(character) the label of the center.}
}
\value{
the nested tables \link[tibble:tibble-package]{tibble} with the
data of just one center.
}
\description{
The function filters the rows of the nested tables returned by
\link{nest_tables} that correspond to a specific center.
}
\details{
The function aims to retrieve one dataset for each center.
The single dataset will be used as basis to implement
the monitor of data quality.
}
\examples{

nested_tab <- tibble::tibble(
   fields = c(rep("demo_clinical", 3L), rep("follow_up", 3L)),
   sheets = c(
       "demo", "risk_factors", "clinical", "discharge",
       "month_1", "year_1"
   ),
   tables = purrr::map(
       .x = seq_len(6),
       ~ tibble::tibble(
           id = rep(glue::glue("id_{1:5}"), 2L),
           center = rep(c("center_1", "center_2"), each = 5L),
           x = c(0.1, 0.2, 0.15, 0.4, 0.5, 0.6, 0.3, 0.2, 0.2, 0.1),
           y = c(0.1, 0.11, 0.12, 0.3, 0.3, 0.3, 0.4, 0.5, 0.6, 0.55),
           z = c(0.1, 0.2, 0.15, 0.4, 0.5, 0.6, 0.3, 0.2, 0.2, 0.1)
       )
   )
)

select_center(nested_tab, "center", "center_1")

}
