% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHINY_Thresholding_summary_function.R
\name{SHINY_Thresholding_summary_function}
\alias{SHINY_Thresholding_summary_function}
\title{Generates a summary of threshold results designed to work with dedicated shiny app}
\usage{
SHINY_Thresholding_summary_function(
  DATA = NULL,
  DATA_Thresholded = NULL,
  LOCAL = NULL,
  CASE = NULL
)
}
\arguments{
\item{DATA}{Dataframe or tibble containing cell features.}

\item{DATA_Thresholded}{Dataframe or tibble containing thresholded cell features.}

\item{LOCAL}{Logical value indicating if thresholds are calculated locally.}

\item{CASE}{A character indicating which image needs to be summarized.}
}
\value{
A list containing the histogram, the cut-off values and the cells above and below threshold for the selected case
}
\description{
Intended for internal use only
}
\keyword{Internal}
