% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT_entropy_gradient_generator.R
\name{SPIAT_entropy_gradient_generator}
\alias{SPIAT_entropy_gradient_generator}
\title{Analyzes the spatial interaction between cell types according to the SPIAT package}
\usage{
SPIAT_entropy_gradient_generator(
  DATA_SPIAT = NULL,
  Gradient_start = NULL,
  Gradient_stop = NULL,
  Gradient_sampling = NULL,
  Phenotypes_included = NULL
)
}
\arguments{
\item{DATA_SPIAT}{A list containing SPIAT objects generated using \code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.}

\item{Gradient_start}{A numeric value indicating where should the sampling process start.}

\item{Gradient_stop}{A numeric value indicating where should the sampling process end.}

\item{Gradient_sampling}{A numeric value indicating the distance to be sampled.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}
}
\value{
A tibble containing a summary of the analysis result by image.
}
\description{
The function calculates cell interaction pattern based on entropy gradients. For more information see related publication: https://doi.org/10.1038/s41467-023-37822-0.
}
\examples{
#Generate SPIAT object list----------------------------
DATA_SPIAT <-
SPIAT_object_generator(
    DATA_Intensities = CSM_Arrangedcellfeaturedata_test,
    DATA_Phenotypes = CSM_Phenotypecell_test
)

#Calculate entropy gradients--------------------------
SPIAT_entropy_gradient_generator(
    DATA_SPIAT = DATA_SPIAT,
    Gradient_start = 0,
    Gradient_stop = 250,
    Gradient_sampling = 50,
    Phenotypes_included = c("CD8_GZMBpos", "TUMOR")
)

}
\seealso{
\code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.
}
