% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT_neighborhood_identifier.R
\name{SPIAT_neighborhood_identifier}
\alias{SPIAT_neighborhood_identifier}
\title{Identifies cell neighborhoods using the pipeline present in the SPIAT package}
\usage{
SPIAT_neighborhood_identifier(
  N_cores = NULL,
  DATA_SPIAT = NULL,
  Strategy = NULL,
  Cell_types_of_interest = NULL,
  Radius = NULL,
  Min_neighborhood_size = NULL,
  K_phenograph = NULL,
  No_Phenotype_name = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA_SPIAT}{A list containing SPIAT objects generated using \code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.}

\item{Strategy}{A character value indicating the method to find the clusters. One of "hierarchical" or "dbscan".}

\item{Cell_types_of_interest}{A character vector indicating the cell phenotype labels included in the analysis.}

\item{Radius}{A numeric value indicating the distance radius used to calculate neighborhoods.}

\item{Min_neighborhood_size}{A integer value indicating the minimum number of cells to constitute a neighborhood.}

\item{No_Phenotype_name}{(OPTIONAL) A character value indicating the cell phenotype label corresponding to non-phenotyped cells.}
}
\value{
A tibble containing a summary of the analysis result by image.
}
\description{
The function identifies cell neighborhoods using the SPIAT::identify_neighborhoods function.
}
\seealso{
\code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.
}
