% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT_object_generator.R
\name{SPIAT_object_generator}
\alias{SPIAT_object_generator}
\title{Generates SPIAT objects for every image in a dataset}
\usage{
SPIAT_object_generator(DATA_Intensities = NULL, DATA_Phenotypes = NULL)
}
\arguments{
\item{DATA_Intensities}{A dataframe or tibble containing cell feature data with feature expression by cell.}

\item{DATA_Phenotypes}{A dataframe or tibble containing cell feature data with a column named 'Phenotype' containing cell phenotype labels.}
}
\value{
A list with SPIAT objects.
}
\description{
The function transforms image information into a SPIAT object that can be used in subsequent analysis that depend on SPIAT package, like
\code{\link[=SPIAT_Heterogeneity_Analyzer]{SPIAT_Heterogeneity_Analyzer()}}, \code{\link[=SPIAT_entropy_gradient_generator]{SPIAT_entropy_gradient_generator()}}, \code{\link[=SPIAT_Tissue_structuring_function]{SPIAT_Tissue_structuring_function()}} and \code{\link[=SPIAT_neighborhood_identifier]{SPIAT_neighborhood_identifier()}}.
}
\examples{
SPIAT_object_generator(
    DATA_Intensities = CSM_Arrangedcellfeaturedata_test,
    DATA_Phenotypes = CSM_Phenotypecell_test
)

}
\seealso{
\code{\link[=SPIAT_Heterogeneity_Analyzer]{SPIAT_Heterogeneity_Analyzer()}}, \code{\link[=SPIAT_entropy_gradient_generator]{SPIAT_entropy_gradient_generator()}}, \code{\link[=SPIAT_Tissue_structuring_function]{SPIAT_Tissue_structuring_function()}},\code{\link[=SPIAT_neighborhood_identifier]{SPIAT_neighborhood_identifier()}}.
}
