% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT_Tissue_structuring_function.R
\name{SPIAT_Tissue_structuring_function}
\alias{SPIAT_Tissue_structuring_function}
\title{Generates tissue structures according to a cell phenotype label using the SPIAT package.}
\usage{
SPIAT_Tissue_structuring_function(
  N_cores = NULL,
  DATA_SPIAT = NULL,
  DATA_Phenotypes = NULL,
  Cell_type_to_define_cluster = NULL,
  Minimum_number_cells_cluster = NULL,
  Cell_types_of_interest = NULL,
  Layers_margin = NULL,
  Simplify_result = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA_SPIAT}{A list containing SPIAT objects generated using \code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.}

\item{DATA_Phenotypes}{A dataframe or tibble containing cell feature data with a column named 'Phenotype' containing cell phenotype labels.}

\item{Cell_type_to_define_cluster}{A character value indicating which cell phenotype label should be used  to build tissue structures.}

\item{Minimum_number_cells_cluster}{A integer value indicating the minimum number of cells required to be considered a structure.}

\item{Cell_types_of_interest}{Cell types to be related to these structures.}

\item{Simplify_result}{A logical value indicating if the final result should be simplified.}
}
\value{
If Simplify_result is TRUE, the function returns a tibble with simplify cell location labels.
If Simplify_result is FALSE, the function returns a list containing cell location labels for every image.
}
\description{
The function generates tissue structure according to a cell phenotype label (usually tumor cells). Then all the cells are asigned a label according to their distance to the tissue structures.
}
\examples{
#Generate SPIAT object list----------------------------
DATA_SPIAT <-
SPIAT_object_generator(
    DATA_Intensities = CSM_Arrangedcellfeaturedata_test,
    DATA_Phenotypes = CSM_Phenotypecell_test
)

#Calculate tissue structures-------------
SPIAT_Tissue_structuring_function(
    N_cores = 2,
    DATA_SPIAT = DATA_SPIAT,
    DATA_Phenotypes = CSM_Phenotypecell_test,
    Cell_type_to_define_cluster = "TUMOR",
    Minimum_number_cells_cluster = 5,
    Cell_types_of_interest = c("TUMOR", "CD8_GZMBneg", "CD8_GZMBpos", "OTHER"),
    Layers_margin = 5,
    Simplify_result = T
)

}
\seealso{
\code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.
}
