% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_level.R
\name{str_to_level}
\alias{str_to_level}
\title{Extract the string that corresponds to a given level of a
categorical variabl}
\usage{
str_to_level(x)
}
\arguments{
\item{x}{(chr) a string (i.e. character vector of length one)
representing factors' levels in the form of
\verb{level1, label1 | level2, label2 | ... | levelN, labelN}}
}
\value{
(chr) a character string of length \code{N} reporting the labels
of categorical variables.
}
\description{
Extract the string that corresponds to a given level of a
categorical variabl
}
\examples{
coded <- "1, maschio | 2, femmina | 3, ambiguo"
levels <- str_to_level(coded)
}
