% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_morphometry_graph_maker.R
\name{Structure_morphometry_graph_maker}
\alias{Structure_morphometry_graph_maker}
\title{Generates a plot comparing the tile pattern and the structures identified}
\usage{
Structure_morphometry_graph_maker(
  Tiled_Images,
  Variable_name,
  DATA_Morphometry,
  Image_name,
  Color_by = "s.area"
)
}
\arguments{
\item{Tiled_Images}{A list containing tiled images obtained using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}} or the \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}.}

\item{Variable_name}{A character value indicating the column name containing the cell labels.}

\item{DATA_Morphometry}{A list containing structure morphometry data (generated using the \code{\link[=Structure_morphometry_calculator]{Structure_morphometry_calculator()}}).}

\item{Image_name}{A single character value specifying the name of the image to be plotted.}

\item{Color_by}{A single character value indicating the variable defining the color of the structures.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
Plots the resulting graphs and returns a list with the graphs.
}
\description{
Given a tiled image cell data (generated using the \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}} or the \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}) and
Morphometry_data (generated using the \code{\link[=Structure_morphometry_calculator]{Structure_morphometry_calculator()}}) the function will graph a specific image comparing the tile pattern
and the structures identified. Structures can be identified can be colored by different structure metrics.
}
\examples{
\dontrun{
#Divide cells into discrete tiles
Tiled_Images <-
Image_tiling_processing_function(
 N_cores = 1,
 DATA = CSM_Phenotypecell_test,
 Tile_width = 25,
 Tile_height = 25,
 Variables_to_keep = "Phenotype")

#Try to identify morphological structures for OTHER cells according to tile distribution
Structures <- 
Structure_morphometry_calculator(N_cores = 1,
                                
                                Tiled_Images = Tiled_Images,
                                Variable_name = "Phenotype",
                                Feature_to_analyze = "OTHER",
                                
                                Min_tile_number = 1, 
                                Smooth_amount = 0.01, 
                                Tolerance_value = 1, 
                                Neighborhood_distance = 1,
                                Minimum_cell_no_per_tile = 2)

#Analyze the results
Structure_morphometry_analyzer(DATA = Structures)

#Graph the results
Structure_morphometry_graph_maker(Tiled_Images = Tiled_Images,
                                  Variable_name = "Phenotype",
                                  DATA_Morphometry = Structures,
                                  Color_by = "Structure_ID",
                                  Image_name = "ABCM22001_B09_MiniCrop.tif")
}


}
\seealso{
\code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}, \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}, \code{\link[=Structure_morphometry_calculator]{Structure_morphometry_calculator()}}, \code{\link[=Structure_morphometry_analyzer]{Structure_morphometry_analyzer()}}
}
