% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_redcap_url.R
\name{study_redcap_url}
\alias{study_redcap_url}
\title{Return the REDCap's url of the study}
\usage{
study_redcap_url()
}
\value{
(chr) the REDCap url to access the study's data.
}
\description{
Return the REDCap's url of the study
}
\examples{
study_redcap_url()
}
