% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suggested_Tile_Size_Calculator.R
\name{Suggested_Tile_Size_Calculator}
\alias{Suggested_Tile_Size_Calculator}
\title{Calculates appropriate tile size to satisfy user preferences}
\usage{
Suggested_Tile_Size_Calculator(
  DATA = NULL,
  N_rows = NULL,
  N_cols = NULL,
  Based_on_smaller = NULL,
  Draw_preview = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{N_rows}{An integer value indicating the number of rows to divide into tiles.}

\item{N_cols}{An integer value indicating the number of columns to divide into tiles.}

\item{Based_on_smaller}{A logical value indicating if the test should be run using the smallest image. If FALSE the largest image in the dataset will be used.}

\item{Draw_preview}{A logical value indicating if the test should be plotted.}
}
\value{
A tibble containing recommended tile dimensions.
}
\description{
The function can be used  to plan the image tiling strategy. The function calculates the tile dimensions required to divide an image into a user provided number of rows and columns.
The calculation can be performed using the smallest or largest image from the dataset.
}
\examples{
Suggested_Tile_Size_Calculator(
   CSM_Phenotypecell_test,
   N_cols = 4,
   N_rows = 4,
   Based_on_smaller = TRUE,
   Draw_preview = TRUE
)

}
\seealso{
\code{\link[=Image_length_calculator]{Image_length_calculator()}}, \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}
}
