% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tailored_Closest_neighbor_calculator.R
\name{Tailored_Closest_neighbor_calculator}
\alias{Tailored_Closest_neighbor_calculator}
\title{Compute the closest neighbors matrix.}
\usage{
Tailored_Closest_neighbor_calculator(
  N_cores = NULL,
  DATA = NULL,
  Strategy = NULL,
  N_neighbors = NULL,
  Include_COO_in_neighborhood = NULL,
  Max_dist_allowed = NULL,
  Cell_Of_Origin = NULL,
  Target_Cell = NULL
)
}
\arguments{
\item{N_cores}{Number of cores to parallelize your computation}

\item{DATA}{A dataframe or tibble containing cell feature data that and a column named 'Phenotype' containing cell labels.}

\item{Strategy}{A character value indicating the definition of neighborhoods. Either: "Number", "Distance", "Both".}

\item{N_neighbors}{If Strategy is Number or Both, an integer value indicating the N closest neighbors to be calculated.}

\item{Include_COO_in_neighborhood}{A logical value indicating if the Cell of Origin itself should be included in the neighborhood.}

\item{Max_dist_allowed}{If Strategy is Distance or Both, a numeric value indicating the maximum distance allowed to be considere a neighbor.}

\item{Cell_Of_Origin}{A character vector indicating which cell phenotype labels will be visited to compute their neighbors.}

\item{Target_Cell}{A character vector indicating which cell phenotype labels will be considered when calculating the neighbors.}
}
\value{
A list containing two tibbles. Percentage: A tibble containing the neighbor matrix expressed as percentages. Absolute: A tibble containing the neighbor matrix expressed as cell counts.
}
\description{
The function calculates the closest neighbors matrix according to user preferences. The algorithm will visit all the cells in the dataset and compute their neighbors. During this process, the visited cell is considered the Cell of Origin (COO) and the neighbors are the target cells.
The closest neighbor matrix can be used to identify neighborhoods using \code{\link[=Neighborhood_discovery_function]{Neighborhood_discovery_function()}}.
In addition to the matrix, the function generates a correlation plot. The correlation between neighbors composition can be interpreted as attraction repulsion patterns between cell types.
Neighbors may be defined by distance (a cell is considered a neighbor if it's within a defined distance D), by number (neighbors are the closest N cells) or both (cells are N closest neighbors as long as they are within distance D).
}
\seealso{
\code{\link[=Neighborhood_discovery_function]{Neighborhood_discovery_function()}}, \code{\link[=DATA_neighborhoods_renamer]{DATA_neighborhoods_renamer()}}, \code{\link[=Neighborhood_Quantifier]{Neighborhood_Quantifier()}}, \code{\link[=Neighborhood_voting_function]{Neighborhood_voting_function()}}, \code{\link[=Tiled_neighborhoods_graphicator]{Tiled_neighborhoods_graphicator()}}
}
