% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_exploration_function.R
\name{Thresholding_exploration_function}
\alias{Thresholding_exploration_function}
\title{Summarizes thresholded data by image}
\usage{
Thresholding_exploration_function(
  DATA = NULL,
  Calculate_Density = NULL,
  DATA_Area = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data that has been thresholded.}

\item{Calculate_Density}{A logical value indicating if density should be calculated.}

\item{DATA_Area}{If Calculate_Density is TRUE, a dataframe or tibble containing image names and tissue area information.}
}
\value{
Returns a tibble with cell counts, pertentages and optionally cell densities per image.
}
\description{
The function generates positive cell counts, percentages and optionally cell densities according to thresholded feature data.
}
