% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_function.R
\name{Thresholding_function}
\alias{Thresholding_function}
\title{Thresholds cell features}
\usage{
Thresholding_function(
  DATA,
  Strategy,
  Local_thresholding = FALSE,
  Method_autothreshold = "Otsu",
  number_iterations_TriClass = 20,
  Percentile = 0.5,
  Defined_threshold = 0.1,
  Levels = 3
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Strategy}{A character indicating the thresholding approach. It mus be one of the following: "EBI_Otsu", "Kmeans", "Kmeans_Otsu", "Autothreshold", "TriClass_Otsu", "Mean", "Quantile", "Arbitrary", "Multi_level" (see details).}

\item{Local_thresholding}{A logical value indicating if local (per image) threshold should be calculated (see details).}

\item{Method_autothreshold}{If strategy is "Autothreshold", indicate the type of threshold to be used. It must be one of "IJDefault", "Huang", "Huang2", "Intermodes", "IsoData", "Li", "MaxEntropy", "Mean",
"MinErrorI", "Minimum", "Moments", "Otsu", "RenyiEntropy", "Shanbhag", "Triangle", "Yen".}

\item{number_iterations_TriClass}{If strategy is "TriClass_Otsu", an integer value indicating the maximum number of iterations.}

\item{Percentile}{If strategy is "Quantile", a numeric value indicating the quantile to be used as cut-off value.}

\item{Defined_threshold}{If strategy is "Arbitrary", a numeric value indicating the cut-off value.}

\item{Levels}{If strategy is "Multi_level", an integer indicating the number of levels.}
}
\value{
Returns a tibble with thresholded cell features. All features are thresholded using the same approach. For binary thresholding features are converted to logical vectors, for multi-level are converted to numeric vectors.
}
\description{
\code{Thresholding_function()} thresholds cell features following several approaches. The user can test this approaches using the \code{\link[=Thresholding_tester_app]{Thresholding_tester_app()}}.
Thresholded features can be further processed using the \code{\link[=Marker_combinator_generator]{Marker_combinator_generator()}} and \code{\link[=Phenotype_assigner_function]{Phenotype_assigner_function()}} to obtain cell phenotypes.
}
\details{
Local thresholding is generally discouraged. It may capture image-specific artifacts or patterns and not true feature positivity.

EBI_Otsu calculates Otsu threshold using EBImage::otsu function. Kmeans calculates threshold using the mmand::threshold function. Kmeans_Otsu calculates the threhold using imager::threshold function.
Autothreshold calculates thresholds using the autothresholdr::auto_thresh_mask. TriClass_Otsu calculates thresholds using imagerExtra::ThresholdTriclass function.
Multi_level thresholds are calculated using imagerExtra::ThresholdML function.
}
\examples{
\dontrun{
Thresholding_function(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Strategy = "EBI_Otsu",
  Local_thresholding = TRUE,
  Method_autothreshold = "Otsu",
  number_iterations_TriClass = 20,
  Percentile = 0.5,
  Defined_threshold = 0.1,
  Levels = 3
)
}

}
\seealso{
\code{\link[=Thresholding_tester_app]{Thresholding_tester_app()}}, \code{\link[=Thresholding_function_tailored]{Thresholding_function_tailored()}}, \code{\link[=Thresholding_exploration_function]{Thresholding_exploration_function()}}, \code{\link[=Thresholding_summary_function]{Thresholding_summary_function()}}, \code{\link[=Thresholding_graphicator_function]{Thresholding_graphicator_function()}}.
}
