% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_function_tailored.R
\name{Thresholding_function_tailored}
\alias{Thresholding_function_tailored}
\title{Thresholds cell features applying a unique approach for every marker}
\usage{
Thresholding_function_tailored(DATA = NULL, Variables_tibble = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Variables_tibble}{A dataframe or tibble summarizing thresholding approach for every marker. It must contain the following columns: variable, Strategy, Local_thresholding, Method_autothreshold,
number_iterations_TriClass, Percentile, Defined_threshold, Levels. Values for every of this columns follow the same rules as for \code{\link[=Thresholding_function]{Thresholding_function()}}.}
}
\value{
Returns a tibble with thresholded cell features. Every feature is thresholded using specific parameters. For binary thresholding features are converted to logical vectors, for multi-level are converted to numeric vectors.
}
\description{
In contrast to \code{Thresholding_function()} this function allows selecting a feature specific approach for thresholding. The user can test this approaches using the \code{\link[=Thresholding_tester_app]{Thresholding_tester_app()}}.
Thresholded features can be further processed using the \code{\link[=Marker_combinator_generator]{Marker_combinator_generator()}} and \code{\link[=Phenotype_assigner_function]{Phenotype_assigner_function()}} to obtain cell phenotypes.
}
\details{
Local thresholding is generally discouraged. It may capture image-specific artifacts or patterns and not true feature positivity.

EBI_Otsu calculates Otsu threshold using EBImage::otsu function. Kmeans calculates threshold using the mmand::threshold function. Kmeans_Otsu calculates the threhold using imager::threshold function.
Autothreshold calculates thresholds using the autothresholdr::auto_thresh_mask. TriClass_Otsu calculates thresholds using imagerExtra::ThresholdTriclass function.
Multi_level thresholds are calculated using imagerExtra::ThresholdML function.
}
\seealso{
\code{\link[=Thresholding_tester_app]{Thresholding_tester_app()}}, \code{\link[=Thresholding_function]{Thresholding_function()}}, \code{\link[=Thresholding_exploration_function]{Thresholding_exploration_function()}}, \code{\link[=Thresholding_summary_function]{Thresholding_summary_function()}}, \code{\link[=Thresholding_graphicator_function]{Thresholding_graphicator_function()}}.
}
