% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_graphicator_function.R
\name{Thresholding_graphicator_function}
\alias{Thresholding_graphicator_function}
\title{Generates thresholding results summary plots}
\usage{
Thresholding_graphicator_function(
  DATA = NULL,
  DATA_thresholded = NULL,
  Marker_names = NULL,
  Image_name = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{DATA_thresholded}{A dataframe or tibble containing cell feature data that has been thresholded.}

\item{Marker_names}{A character vector indicating features that need to be plotted.}

\item{Image_name}{A character value indicating the image to be plotted.}
}
\value{
Returns a tibble with cell counts, percentages and optionally cell densities per image.
}
\description{
The function generates plots comparing feature expression values and thresholding results for individual images.
}
