% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_summary_function.R
\name{Thresholding_summary_function}
\alias{Thresholding_summary_function}
\title{Calculates cut-off points of thresholded data and generates summary plots}
\usage{
Thresholding_summary_function(DATA = NULL, DATA_thresholded = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{DATA_thresholded}{A dataframe or tibble containing cell feature data that has been thresholded.}
}
\value{
Returns a tibble with cut-off values and generates summary plots.
}
\description{
Given the original and thresholded dataset, the function calculates cut-off values and generates summary graphs.
}
\examples{
\dontrun{
#Threshold data
DATA_thresholded <- Thresholding_function(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Strategy = "EBI_Otsu",
  Local_thresholding = FALSE,
  Method_autothreshold = "Otsu",
  number_iterations_TriClass = 20,
  Percentile = 0.5,
  Defined_threshold = 0.1,
  Levels = 3
)

#Calculate the summary
Thresholding_summary_function(
   DATA = CSM_Arrangedcellfeaturedata_test,
   DATA_thresholded = DATA_thresholded
)
}


}
