% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_extract.R
\name{tidy_extract}
\alias{tidy_extract}
\title{Extract and tidy data from REDCap import.}
\usage{
tidy_extract(data, type = c("data", "meta"))
}
\arguments{
\item{data}{the output of \link{read_redcap}}

\item{type}{(chr) it can takes two values: "data" or "meta". If
"data" is passed, then the data of the study are
returned. If "meta" is passed, then the meta-data
of the study are returned.}
}
\value{
\link[tibble:tibble-package]{tibble}
}
\description{
The function extract the data or the metadata from the output
returned by \link{read_redcap}. After the extraction, variables names
are cleaned, textual columns are lowered, and the event fields
like "fields" are renamed.
}
\examples{
\dontrun{

url     <- "https://bbmc.ouhsc.edu/redcap/api/"
token   <- "9A81268476645C4E5F03428B8AC3AA7B"
data <- read_redcap(url = url, token = token)

tidy_extract(data, type = "data")
tidy_extract(data, type = "meta")

}

}
