% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_fields_names.R
\name{tidy_fields_names}
\alias{tidy_fields_names}
\title{Tidy the names of the fields}
\usage{
tidy_fields_names(data)
}
\arguments{
\item{data}{(tibble) a tibble with one column that contains the
names of the fields with missing information}
}
\value{
\link[tibble:tibble-package]{tibble} a tibble with one column
that contains the names with
missing information without
duplicates
}
\description{
The function takes a character vector which elements are the names
of the fields for which the information is missing. If a field has
more than one column, it remove the duplicated columns.
}
\examples{

a <- tibble::tibble(
  variable = c("aa", "dd_1", "dd_2", "dd_3", "cc")
)

tidy_fields_names(data = a)

}
