% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile_generator_function.R
\name{Tile_generator_function}
\alias{Tile_generator_function}
\title{Generates tiles from any given image}
\usage{
Tile_generator_function(
  Image_name,
  DATA = NULL,
  Tile_width = NULL,
  Tile_height = NULL,
  Variables_to_keep = NULL
)
}
\arguments{
\item{Image_name}{A character value indicating the name of the image being processed.}

\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Tile_width}{A numeric value indicating the width of the tiles.}

\item{Tile_height}{A numeric value indicating the height of the tiles.}

\item{Variables_to_keep}{A character vector indicating the column names to be kept when tiling the image.}
}
\value{
A list containing two elements. Tile_info contains information of tiles. Final_tibble contains cell information and the tile where they are located.
}
\description{
Intended for internal use only
}
\seealso{
\code{\link[=Suggested_Tile_Size_Calculator]{Suggested_Tile_Size_Calculator()}}, \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}
}
\keyword{Internal}
