% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile_to_image_cell_arrange_function.R
\name{Tile_to_image_cell_arrange_function}
\alias{Tile_to_image_cell_arrange_function}
\title{Arrange cell coordinates according to position in whole image using tile information}
\usage{
Tile_to_image_cell_arrange_function(
  DATA,
  Dist_to_edge = NULL,
  Tile_overlap = 0
)
}
\arguments{
\item{DATA}{Dataframe or tibble with cell features. This data should have been processed at some point using \code{\link[=Data_arrange_function]{Data_arrange_function()}}.}

\item{Dist_to_edge}{(OPTIONAL) For overlapping tiles, cells closer to tile edge than Dist_to_edge will be removed.}

\item{Tile_overlap}{(OPTIONAL) A numeric value indicating the tile overlap. If provided, automatic tile overlap detection will be aborted. If a value of 0 is provided, no overlap correction will be performed at all.}
}
\value{
A tibble containing cell features with modified X and Y coordinates and Subject_Names information.
}
\description{
\code{Tile_to_image_cell_arrange_function()} modifies cell X and Y coordinates to reflect cell position in the entire image. For cell segmentation pipelines that
are based on tiled images (those processed using the \code{\link[=Image_tile_deconstruction_function]{Image_tile_deconstruction_function()}} function), the function will switch the tile-based X Y coordinates to image-based coordinates.
The function performs the operation based on the tile X Y coordinates info contained in the Subject_Names variable. If names have been modified the function could return unexpected results. In addition,
the Subject_Names variable of the dataset is also modified to reflect entire image information. If tiles present overlap, duplicated cells are removed. Further, cells close to the tile edge can
also be discarded according to user preferences using the 'Dist_to_edge' argument.
}
\examples{
\dontrun{
#Returns an error since Subject_Names in dataframe does not contain adequate tile information
Tile_to_image_cell_arrange_function(
 DATA = CSM_Arrangedcellfeaturedata_test,
 Dist_to_edge = 0
)

}



}
\seealso{
\code{\link[=Image_tile_deconstruction_function]{Image_tile_deconstruction_function()}}, \code{\link[=Mask_conflict_resolver]{Mask_conflict_resolver()}}, \code{\link[=Image_from_tile_rebuilder]{Image_from_tile_rebuilder()}}
}
