% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiled_image_heterogeneity_calculator.R
\name{Tiled_image_heterogeneity_calculator}
\alias{Tiled_image_heterogeneity_calculator}
\title{Calculates heterogeneity by tile}
\usage{
Tiled_image_heterogeneity_calculator(
  Tiled_images = NULL,
  Minimum_cell_no_per_tile = NULL,
  Phenotypes_included = NULL
)
}
\arguments{
\item{Tiled_images}{A list containing tiled images obtained using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}.}

\item{Minimum_cell_no_per_tile}{An integer indicating the minimum number of cells that a tile must contain. Tiles below the limit will not be included in the analysis.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}
}
\value{
A list containing image information with by-tile heterogeneity metrics.
}
\description{
The function calculates various heterogeneity metrics based on cell composition distribution of every tile in an image.
}
\details{
Shannon, Simpson,Inverse Simpson and Renyi entropy indexes are calculated using the vegan package.

Rao index is calculated using the picante::raoD function.

Gini index is calculated using the DescTools::Gini function.

Kullback_Leibler and Jensen_Shannon indexes are calculated using the philentropy package. They compare the observed cell composition within each tile against the global cell composition of the sample.
}
