% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiled_image_heterogeneity_graph_maker.R
\name{Tiled_image_heterogeneity_graph_maker}
\alias{Tiled_image_heterogeneity_graph_maker}
\title{Generates a plot illustrating heterogeneity by tile.}
\usage{
Tiled_image_heterogeneity_graph_maker(
  DATA = NULL,
  Tiled_images = NULL,
  Image_name = NULL,
  Metric = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Tiled_images}{A list containing tiled images obtained using \code{\link[=Tiled_image_heterogeneity_calculator]{Tiled_image_heterogeneity_calculator()}}.}

\item{Image_name}{A character value indicating the image to be plotted.}

\item{Metric}{A character value indicating the heterogeneity metric to be plotted.}
}
\value{
Returns a plot with the cells and the overlying tiles shaded according to the heterogeneity index.
}
\description{
Generates a plot illustrating heterogeneity by tile.
}
