% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiled_neighborhoods_graphicator.R
\name{Tiled_neighborhoods_graphicator}
\alias{Tiled_neighborhoods_graphicator}
\title{Generates a plot illustrating heterogeneity by tile.}
\usage{
Tiled_neighborhoods_graphicator(
  DATA_elections = NULL,
  Image_name = NULL,
  Graph_only_winner_neighborhood = NULL
)
}
\arguments{
\item{DATA_elections}{An object created using the \code{\link[=Neighborhood_voting_function]{Neighborhood_voting_function()}} function containing neighborhood election results.}

\item{Image_name}{A character value indicating the image to be plotted.}

\item{Graph_only_winner_neighborhood}{A logical value indicating if only the winner neighborhood by tile should be plotted.}
}
\value{
Returns a plot with the tiles and the neighborhood elections results.
}
\description{
Generates a plot illustrating heterogeneity by tile.
}
