% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangle_generator.R
\name{Triangle_generator}
\alias{Triangle_generator}
\title{Generates triangles given two point coordinates (a COO and a target point)}
\usage{
Triangle_generator(DATA, Angle = 10)
}
\arguments{
\item{DATA}{A dataframe or tibble containing information of coordinates of two data points.}

\item{Angle}{A numeric value indicating the angle of the triangle edges in the COO vertex}
}
\value{
A tibble containing the coordinates of triangle vertex points. Also returns distance and angular position of COO and target cells.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}}
}
\keyword{Internal}
