% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trio_Cells_in_Radius_analyzer.R
\name{Trio_Cells_in_Radius_analyzer}
\alias{Trio_Cells_in_Radius_analyzer}
\title{Calculates the average TRIO score for every image}
\usage{
Trio_Cells_in_Radius_analyzer(
  DATA = NULL,
  DATA_RANDOM = NULL,
  Radius = NULL,
  Include_Random = NULL,
  By_Sample_Random = NULL
)
}
\arguments{
\item{DATA}{A list containing cumulative interaction data and TRIO scores. It can be computed using \code{\link[=Trio_Cumulative_Interaction_generator]{Trio_Cumulative_Interaction_generator()}} function.}

\item{DATA_RANDOM}{(OPTIONAL) A list containing random cumulative interaction data and TRIO scores. It can be computed using \code{\link[=Trio_Cumulative_Interaction_generator]{Trio_Cumulative_Interaction_generator()}} function.}

\item{Radius}{A numeric value indicating the size of the radius to perform analysis. It must have been computed during the generation of the cumulative interaction matrix.}

\item{Include_Random}{A logical value indicating if the Random background should be used.}

\item{By_Sample_Random}{If a random background needs to be included, a logical value indicating if the random background should be calculated by sample or by experiment (see details).}
}
\value{
A tibble containing a summary by sample of spatial interactions.
}
\description{
The function calculates the average TRIO score for every image given a radius size. For more information on the TRIO score please see \code{\link[=Trio_Cumulative_Interaction_generator]{Trio_Cumulative_Interaction_generator()}}.
}
\details{
If By_Sample_Random is TRUE, for every sample, a specific random background will be calculated to compute expected spatial interaction metrics. If By_Sample_Random is FALSE, the random background will be calculated using all image information. The same random background will be used for all images.
}
\seealso{
\code{\link[=Trio_Random_Distance_matrix_generator]{Trio_Random_Distance_matrix_generator()}}, \code{\link[=Trio_Cumulative_Interaction_generator]{Trio_Cumulative_Interaction_generator()}}, \code{\link[=Trio_Min_Distance_analyzer]{Trio_Min_Distance_analyzer()}}, \code{\link[=Trio_graph_maker]{Trio_graph_maker()}}.
}
