% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trio_graph_maker.R
\name{Trio_graph_maker}
\alias{Trio_graph_maker}
\title{Generates a plot of the TRIO spatial interaction for a single image}
\usage{
Trio_graph_maker(
  Image_name,
  DATA_Phenotypes,
  Strategy,
  TRIO_Distances = NULL,
  TRIO_Cumulative = NULL,
  Radius = NULL
)
}
\arguments{
\item{Image_name}{A character value indicating the image to be plotted.}

\item{DATA_Phenotypes}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Strategy}{A character value indicating the plot that will be generated. One of the following: "Min_Distance" or "TRIO_in_Radius".}

\item{TRIO_Distances}{If Strategy is Min_Distance, a trio distance matrix list generated using \code{\link[=Trio_Distance_matrix_generator]{Trio_Distance_matrix_generator()}} function.}

\item{TRIO_Cumulative}{If Strategy is TRIO_in_Radius, a trio cumulative interaction matrix list generated using \code{\link[=Trio_Cumulative_Interaction_generator]{Trio_Cumulative_Interaction_generator()}} function.}

\item{Radius}{If strategy is TRIO_in_Radius, a numeric value indicating the radius to be plotted. It must have been computed during the generation of the cumulative interaction matrix.}
}
\value{
A graph summarizing spatial interactions for the indicated image.
}
\description{
The function generates a summary plot of TRIO spatial interactions. It can graph results for min dist to TRIO and TRIO score within radius approaches
}
\examples{
\dontrun{
#Generate distance matrix-----------------------------------
TRIO_Distance <-
Trio_Distance_matrix_generator(
    N_cores = 1,
    DATA = CSM_Phenotypecell_test,
    Cell_Of_Origin = "TUMOR",
    Target_Cell_1 = "CD8_GZMBneg",
    Target_Cell_2 = "CD8_GZMBpos",
    Perform_edge_correction = FALSE
)

#Generate cumulative interactions (must contain the actual radius distance)--
TRIO_Cumulative <-
Trio_Cumulative_Interaction_generator(
  N_cores = 1,
  DATA = TRIO_Distance,
  Start_from = 25,
  Stop_at = 100,
  Sampling_frequency = 25
)

#Analyze min distance
Trio_graph_maker(
   Image_name = "ABCM22001_B14_MiniCrop.tif",
   DATA_Phenotypes = CSM_Phenotypecell_test,
   Strategy = "Min_Distance",
   TRIO_Distances = TRIO_Distance
)

Trio_graph_maker(
   Image_name = "ABCM22001_B14_MiniCrop.tif",
   DATA_Phenotypes = CSM_Phenotypecell_test,
   Strategy = "TRIO_in_Radius",
   TRIO_Cumulative = TRIO_Cumulative,
   Radius = 50
)
}

}
\seealso{
\code{\link[=Trio_Distance_matrix_generator]{Trio_Distance_matrix_generator()}}, \code{\link[=Trio_Cumulative_Interaction_generator]{Trio_Cumulative_Interaction_generator()}}, \code{\link[=Trio_Min_Distance_analyzer]{Trio_Min_Distance_analyzer()}}, \code{\link[=Trio_Cells_in_Radius_analyzer]{Trio_Cells_in_Radius_analyzer()}}
}
