% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trio_Min_Distance_analyzer.R
\name{Trio_Min_Distance_analyzer}
\alias{Trio_Min_Distance_analyzer}
\title{Calculates the min distance to generate a TRIO between three cell types}
\usage{
Trio_Min_Distance_analyzer(
  N_cores = NULL,
  DATA = NULL,
  DATA_RANDOM = NULL,
  Include_Random = NULL,
  By_Sample_Random = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A list containing trio cell distance matrix. It can be computed using \code{\link[=Trio_Distance_matrix_generator]{Trio_Distance_matrix_generator()}} function.}

\item{DATA_RANDOM}{(OPTIONAL) A list containing random  trio cell distance matrix. It can be computed using \code{\link[=Trio_Random_Distance_matrix_generator]{Trio_Random_Distance_matrix_generator()}} function.}

\item{Include_Random}{A logical value indicating if the Random background should be used.}

\item{By_Sample_Random}{If a random background needs to be included, a logical value indicating if the random background should be calculated by sample or by experiment (see details).}
}
\value{
A tibble containing a summary by sample of spatial interactions.
}
\description{
The function calculates the average min distance to generate a TRIO between three cell types for every image.
}
\details{
If By_Sample_Random is TRUE, for every sample, a specific random background will be calculated to compute expected spatial interaction metrics. If By_Sample_Random is FALSE, the random background will be calculated using all image information. The same random background will be used for all images.
}
