% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trio_Random_Distance_matrix_generator.R
\name{Trio_Random_Distance_matrix_generator}
\alias{Trio_Random_Distance_matrix_generator}
\title{Calculates random distance backgrounds between a cell of origin and two target cell types}
\usage{
Trio_Random_Distance_matrix_generator(
  N_cores = NULL,
  DATA = NULL,
  Cell_Of_Origin = NULL,
  Target_Cell_1 = NULL,
  Target_Cell_2 = NULL,
  Random_cells_per_sample = NULL,
  Perform_edge_correction = NULL,
  Hull_ratio = NULL,
  Distance_to_edge = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Cell_Of_Origin}{A character value indicating the cell phenotype label of the Cell of Origin.}

\item{Target_Cell_1}{A character value indicating the cell phenotype label of the first Target cell.}

\item{Target_Cell_2}{A character value indicating the cell phenotype label of the second Target cell.}

\item{Random_cells_per_sample}{A integer value indicating the number of random iterations to calculate the distances.}

\item{Perform_edge_correction}{A logical value indicating if COO close to the edge of the tissue should be removed from the analysis.}

\item{Hull_ratio}{If edge correction needs to be performed, a numeric value indicating the hull ratio. Smaller values calculate more precise edge silhouettes at the cost of being more computationally demanding.}

\item{Distance_to_edge}{If edge correction needs to be performed, the distance the maximum distance to the edge allowed. Cells closer to the edge will be removed.}
}
\value{
A list containing the random distance matrix for each image. Rows represent Cells of Origin and columns represent Target cells.
}
\description{
Given a dataframe with cell features and phenotype labels, random spatial distances between three cell populations is calculated. This is done by
randomly permuting cell labels in an iterative manner. For every image the same number of random COO to target cell distances are calculated.
}
\examples{
\dontrun{
Trio_Random_Distance_matrix_generator(
  N_cores = 1,
  DATA = CSM_Phenotypecell_test,
  Cell_Of_Origin = "TUMOR",
  Target_Cell_1 = "CD8_GZMBneg",
  Target_Cell_2 = "CD8_GZMBpos",
  Random_cells_per_sample = 10,
  Perform_edge_correction = FALSE
)
}


}
\seealso{
\code{\link[=Trio_Distance_matrix_generator]{Trio_Distance_matrix_generator()}}, \code{\link[=Trio_Cumulative_Interaction_generator]{Trio_Cumulative_Interaction_generator()}}, \code{\link[=Trio_Min_Distance_analyzer]{Trio_Min_Distance_analyzer()}}, \code{\link[=Trio_Cells_in_Radius_analyzer]{Trio_Cells_in_Radius_analyzer()}}, \code{\link[=Trio_graph_maker]{Trio_graph_maker()}}
}
