% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTAG_message_passing_Image_based_phenotyper.R
\name{UTAG_message_passing_Image_based_phenotyper}
\alias{UTAG_message_passing_Image_based_phenotyper}
\title{Performs message passing tailored for the image based phenotyper shiny App}
\usage{
UTAG_message_passing_Image_based_phenotyper(
  DATA = NULL,
  COO_to_visit = NULL,
  Neighbor_strategy = NULL,
  Message_strategy = NULL,
  N_neighbors = NULL,
  Max_dist_allowed = NULL,
  Weighting_Strategy = NULL,
  N_cores = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{COO_to_visit}{(OPTIONAL) A logical vector indicating which cells from DATA should be visited by the algorithm.}

\item{Neighbor_strategy}{A character value indicating how neighboring cells are defined: "Number", "Distance" or "Both".}

\item{Message_strategy}{A character value indicating how features are shared between neighboring cells: "Averaging" or "Sum".}

\item{N_neighbors}{If strategy is Number or Both, an integer indicating the number of closest neighbors to calculate for each cell.}

\item{Max_dist_allowed}{If strategy is Distance or Both, a numeric value indicating the distance from every cell to search for neighbors}

\item{Weighting_Strategy}{A character value indicating the weighting strategy for message passing: "None", "Proximity", "Disregarded_minority", "Both" (see details).}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
A tibble containing cell feature data that have undergone neighbor message passing.
}
\description{
Intended for internal use only
}
\details{
Used in \code{\link[=Image_based_phenotyper_App_launcher]{Image_based_phenotyper_App_launcher()}}, \code{\link[=Model_cell_phenotyper]{Model_cell_phenotyper()}}

Weighting strategy modifies the way neighbors share information:

Proximity weighting allows closer neighbors to have a higher influence compared to distant neighbors. For a set of neighbors each located at a distance d from the cell of origin, and being the sum of all neighbor distances D, weight is defined as 1/(d/D).

Disregarded_minority weighting allows low expressed features to have higher influence on message passing: First the -log10 of average feature expression is calculated for every feature. The weight is defined as -log10 average / min(-log10 average)

Proximity weighting is available for averaging and sum message passing strategies. Disregarded_minority can only be used with averaging.
}
\keyword{Internal}
