% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_check.R
\name{var_check}
\alias{var_check}
\title{Perform Variance Check}
\usage{
var_check(data, title = "")
}
\arguments{
\item{data}{A data frame with the site name/number (can be string or numeric) in the first column and any continuous variables to check in the following columns.}

\item{title}{This provides the prefix of the title for table. Default is empty.}
}
\value{
sites with extremely low variance
}
\description{
The function is to compare the variance in one site to variance in the other sites.
}
\examples{
\dontrun{
var_check(data, '')
}
}
\section{Last Edited By}{
Feinan Lu
}

\section{Last Edit Date}{
2022-11-16
}

