% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_out_range.R
\name{which_out_range}
\alias{which_out_range}
\title{Check which continuous variable has an out-of-range value}
\usage{
which_out_range(data, range)
}
\arguments{
\item{data}{(data.frame) a dataframe with the continuous variables
for each patient.}

\item{range}{(character) output of the function \code{extract_range}.}
}
\value{
(character) a character vector with the names of the
continuous variables that had an out-of-range
value.
}
\description{
For each patient enrolled in the study, the function checks which
continuous variable assumes a value that it is not included in the
range.
}
\examples{

range <- list("age" = c(18, 70), "hemoglobin" = c(6, 18))
df <- tibble::tibble(age = 16, hemoglobin = 9)

which_out_range(data = df, range = range)

}
