% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barrier_effect_analyzer.R
\name{Barrier_effect_analyzer}
\alias{Barrier_effect_analyzer}
\title{Computes sample summary metrics of barrier elements}
\usage{
Barrier_effect_analyzer(DATA = NULL, Analysis_type = "Area")
}
\arguments{
\item{DATA}{A tibble containing barrier cell data obtained using the \code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}} function.}

\item{Analysis_type}{A character value indicating the type of metric to be summarized. One of the following: 'Area' (default), 'Distance', 'N_targets'.}
}
\value{
A list containing tibbles with by sample summary metrics.
}
\description{
Givena dataset obtained using the \code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}}, the function computes sample summary metrics for every element analyzed.
}
\examples{
\dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
          function(Image){
            EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
                               file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
          })

#Generate the Image_parameter_list------------------------------
Image_Parameters <- list(
 PDL1 = list(
   Directory = Input_Dir,
   Channel = 2,
   Image_rotate = NULL,
   Image_x_flip = FALSE,
   Image_y_flip = TRUE),
 GZMB = list(
   Directory = Input_Dir,
   Channel = 3,
   Image_rotate = NULL,
   Image_x_flip = FALSE,
   Image_y_flip = TRUE))

#Compute the cells and pixels in the barrier------------------------------
Barrier_DATA <- Barrier_effect_calculator(
 N_cores = 1,

 DATA = CSM_Phenotypecell_test,


 Cell_Of_Origin = "CD8_GZMBneg",
 Target_Cell = "TUMOR",
 Barrier_Cell = c("OTHER", "CD8_GZMBpos"),
 Image_parameter_list = Image_Parameters,

 Distance_sampled = 30,

 Perform_edge_correction = FALSE,

 Polygon_angle = 30,
 Target_cell_Angle_tolerance = 30
)

#Summarize the results by sample------------------------------
Barrier_effect_analyzer(
 DATA = Barrier_DATA,
 Analysis_type = "Area"
 )

#Plot the results------------------------------
Barrier_graph_maker(
  DATA_Phenotypes = CSM_Phenotypecell_test,
  Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype),

  Barrier_DATA = Barrier_DATA,
  Image_name = "ABCM22001_B14_MiniCrop.tif",

  Color_by = "GZMB_PixelValueSum_Area_Ratio",
  Point_size = 2,

  Image_directory = dir(Input_Dir, full.names = TRUE)[[2]],
  Pixel_distance_ratio = NULL,
  Channel_to_display = 3,
  Image_rotate = 0,
  Image_x_flip = FALSE,
  Image_y_flip = TRUE
)
}

}
\seealso{
\code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}}, \code{\link[=Barrier_graph_maker]{Barrier_graph_maker()}}
}
