% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barrier_graph_maker.R
\name{Barrier_graph_maker}
\alias{Barrier_graph_maker}
\title{Generate a plot of the barrier metrics of single samples}
\usage{
Barrier_graph_maker(
  DATA_Phenotypes,
  Phenotypes_included,
  Barrier_DATA,
  Image_name,
  Color_by,
  Point_size = 1,
  Image_directory = NULL,
  Pixel_distance_ratio = NULL,
  Channel_to_display = 1,
  Image_rotate = 0,
  Image_x_flip = FALSE,
  Image_y_flip = FALSE
)
}
\arguments{
\item{DATA_Phenotypes}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Phenotypes_included}{A character vector indicating the cell phenotypes that will be included in the plot.}

\item{Barrier_DATA}{A tibble containing barrier cell data obtained using the \code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}} function.}

\item{Color_by}{A character value indicating the column name in Barrier_DATA that should be used to color cell points.}

\item{Point_size}{A numeric value indicating the size of the points.}

\item{Image_directory}{(OPTIONAL) Character string specifying the path to the image file that needs to be plotted.}

\item{Pixel_distance_ratio}{(OPTIONAL) A numeric value indicating the ratio between pixel size / cell coordinates. Use this argument when cell coordinates have been transformed to a distance unit like microns.}

\item{Channel_to_display}{A integer value indicating the Channel index to be displayed (1 for single channel images).}

\item{Image_rotate}{(OPTIONAL) A integer value indicating the degrees of rotation of the image.}

\item{Image_x_flip}{A logical value indicating if X image flip should be performed.}

\item{Image_y_flip}{A logical value indicating if Y image flip should be performed.}

\item{Image_nameA}{A character value indicating the image to be plotted (it must be present in DATA_Phenotypes and Barrier_DATA).}
}
\value{
A plot of the barrier effect for a given sample.
}
\description{
Generate a plot of the barrier metrics of single samples
}
\details{
If pixel_distance_ratio is provided, it will modify the X and Y coordinates of the provided DATA_Phenotypes. The function will assume that
X and Y coordinates of the Barrier_DATA has already been modified with the pixel_distance_ratio argument of the \code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}}.
}
\examples{
\dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
          function(Image){
            EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
                               file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
          })

#Generate the Image_parameter_list------------------------------
Image_Parameters <- list(
 PDL1 = list(
   Directory = Input_Dir,
   Channel = 2,
   Image_rotate = NULL,
   Image_x_flip = FALSE,
   Image_y_flip = TRUE),
 GZMB = list(
   Directory = Input_Dir,
   Channel = 3,
   Image_rotate = NULL,
   Image_x_flip = FALSE,
   Image_y_flip = TRUE))

#Compute the cells and pixels in the barrier------------------------------
Barrier_DATA <- Barrier_effect_calculator(
 N_cores = 1,
 
 DATA = CSM_Phenotypecell_test,
 
 
 Cell_Of_Origin = "CD8_GZMBneg",
 Target_Cell = "TUMOR",
 Barrier_Cell = c("OTHER", "CD8_GZMBpos"),
 Image_parameter_list = Image_Parameters,
 
 Distance_sampled = 30,
 
 Perform_edge_correction = FALSE,
 
 Polygon_angle = 30,
 Target_cell_Angle_tolerance = 30
)

#Summarize the results by sample------------------------------
Barrier_effect_analyzer(
 DATA = Barrier_DATA,
 Analysis_type = "Area"
 )
 
#Plot the results------------------------------
Barrier_graph_maker(
  DATA_Phenotypes = CSM_Phenotypecell_test,
  Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype),

  Barrier_DATA = Barrier_DATA,
  Image_name = "ABCM22001_B14_MiniCrop.tif",

  Color_by = "GZMB_PixelValueSum_Area_Ratio",
  Point_size = 2,

  Image_directory = dir(Input_Dir, full.names = TRUE)[[2]],
  Pixel_distance_ratio = NULL,
  Channel_to_display = 3,
  Image_rotate = 0,
  Image_x_flip = FALSE,
  Image_y_flip = TRUE
)
}

}
\seealso{
\code{\link[=Barrier_effect_calculator]{Barrier_effect_calculator()}}, \code{\link[=Barrier_effect_analyzer]{Barrier_effect_analyzer()}}
}
