% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary_threshold_image_combinator.R
\name{Binary_threshold_image_combinator}
\alias{Binary_threshold_image_combinator}
\title{Performs combination of multiple binary thresholded images using boolean operators}
\usage{
Binary_threshold_image_combinator(
  ...,
  Save_processed_images = NULL,
  Output_Directory = NULL,
  Operators = NULL,
  N_cores = NULL
)
}
\arguments{
\item{...}{Strings indicating the path to folders containing binary thresholded images. Images from different directories will be matched by name.}

\item{Save_processed_images}{Logical value indicating if thresholded images should be written (see details).}

\item{Operators}{Character vector indicating how the provided directories should be combined. Accepted values are 'AND', 'OR', 'NOT'}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Output_directory}{Character specifying the path to the folder where  output images are written. It must be an empty folder.}
}
\value{
Returns a tibble with the total positive pixels per image after combining binary thresholded images
}
\description{
This function can be used to generate binary thresholded images that result from combining several thresholded images using the operators AND, OR, NOT
}
\details{
If processed images are saved, these can be further combined with cell position data using \code{\link[=DATA_Phenotypes_distance]{DATA_Phenotypes_distance()}}.
}
\seealso{
\code{\link[=Image_thresholding_app_launcher]{Image_thresholding_app_launcher()}}, \code{\link[=Binary_threshold_image_combinator]{Binary_threshold_image_combinator()}}, \code{\link[=MFI_Experimet_Calculator]{MFI_Experimet_Calculator()}}
}
