% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{branching_logic}
\alias{branching_logic}
\title{Create the branching logic conditions}
\usage{
branching_logic(meta_data, fields_names, branching_logic)
}
\arguments{
\item{meta_data}{(tibble) a tibble containing the meta_data
information of the study.}

\item{fields_names}{(character) a character with the name of the
column of \code{meta_data} that contains
the names of fields.}

\item{branching_logic}{(character)) a character with the name of the
column of \code{meta_data} that the
branching logics associated to
each field.}
}
\value{
(tibble) a tibble with two columns: one column must contains
the names of the fields. The other column must
contains the associated condition for missing data
checking; if it is NA, then it means that no
branching logic is associated to the field.
}
\description{
Create the conditions for missing data checking. If a variable has no
branching logic and it is NA, then the information is missing. If a
variable is missing and it has a branching logic, then it is missing
only if it is missing and the branching logic is not satisfied.
}
