% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_feature_extractor.R
\name{Cell_feature_extractor}
\alias{Cell_feature_extractor}
\title{Generates a cell feature matrix based on images and cell mask}
\usage{
Cell_feature_extractor(
  N_cores = 1,
  Image_directory = NULL,
  Mask_directory = NULL,
  Ordered_Channels = NULL,
  Compute_basic_features = TRUE,
  quantiles_to_calculate = c(0.25, 0.5, 0.75),
  Compute_texture_features = FALSE,
  Texture_pixel_distance = 1,
  Subcellular_compartment_parameter_list = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Image_directory}{Character specifying the path to the folder where images are present.}

\item{Mask_directory}{Character specifying the path to the folder where cell segmentation masks are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{Compute_basic_features}{A logical value indicating if expression values should be computed.}

\item{quantiles_to_calculate}{A numeric vector containing values between 0 and 1 indicating the expression quantiles to be computed (0.25, 0.5 and 0.75 by default).}

\item{Compute_texture_features}{A logical value indicating if texture features should be computed for all the markers measured.}

\item{Texture_pixel_distance}{If Compute_texture_features is TRUE, de pixel distance taken into account to compute the GLCM and texture features.}

\item{Subcellular_compartment_parameter_list}{A list containing parameters to be employed for subcellular feature expression computation (see details).}
}
\value{
A tibble containing cell data.
}
\description{
The function computes a cell feature matrix containing spatial coordinates, morphologic features and expression values of cells given an image and its corresponding cell segmentation mask.
}
\details{
The function first performs name matching between Images and their cell masks. If non-unique matches are found, the function will
return an error. By default, the function will compute basic cell shape and location. In addition, the user may request the computation of feature expression data for
every cell, the texture features of marker expression and subcellular marker expression. In order to perform subcellular expression computation, the user needs to provide
the Subcellular_compartment_parameter_list.

The Subcellular_compartment_parameter_list should be a named list. The names of the list should be the name of the subcellular compartment to be analyzed.
Each element should be a list with the following named elements: Channel_name, Type, Threshold_type, Threshold_value, Blurr, Sigma.
These parameters should follow the same rules as the ones used to calculate tissueMask in the \code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}} or the \code{\link[=MFI_Experimet_Calculator]{MFI_Experimet_Calculator()}} (see examples).

\itemize{
\item{Channel_name: A single character value or a character vector indicating the channels present In Ordered_Channels that should be used for subcellular mask generation.}
\item{Type: One of the following: Positive - the subcellular mask will be computed and applied directly. Negative - the inverse (negative) of the subcellular mask will be used.}
\item{Threshold_type: One of the following: Arbitrary, Otsu.}
\item{Threshold_value: If Threshold_type is arbitrary, the threshold value to be applied.}
\item{Blurr: A logical value indicating if image blurring should be performed before thresholding.}
\item{Sigma: If Blurr is set to TRUE, the Sigma value indicates the sigma for Gaussian blurr filtering.}
}
}
\examples{
\dontrun{
#Create temporary input directory----------------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

Output_Dir <- tempfile(pattern = "tempdir2_Output")
dir.create(Output_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
          function(Image){
            EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
                                file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
         })

#Check a segmentation parameters list obtained using the dedicated function----------
print(CSM_SegmentParams_test)


#Generate the whole cell masks using watershed---------------------------------------
Watershed_cell_mask_generator(
 Directory = Input_Dir,
 Output_directory = Output_Dir,
 Parameter_list = CSM_SegmentParams_test
)

#Prepare parameters to obtain subcellular nuclear and cytoplasmic compartments---
Subcellular_compartment_parameters <- 
 list(Nuclear = list(Channel_name = "DAPI", 
                     Type = "Positive",
                     Threshold_type = "Arbitrary", 
                     Threshold_value = 0.05,
                     Blurr = TRUE, 
                     Sigma = 0.4),
      Cytoplasmic = list(Channel_name = "DAPI", 
                         Type = "Negative",
                         Threshold_type = "Arbitrary", 
                         Threshold_value = 0.05,
                         Blurr = TRUE, 
                         Sigma = 0.4)
 )

#Obtain the cell feature matrix--------------------------------------------
Cell_feature_extractor(Image_directory = Input_Dir,
                      Mask_directory = Output_Dir,
                      Ordered_Channels = CSM_SegmentParams_test[["Channels_to_keep"]],
                      Subcellular_compartment_parameter_list = Subcellular_compartment_parameters)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
unlink(Output_Dir, recursive = TRUE)
}

}
\seealso{
\code{\link[=Watershed_cell_mask_generator]{Watershed_cell_mask_generator()}}, \code{\link[=Segmentator_tester_app]{Segmentator_tester_app()}}
}
