% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_image_plot_generator.R
\name{Cell_image_plot_generator}
\alias{Cell_image_plot_generator}
\title{Plot an image with a cell overlay.}
\usage{
Cell_image_plot_generator(
  Image_directory,
  Channel_to_display = 1,
  Image_rotate = 0,
  Image_x_flip = FALSE,
  Image_y_flip = FALSE,
  Gamma_level = 0,
  Equalize = FALSE,
  Black_level = 0,
  White_level = 100,
  DATA,
  Image_name,
  Color_by = NULL,
  Point_size = 1,
  Pixel_distance_ratio = NULL
)
}
\arguments{
\item{Image_directory}{Character string specifying the path to the image file that needs to be plotted.}

\item{Channel_to_display}{A integer value indicating the Channel index to be displayed (1 for single channel images).}

\item{Image_rotate}{(OPTIONAL) A integer value indicating the degrees of rotation of the image.}

\item{Image_x_flip}{A logical value indicating if X image flip should be performed.}

\item{Image_y_flip}{A logical value indicating if Y image flip should be performed.}

\item{Gamma_level}{A numeric value between -3 and +3 indicating the gamma level to apply to the image.}

\item{Equalize}{A logical value indicating if the image should be equalized.}

\item{Black_level}{A integer value between 0 and 100 indicating the minimum pixel intensity.}

\item{White_level}{A integer value between 0 and 100 indicating the maximum pixel intensity.}

\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Image_name}{A character value indicating the image to be plotted.}

\item{Color_by}{(OPTIONAL) A character value indicating the column name in DATA that should be used to color cell points.}

\item{Point_size}{A numeric value indicating the size of the points.}

\item{Pixel_distance_ratio}{(OPTIONAL) A numeric value indicating the ratio between pixel size / cell coordinates. Use this argument when cell coordinates have been transformed to a distance unit like microns.}
}
\value{
Generates a plot with the image and the cells overlay.
}
\description{
The function generates a plot with an image in the background and cells overlayed. Cell information must have been obtained from the image in order to match X Y coordinates with image pixel coordinates.
}
\examples{
\dontrun{
#Create temporary directory------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Draw the plot
Cell_image_plot_generator(
    Image_directory = dir(Input_Dir, full.names = TRUE)[2],
    Channel_to_display = 5,
    Image_rotate = NULL,
    Image_x_flip = FALSE,
    Image_y_flip = TRUE,
    Gamma_level = 0,
    Equalize = FALSE,
    Black_level = 1,
    White_level = 20,

    DATA = CSM_Phenotypecell_test,
    Image_name ="ABCM22001_B14_MiniCrop.tif",
    Color_by = "Phenotype",
    Point_size = 2,
    Pixel_distance_ratio = NULL
)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
}


}
\seealso{
\code{\link[=Cell_plotter]{Cell_plotter()}}
}
