% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Channel_Sample_diagnostics.R
\name{Channel_Sample_diagnostics}
\alias{Channel_Sample_diagnostics}
\title{Identify image issues based on correlation hubs}
\usage{
Channel_Sample_diagnostics(
  N_cores = 1,
  Directory = NULL,
  Ordered_Channels = NULL,
  Pixel_downsize = NULL,
  Channel_for_test = NULL,
  Channels_to_keep = NULL,
  Threshold_type_tissueMask = NULL,
  Threshold_value_tissueMask = NULL,
  Blurr_tissueMask = FALSE,
  Sigma_tissueMask = NULL,
  DATA = NULL,
  List_correlation_hubs,
  Spearman_threshold = 0.05
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Directory}{Character string specifying the path to the folder where images are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{Pixel_downsize}{(OPTIONAL) A numeric value between 0 and 1 indicating the percentage of pixel downsize. If NULL no pixel downsize will be performed.}

\item{Channel_for_test}{A character vector indicating the channel name to be displayed in the pre-processing test.}

\item{Channels_to_keep}{Character vector indicating channels used in tissue mask generation.}

\item{Threshold_type_tissueMask}{Type of threshold to performe tissue mask. Either 'Otsu', 'Arbitrary' or 'Absolute'.}

\item{Threshold_value_tissueMask}{Numeric value used if Arbitrary is the threshold type of choice.}

\item{Blurr_tissueMask}{Logical value indicating if image blurring be performed before tissue mask generation.}

\item{Sigma_tissueMask}{Numeric value indicating the sigma value to perform Gaussian blurring.}

\item{DATA}{A tibble containing the cell features.}

\item{List_correlation_hubs}{A list containing vectors of channel names where positive correlation is expected.}

\item{Spearman_threshold}{A single numeric value indicating the positive threshold to consider the correlation value to be adequate (0.05 by default).}
}
\value{
A tibble containing flagged image channel and samples according to correlation metrics
}
\description{
Identify image issues based on correlation hubs
}
\details{
The function is based on the prior knowledge of protein/biomolecule spatial distribution. If two proteins are expected to be expressed in the same cell/tissue structure they should show
positive correlation. If the correlation is not found, this may warrant the review of specific images or channels. The function is based on computing the Spearman correlation index.
In order to ameliorate the computational burden, the images can be downsized with the Pixel_downsize argument.
}
\examples{
\dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
          function(Image){
            EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
                               file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
          })
          
#Generate list of correlation hubs
Correlation_hubs <- list(Nuclear_markers = c("DAPI", "FOXP3"),
                        Tumor_markers = c("CK-EPCAM", "PDL1"),
                        Lymphocyte_markers = c("CD8a", "GZMB", "PD1"))


#Run the test
Channel_Sample_diagnostics(
  N_cores = 1,
 
  Directory = Input_Dir,
  Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3", "RANDOM_1", "RANDOM_2", "RANDOM_3", "RANDOM_4"),
  Pixel_downsize = 0.2,
  Channel_for_test = "DAPI",
  Channels_to_keep = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3"),
  Threshold_type_tissueMask = "Arbitrary",
  Threshold_value_tissueMask = 0.05,
  Blurr_tissueMask = FALSE,
  Sigma_tissueMask = NULL,
 
  DATA = NULL,
 
  List_correlation_hubs = Correlation_hubs,
  Spearman_threshold = 0.05
)
}


}
\seealso{
\code{\link[=Data_QC_Check_function]{Data_QC_Check_function()}}
}
