% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinical_Data_analyzer.R
\name{Clinical_Data_analyzer}
\alias{Clinical_Data_analyzer}
\title{Calculates association between image feature data and image metadata}
\usage{
Clinical_Data_analyzer(
  DATA,
  DATA_var,
  DATA_Clinical,
  Clinical_var,
  Perform_time_to_event = FALSE,
  Time_variable = NULL,
  Event_variable = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing image feature data.}

\item{DATA_var}{A character vector indicating which features from DATA will be analyzed.}

\item{DATA_Clinical}{A dataframe or tibble containing image metadata, formatted using \code{\link[=Clinical_Data_arrange_function]{Clinical_Data_arrange_function()}}.}

\item{Clinical_var}{A character value indicating the name of the column to be analyzed from image metadata.}

\item{Perform_time_to_event}{A logical value indicating if time to event analysis should be performed (see details).}

\item{Time_variable}{If time to event is required, this argument is used to provide the column name specifying the time variable (Numeric).}

\item{Event_variable}{If time to event is required, this argument is used to provide the column name specifying the event variable (a numeric variable containing 0 or 1 values).}
}
\value{
Returns a summary of association and generates summary plots.
}
\description{
\code{Clinical_Data_analyzer()} will calculate the association between image feature data (a numeric variable) and image metadata.
Image feature data can be obtained in various ways using CSM, for example running \code{\link[=Phenotype_quantifier]{Phenotype_quantifier()}}, \code{\link[=Global_heterogeneity_calculator]{Global_heterogeneity_calculator()}}, \code{\link[=Distance_analyzer]{Distance_analyzer()}} among others.
}
\details{
If image metadata variable is a character, the function will calculate t-tests or ANOVA depending on the number of categories.
If image metadata is numeric, Pearson correlation will be calculated. For time-to event analysis, cut-off points for the image feature will be calculated using the survminer::surv_cutpoint function.
}
\examples{
\dontrun{
#Calculate the area of TMA samples-------------------------------------------
DATA_AREA <-
 Image_size_calculator(
    DATA = CSM_PhenotypeTMA_test,
    Strategy = "Concave_hull",
    Hull_ratio = 0.4
)

#Calculate the cell densities by samples-------------------------------------
Cells_by_sample <-
 Phenotype_quantifier(
    CSM_PhenotypeTMA_test,
    Calculate_Density = TRUE,
    DATA_Area = DATA_AREA
)

#Obtain clinical data--------------------------------------------------------
DATA_CLINICAL <-
Clinical_Data_arrange_function(
     DATA = CSM_ClinicalTMA_test,
     Subject_Names = "Sample",
     Outcomes_to_keep = c("AGE", "MMRP_status", "DEATH", "OS_m")
)

#Analyze sample features-----------------------------------------------------
Clinical_Data_analyzer(
   DATA = Cells_by_sample,
   DATA_var = c("Density_CD8", "Density_MACROPHAGE"),
   DATA_Clinical = DATA_CLINICAL,
   Clinical_var = "MMRP_status",
   Perform_time_to_event = FALSE
)

Clinical_Data_analyzer(
   DATA = Cells_by_sample,
   DATA_var = c("Density_CD8", "Density_MACROPHAGE"),
   DATA_Clinical = DATA_CLINICAL,
   Clinical_var = "AGE",
   Perform_time_to_event = FALSE
)

Clinical_Data_analyzer(
   DATA = Cells_by_sample,
   DATA_var = c("Density_CD8", "Density_MACROPHAGE"),
   DATA_Clinical = DATA_CLINICAL,
   Perform_time_to_event = TRUE,
   Time_variable = "OS_m",
   Event_variable = "DEATH"
)
}

}
