% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustered_Tiled_Images_analyzer.R
\name{Clustered_Tiled_Images_analyzer}
\alias{Clustered_Tiled_Images_analyzer}
\title{Generates a summary of neighborhood analysis based on clustering of tiles}
\usage{
Clustered_Tiled_Images_analyzer(
  Tiled_images,
  Perform_heterogeneity_analysis = FALSE,
  Graph_Modularity_resolution = NULL
)
}
\arguments{
\item{Tiled_images}{A list containing neighborhood information generated using \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}} function.}

\item{Perform_heterogeneity_analysis}{A logical value indicating if heterogeneity analysis should be performed.}

\item{Graph_Modularity_resolution}{The resolution parameter for graph modularity analysis.}
}
\value{
Generates a tibble containing neighborhood summary by sample and optionally heterogeneity metrics.
}
\description{
The function returns neighborhood counts and percentages. Optionally the function may compute neighborhood heterogeneity metrics.
}
\details{
Shannon, Simpson,Inverse Simpson and Renyi entropy indexes are calculated using the vegan package.

Rao index is calculated using the picante::raoD function.

Gini index is calculated using the DescTools::Gini function.

Graph modularity is computed using the igraph::modularity function.
}
\examples{
\dontrun{
#Tile images with cell phenotype information---------------------------------
Tiled_Images <-
 Image_tiling_processing_function(
   N_cores = 1,
   DATA = CSM_Phenotypecell_test,
   Tile_width = 125,
   Tile_height = 125,
   Variables_to_keep = "Phenotype"
)

#Cluster cell composition by tile to find neighborhoods---------------------
Clustered_Tiled_Images <-
Tiled_Image_Clustering_function(
    Tiled_images = Tiled_Images,
    Minimum_cell_no_per_tile = 4,
    Minimum_valid_tiles_per_image = 4,
    Phenotypes_included = unique(CSM_Phenotypecell_test$Phenotype),

    Cluster_Data = "Cell_Density",

    Perform_Dimension_reduction = FALSE,
    Cluster_on_Reduced = FALSE,

   Strategy = "Consensus_Clustering",
   Max_N_Clusters = 5,
   Consensus_reps = 2,
   Consensus_p_Items = 1,
   Consensus_Cluster_Alg = "pam",
   Consensus_Distance = "euclidean",
   Consensus_Name = "Consensus_clustering_test"
)

#Run the analysis-------------------------------------------------------------
Clustered_Tiled_Images_analyzer(
   Tiled_images = Clustered_Tiled_Images,
   Perform_heterogeneity_analysis = TRUE,
   Graph_Modularity_resolution = 0.5
)
}

}
