% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustered_Tiled_Images_graphicator.R
\name{Clustered_Tiled_Images_graphicator}
\alias{Clustered_Tiled_Images_graphicator}
\title{Generates plots of tiled neighborhood identity}
\usage{
Clustered_Tiled_Images_graphicator(Tiled_images = NULL, Image_name = NULL)
}
\arguments{
\item{Tiled_images}{A list containing tiled images obtained using \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}.}

\item{Image_name}{A character value indicating the image to be plotted.}
}
\value{
A plot of the neighborhood identity of each tile.
}
\description{
The function generates plots of the neighborhood labels of each tile after running the \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}.
}
