% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_deconvolution_App_launcher.R
\name{Color_deconvolution_App_launcher}
\alias{Color_deconvolution_App_launcher}
\title{Launches a shiny APP to explore color deconvolution parameters}
\usage{
Color_deconvolution_App_launcher(Directory = NULL)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where RGB color images are present.}
}
\description{
\code{Color_deconvolution_App_launcher()} launches an APP to interactively explore color deconvolution parameters
Parameters can then be used to feed the \code{\link[=Image_deconvolution_function]{Image_deconvolution_function()}}.
}
