% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordance_calculator.R
\name{Concordance_calculator}
\alias{Concordance_calculator}
\title{Calculates concordance between different methods.}
\usage{
Concordance_calculator(..., Variable, Strategy)
}
\arguments{
\item{...}{Two or more dataframes or tibbles containing a column with cell labels. A name must be provided for every tibble.}

\item{Variable}{A character value of the column name containing the labels. The name must be shared across datasets provided.}

\item{Strategy}{One of the following: 'Rand' or 'FM' (Fowlkes–Mallows) (see details).}
}
\value{
Returns a tibble containing concordance values by sample. Generates summary plots.
}
\description{
The function calculates concordance between two or more methods that generate cell labels.
}
\details{
Rand index is calculated using catsim::rand_index function.

Fowlkes-Mallows index is calculated using dendextend::FM_index function.
}
\examples{
\dontrun{
#Generate datasets with random cell phenotype labels----------------------
CSM_Phenotypecell_test_1 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))
CSM_Phenotypecell_test_2 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))
CSM_Phenotypecell_test_3 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))
#Find concordance between results--------------------------------------
Concordance_calculator(
  Random_1 = CSM_Phenotypecell_test_1,
  Random_2 = CSM_Phenotypecell_test_2,
  Random_3 = CSM_Phenotypecell_test_3,
  Variable = "Phenotype",
  Strategy = "FM"
)
}

}
