% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confusion_matrix_plotter.R
\name{Confusion_matrix_plotter}
\alias{Confusion_matrix_plotter}
\title{Generates a confusion matrix between two methods used to label cells.}
\usage{
Confusion_matrix_plotter(..., Variable = NULL)
}
\arguments{
\item{...}{Two dataframes or tibbles containing a column with cell labels.}

\item{Variable}{A character value of the column name containing the labels. The name must be shared across the two datasets provided.}
}
\value{
Generates a confusion matrix plot
}
\description{
The function generates a confusion matrix between two cell labeling methods (cell phenotype or neighborhood labels for example).
}
\examples{
\dontrun{
#Generate datasets with random cell phenotype labels----------------------
CSM_Phenotypecell_test_1 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))
CSM_Phenotypecell_test_2 <-
 CSM_Phenotypecell_test \%>\%
     mutate(Phenotype = sample(unique(CSM_Phenotypecell_test$Phenotype),
                               size = nrow(CSM_Phenotypecell_test),
                               replace = TRUE))

#Generate the confusion matrix-------------------------------------------

Confusion_matrix_plotter(
   Random_1 = CSM_Phenotypecell_test_1,
   Random_2 = CSM_Phenotypecell_test_2,
   Variable = "Phenotype"
   )
}

}
