% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM-package.R
\docType{package}
\name{CSM-package}
\alias{CSM}
\alias{CSM-package}
\title{CSM: Toolbox to analyze tissue images with spatial resolution}
\description{
Contains functions to perform tissue spatial analysis. Functions are organized in 7 modules. MODULE0 Generates cell and pixel level information and formats data. MODULE1 normalizes cell features. MODULE2 thresholds features. MODULE3 performs cell phenotyping. MODULE4 analyzes heterogeneity. MODULE5 performs cell-to-cell spatial interaction analysis. MODULE6 performs neighborhood analysis.
}
\author{
\strong{Maintainer}: Alvaro Lopez-Janeiro \email{alopezj@unav.es} (\href{https://orcid.org/0000-0002-5744-3115}{ORCID})

Other contributors:
\itemize{
  \item Paulo Perez-Dominguez \email{pperezdomin@unav.es} [contributor]
  \item Eduardo Miraval \email{emiraval@unav.es} [contributor]
  \item Raluca Alexandru \email{ralexandru@unav.es} [contributor]
  \item Carlos E. de Andrea \email{ceandrea@unav.es} [contributor]
}

}
\keyword{internal}
