% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm-package.R
\docType{package}
\name{csm-package}
\alias{csm}
\alias{csm-package}
\title{csm: Central Statistical Monitoring}
\description{
The goal of `csm` is to build a Central Statistical 
    Monitoring (CSM) tool that aims to systematically check data quality 
    of the clinical studies managed by the Center for Clinical Trials 
    and Biometrics (CCTB) of the *Unit of Biostatistics, Epidemiology and 
    Public Health (UBEPH)*, University of Padova, Italy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/csm}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/csm/issues}
}

}
\author{
\strong{Maintainer}: Daniele Bottigliengo \email{daniele.bottigliengo@phd.unipd.it}

}
\keyword{internal}
