% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cumulative_Interaction_generator.R
\name{Cumulative_Interaction_generator}
\alias{Cumulative_Interaction_generator}
\title{Calculates the cumulative interaction matrix}
\usage{
Cumulative_Interaction_generator(
  N_cores = 1,
  DATA,
  Start_from,
  Stop_at,
  Sampling_frequency
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A list containing cell distance matrix. It can be computed using \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}} or \code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}} functions.}

\item{Start_from}{A numeric value indicating the minimum distance to be analyzed.}

\item{Stop_at}{A numeric value indicating the maximum distance to be analyzed.}

\item{Sampling_frequency}{A numeric value indicating the distance sampling frequency.}
}
\value{
A list containing the cumulative interaction matrix for every image.
}
\description{
The cumulative interaction matrix contains information regarding the total potential interactions between COO and target cells. For a a given set of increasing distances,
the number of target cells present from every COO is calculated.
}
\examples{
\dontrun{
#' #Generate a data distance or random distance matrix--------------------------
DATA_Distances <-
Distance_matrix_generator(
    N_cores = 1,
    DATA = CSM_Phenotypecell_test,
    Cell_Of_Origin = "CD8_GZMBneg",
    Target_Cell = "TUMOR",
    Allow_Cero_Distance = FALSE,
    Perform_edge_correction = FALSE
)

#Calculate the cumulative interaction every 25 pixels up to 100 pixels------
Cumulative_Interaction_generator(
   N_cores = 1,
   DATA = DATA_Distances,
   Start_from = 25,
   Stop_at = 100,
   Sampling_frequency = 25
)
}

}
\seealso{
\code{\link[=Random_Distance_matrix_generator]{Random_Distance_matrix_generator()}}, \code{\link[=Distance_matrix_generator]{Distance_matrix_generator()}}, \code{\link[=Distance_analyzer]{Distance_analyzer()}}, \code{\link[=Cells_in_Radius_analyzer]{Cells_in_Radius_analyzer()}}.
}
