% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Describe_my_computing_unit.R
\name{Describe_my_computing_unit}
\alias{Describe_my_computing_unit}
\title{Obtain CPU information and available RAM}
\usage{
Describe_my_computing_unit()
}
\value{
A list containing information regarding CPU, RAM, R version being used and Operating system
}
\description{
The function returns machine characteristics like CPU model and number of cores. It will return the available RAM. It will also return the OS and R version.
}
\examples{
\dontrun{
Describe_my_computing_unit()
}

}
