% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid.unique.R
\name{expand.grid.unique}
\alias{expand.grid.unique}
\title{Generates a tibble with unique combinations between two vectors}
\usage{
expand.grid.unique(x, y, include.equals = FALSE)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}

\item{include.equals}{A logical value}
}
\value{
A dataframe containing unique combinations
}
\description{
Intended for internal use only. In contrast to expand.grid only unique combinations are preserved.
}
\details{
Used in \code{\link[=Concordance_calculator]{Concordance_calculator()}}, \code{\link[=Neighborhood_discovery_function]{Neighborhood_discovery_function()}}, \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}, \code{\link[=Clustered_Tiled_Images_analyzer]{Clustered_Tiled_Images_analyzer()}}
}
\keyword{Internal}
