% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_based_phenotyper_App_launcher.R
\name{Image_based_phenotyper_App_launcher}
\alias{Image_based_phenotyper_App_launcher}
\title{Launches a shiny APP that can generate cell phenotyping models based on user expertise}
\usage{
Image_based_phenotyper_App_launcher(
  DATA = NULL,
  Directory = NULL,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Directory}{Character specifying the path to the folder where images are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{Channels_to_keep}{Character vector indicating the channels to be kept in the analysis.}
}
\description{
Launches an APP to interactively explore cell segmented images. The user can manually assign cell phenotype labels to cells and train a model.
The model can then be used in \code{\link[=Model_cell_phenotyper]{Model_cell_phenotyper()}} function to generalize it to all the cells in a dataset.
}
\seealso{
\code{\link[=Model_cell_phenotyper]{Model_cell_phenotyper()}}
}
