% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_deconvolution_function.R
\name{Image_deconvolution_function}
\alias{Image_deconvolution_function}
\title{Separates RGB color images into color channels according to user preferences}
\usage{
Image_deconvolution_function(
  Directory = NULL,
  Output_directory = NULL,
  Deconvolution_parameters = NULL,
  N_cores = NULL
)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where RGB images are present.}

\item{Output_directory}{Character string specifying the path to the folder where  output images are written. It must be an empty folder.}

\item{Deconvolution_parameters}{A list containing parameters. Created using \code{\link[=Color_deconvolution_App_launcher]{Color_deconvolution_App_launcher()}}.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
The function writes multi-tiff images in the ouput directory.
}
\description{
\code{Image_deconvolution_function()} generates multi-channel tiff images according to color deconvolution parameters.
}
\examples{
|dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
Output_Dir <- tempfile(pattern = "tempdir2_Output")
dir.create(Input_Dir, recursive = TRUE)
dir.create(Output_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniHE_test[[Image]], file.path(Input_Dir, names(CSM_MiniHE_test)[Image]))
})

#Print an example of the deconvolution parameters obtained using the App------------
print(CSM_colordeconv_test)

#Run color channel separation------------------------------------------
Image_deconvolution_function(
       Directory = Input_Dir,
       Output_directory = Output_Dir,
       Deconvolution_parameters = CSM_colordeconv_test,
       N_cores = 1
)

#Check that files have been created-----------------------------------------
list.files(Output_Dir)

#Remove directories---------------------------------------------------------
unlink(c(Input_Dir, Output_Dir), recursive = TRUE)
}

}
\seealso{
\code{\link[=Color_deconvolution_App_launcher]{Color_deconvolution_App_launcher()}}
}
