% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_from_tile_rebuilder.R
\name{Image_from_tile_rebuilder}
\alias{Image_from_tile_rebuilder}
\title{Build images from tile image chunks}
\usage{
Image_from_tile_rebuilder(
  Directory = NULL,
  Output_directory = NULL,
  RGB_Color_images = NULL,
  N_cores = NULL
)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where image tiles are present.}

\item{Output_directory}{Character string specifying the path to the folder where  output images are written. It must be an empty folder.}

\item{RGB_Color_images}{Logical. Is the image a RGB color image?}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
The function writes the images in the ouput directory.
}
\description{
\code{Image_from_tile_rebuilder()} builds a complete image from it's tiles.
Warning: Image reconstruction is limited to ≈ 2-3Gb.
Image reconstruction is designed to rebuild complete images from tiles once these have undergone pixel thresholding using \code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}}.
}
