% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_thresholding_app_launcher.R
\name{Image_thresholding_app_launcher}
\alias{Image_thresholding_app_launcher}
\title{Launches a shiny APP to explore image pixel thresholding parameters}
\usage{
Image_thresholding_app_launcher(Directory, Ordered_Channels)
}
\arguments{
\item{Directory}{Character specifying the path to the folder where images to be thresholded are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}
}
\description{
\code{Image_thresholding_app_launcher()} launches an APP to interactively explore image pixel segmentation parameters interactively.
}
\details{
Image parameters control the image and channel to be displayed.

Tissue mask parameters allow customizing how the tissue mask will be computed (decide foreground / background pixels).

Tissue thresholding parameters control the tresholding process. If arbitrary, the "value" box is used to write the desired threshold value.
If Multilevel, the "value" box can be used to write a vector with the thresholds (for example "c(0.1, 0.5, 0.9)") or the number of levels
can be specified with the Levels box. If "value" is not provided, the app will calculate threshold values using the imagerExtra::ThresholdML function.

The GO! button executes the thresholding according to active parameters.

Image Panels:
\itemize{
\item{Upper left: Image being thresholded. Use it to zoom in.}
\item{Upper right: Tissue mask. Shows the tissue mask generated.}
\item{Lower left: Pixels above the threshold included in the tissue mask.}
\item{Lower right: Final result after applying blurring to the target (if required).}
}
}
\examples{
\dontrun{
#Create temporary directory------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Deploy app----------------------------------------------
Image_thresholding_app_launcher(
    Directory = Input_Dir,
    Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3")
)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
}

}
\seealso{
\code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}}, \code{\link[=Binary_threshold_image_combinator]{Binary_threshold_image_combinator()}}, \code{\link[=MFI_Experimet_Calculator]{MFI_Experimet_Calculator()}}
}
