% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_thresholding_app_launcher.R
\name{Image_thresholding_app_launcher}
\alias{Image_thresholding_app_launcher}
\title{Launches a shiny APP to explore image pixel thresholding parameters}
\usage{
Image_thresholding_app_launcher(Directory = NULL, Ordered_Channels = NULL)
}
\arguments{
\item{Directory}{Character specifying the path to the folder where images to be thresholded are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}
}
\description{
\code{Image_thresholding_app_launcher()} launches an APP to interactively explore image pixel segmentation parameters interactively.
}
\examples{
\dontrun{
#Create temporary directory------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]], file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Deploy app----------------------------------------------
Image_thresholding_app_launcher(
    Directory = Input_Dir,
    Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3")
)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
}

}
\seealso{
\code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}}, \code{\link[=Binary_threshold_image_combinator]{Binary_threshold_image_combinator()}}, \code{\link[=MFI_Experimet_Calculator]{MFI_Experimet_Calculator()}}
}
