% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_tile_deconstruction_function.R
\name{Image_tile_deconstruction_function}
\alias{Image_tile_deconstruction_function}
\title{Divide images into tile chuncks}
\usage{
Image_tile_deconstruction_function(
  Directory = NULL,
  Output_directory = NULL,
  RGB_Color_images = NULL,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL,
  Tile_pixel_size = NULL,
  Tile_Overlap = 0,
  N_cores = NULL
)
}
\arguments{
\item{Directory}{Character string specifying the path to the folder where images to be chopped into tiles are present.}

\item{Output_directory}{Character string specifying the path to the folder where  output images are written. It must be an empty folder.}

\item{RGB_Color_images}{Logical. Is the image a RGB color image?}

\item{Ordered_Channels}{Character vector specifying image channels. If RGB_Color_images is TRUE it will be set automatically to c("R", "G", "B").}

\item{Channels_to_keep}{Character vector specifying image channels to be kept after image chopping.}

\item{Tile_pixel_size}{Integer specifying the size of the tiles in pixels.}

\item{Tile_Overlap}{Integer specifying the overlap amount between tiles in pixels.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
The function writes the tiles in the ouput directory. The names of the resulting images should not be modified. If modified this can
result in unexpected results when using companion functions \code{\link[=Image_from_tile_rebuilder]{Image_from_tile_rebuilder()}}, \code{\link[=Tile_to_image_cell_arrange_function]{Tile_to_image_cell_arrange_function()}}.
}
\description{
\code{Image_tile_deconstruction_function()} divides into smaller chunks all the images in a directory.
This allows faster computations when using CSM apps, performing image thresholding and cell segmentation.
}
