% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_tiling_processing_function.R
\name{Image_tiling_processing_function}
\alias{Image_tiling_processing_function}
\title{Performs image tiling.}
\usage{
Image_tiling_processing_function(
  N_cores = NULL,
  DATA = NULL,
  Tile_width = NULL,
  Tile_height = NULL,
  Variables_to_keep = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Tile_width}{A numeric value indicating the width of the tiles.}

\item{Tile_height}{A numeric value indicating the height of the tiles.}

\item{Variables_to_keep}{A character vector indicating the column names to be kept when tiling the image.}
}
\value{
A list containing tile information for every image in DATA.
}
\description{
The function assigns cells to unique tiles. Afterwards, tile information can be used to run subsequent analyses.
}
\seealso{
\code{\link[=Image_length_calculator]{Image_length_calculator()}}, \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}
}
