% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_tiling_processing_function.R
\name{Image_tiling_processing_function}
\alias{Image_tiling_processing_function}
\title{Performs image tiling.}
\usage{
Image_tiling_processing_function(
  N_cores = 1,
  DATA,
  Tile_width,
  Tile_height,
  Variables_to_keep
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Tile_width}{A numeric value indicating the width of the tiles.}

\item{Tile_height}{A numeric value indicating the height of the tiles.}

\item{Variables_to_keep}{A character vector indicating the names of the columns to be kept when tiling the image.}
}
\value{
A list containing tile information for every image in DATA.
}
\description{
The function assigns cells to unique tiles. Afterwards, tile information can be used to run subsequent analyses.
}
\examples{
\dontrun{
Image_tiling_processing_function(
   N_cores = 2,
   DATA = CSM_Phenotypecell_test,
   Tile_width = 125,
   Tile_height = 125,
   Variables_to_keep = "Phenotype"
)
}

}
\seealso{
\code{\link[=Image_length_calculator]{Image_length_calculator()}}, \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}
}
