% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker_combinator_generator.R
\name{Marker_combinator_generator}
\alias{Marker_combinator_generator}
\title{Generates all thresholded feature combinations present in a thresholded cell feature dataset}
\usage{
Marker_combinator_generator(DATA = NULL, Markers = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing thresholded cell feature data.}

\item{Markers}{A character vector indicating which features need to be used.}
}
\value{
Returns a tibble where each row are unique positive feature combinations found in DATA.
}
\description{
Given a thresholded cell feature dataset, the function will find all unique combinations of positive features present. These combinations are actually cell phenotypes.
The result can then be used to feed the \code{\link[=Phenotype_assigner_function]{Phenotype_assigner_function()}} to obtain cell phenotypes.
}
\seealso{
\code{\link[=Phenotype_assigner_function]{Phenotype_assigner_function()}}
}
