% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFI_Experimet_Calculator.R
\name{MFI_Experimet_Calculator}
\alias{MFI_Experimet_Calculator}
\title{Calculates MFI for a set of images}
\usage{
MFI_Experimet_Calculator(
  N_cores = 1,
  Directory = NULL,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL,
  Target_channel = NULL,
  Target_masks = NULL,
  Threshold_type_tissueMask = NULL,
  Threshold_value_tissueMask = NULL,
  Blurr_tissueMask = NULL,
  Sigma_tissueMask = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Directory}{Character string specifying the path to the folder where images are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{Channels_to_keep}{Character vector indicating channels used in tissue mask generation.}

\item{Target_channel}{Character value indicating the target channel to be thresholded.}

\item{Target_masks}{(OPTIONAL) A named list of the parameters used to calculate the compartment where MFI will be calculated (see details).}

\item{Threshold_type_tissueMask}{Type of threshold to performe tissue mask. Either 'Otsu', 'Arbitrary' or 'Absolute'.}

\item{Threshold_value_tissueMask}{Numeric value used if Arbitrary is the threshold type of choice.}

\item{Blurr_tissueMask}{Logical value indicating if image blurring be performed before tissue mask generation.}

\item{Sigma_tissueMask}{Numeric value indicating the sigma value to perform Gaussian blurring.}
}
\value{
Returns a tibble with the MFI per image.
}
\description{
Calculates the Mean Fluorescence Intensity of a target channel for every image in a directory. MFI can be calculated inside compartments using the Target_mask argument.
}
\details{
Target_masks should be a named lists. The names of the list should be any of the Channels_to_keep. Each element should be a list with the following named elements:
Mask_name, Threshold_type, Threshold_value, Blurr, Sigma. These parameters should follow the same rules as the ones used to calculate tissueMask (see examples).
}
\examples{
\dontrun{
#Create temporary input and output directories------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory--------------------------------------------
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]], file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Optionally generate marker masks to obtain the target MFI----------------------------
#Calculate the target in a combined CK, CD8 mask that will be overlayed to the tissue mask
Target_mask_list <- list(CK = list(Mask_name = "CK-EPCAM",
                                  Threshold_type = "Arbitrary",
                                  Threshold_value = 0.1,
                                  Blurr = TRUE,
                                  Sigma = 1),
                      CD8 = list(Mask_name = "CD8a",
                                 Threshold_type = "Arbitrary",
                                 Threshold_value = 0.001,
                                 Blurr = TRUE,
                                 Sigma = 1)
)


#Then calculate the MFI---------------------------
MFI_Experimet_Calculator(
   N_cores = 1,
   Directory = Input_Dir,
   Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3"),
   Channels_to_keep = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3"),
   Target_channel = "GZMB",

   Target_masks = Target_mask_list,

   Threshold_type_tissueMask = "Arbitrary",
   Threshold_value_tissueMask = 0.01,
   Blurr_tissueMask = TRUE,
   Sigma_tissueMask = 0.5
)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
}

}
\seealso{
\code{\link[=Image_thresholding_app_launcher]{Image_thresholding_app_launcher()}}, \code{\link[=Binary_threshold_image_combinator]{Binary_threshold_image_combinator()}}, \code{\link[=MFI_Experimet_Calculator]{MFI_Experimet_Calculator()}}
}
