% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_cell_phenotyper.R
\name{Model_cell_phenotyper}
\alias{Model_cell_phenotyper}
\title{Executes cell phenotype label prediction}
\usage{
Model_cell_phenotyper(DATA, Model_parameters, N_cores = 1)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Model_parameters}{The model obtained using \code{\link[=Image_based_phenotyper_App_launcher]{Image_based_phenotyper_App_launcher()}}.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\description{
Given a cell feature dataset and a cell phenotype label prediction model, the function assigns cell phenotype labels to all cells in the dataset.
The model must have been fitted using \code{\link[=Image_based_phenotyper_App_launcher]{Image_based_phenotyper_App_launcher()}}.
}
\examples{
\dontrun{
#Check the model generated using the dedicated app
print(CSM_ModelParam_test)

#Use the model to assign cell labels to all cells in the dataset
Model_cell_phenotyper(
 DATA = CSM_Arrangedcellfeaturedata_test,
 Model_parameters = CSM_ModelParam_test,
 N_cores = 1
)
}


}
\seealso{
\code{\link[=Image_based_phenotyper_App_launcher]{Image_based_phenotyper_App_launcher()}}
}
