% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_level_modelling_function.R
\name{Multi_level_modelling_function}
\alias{Multi_level_modelling_function}
\title{Analyzes the spatial interaction between two cell types using Multi-level modelling}
\usage{
Multi_level_modelling_function(
  DATA_cumulative = NULL,
  DATA_Clinical = NULL,
  Clinical_var = NULL,
  DATA_Densities = NULL,
  Cell_Of_Origin = NULL,
  Target_cell = NULL,
  Calculate_R2 = NULL,
  N_bootstrap = NULL
)
}
\arguments{
\item{DATA_cumulative}{A list containing cumulative interaction data. It can be computed using \code{\link[=Cumulative_Interaction_generator]{Cumulative_Interaction_generator()}} function.}

\item{DATA_Clinical}{A dataframe or tibble containing image metadata.}

\item{Clinical_var}{A character value indicating the column name used in the clinical data analysis.}

\item{DATA_Densities}{A dataframe or tibble containind the cell phenotype density by image. It can be computed using \code{\link[=Phenotype_quantifier]{Phenotype_quantifier()}}.}

\item{Cell_Of_Origin}{A character value indicating the cell phenotype label of the Cell of Origin.}

\item{Target_cell}{A character value indicating the cell phenotype label of the Target cell.}

\item{Calculate_R2}{A logical value indicating if R2 should be calculated after model fitting.}

\item{N_bootstrap}{A integer value indicating the number of permutations to calculate R2 values.}
}
\value{
A list containing two outputs: Raw_DATA contains the expected interaction probability for every distance point. Simplified_DATA cotains the AUC data by image.
}
\description{
The function calculates a multilevel model to analyze the spatial interaction between two cell types. The spatial cumulative interaction between two cell types,
one being the Cell of Origin (COO) and the other being the target cell, is dependent on two main variables: the target cell density and distance from the cell of origin.
If the tissue target cell density is higher, or the distance from the COO is larger, the spatial interaction is going to increase.
The function tries to model the influence of image metadata on this spatial interaction. It does so by considering cells in the dataset as nested observations,
where cells within the same image share specific characteristics (like exposure to the same target cell density or image metadata condition).
The model formula will look something like this: N umber of Target cells ~ Image_Metadata*Distance_COO + Density_Target + (1|Cell_Of_Origin_ID)
The multilevel model is calculated using lme4::lmer function. The contribution of each variables to the total variance explanation is calculated using the partR2::partR2 function.
}
\examples{

\dontrun{
#Obtain the Cumulative interaction between two cells types accross the desired distance--------
DATA_Distances <-
Distance_matrix_generator(
    N_cores = 1,
    DATA = CSM_PhenotypeTMA_test,
    Cell_Of_Origin = "CD8",
    Target_Cell = "MACROPHAGE",
    Allow_Cero_Distance = FALSE,
    Perform_edge_correction = FALSE
)
DATA_Cumulative <-
Cumulative_Interaction_generator(
    N_cores = 1,
    DATA = DATA_Distances,
    Start_from = 10,
    Stop_at = 210,
    Sampling_frequency = 50
)

#Obtain cell density by sample to account for density of target cells in the model-------------
DATA_AREA <-
 Image_size_calculator(
    DATA = CSM_PhenotypeTMA_test,
    Strategy = "Concave_hull",
    Hull_ratio = 0.4
)
Cells_by_sample <-
 Phenotype_quantifier(
    CSM_PhenotypeTMA_test,
    Calculate_Density = TRUE,
    DATA_Area = DATA_AREA
)

#Arrange clinical data------------------------------------------------------------------------
DATA_CLINICAL <-
Clinical_Data_arrange_function(
     DATA = CSM_ClinicalTMA_test,
     Subject_Names = "Sample",
     Outcomes_to_keep = c("AGE", "MMRP_status", "DEATH", "OS_m")
)

#Calculate the model--------------------------------------------------------------------------
Multi_level_modelling_function(
    DATA_cumulative = DATA_Cumulative,
    DATA_Clinical = DATA_CLINICAL,
    Clinical_var = "MMRP_status",
    DATA_Densities = Cells_by_sample,
    Cell_Of_Origin = "CD8",
    Target_cell = "MACROPHAGE",
    Calculate_R2 = FALSE
)
}

}
