% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighborhood_voting_function.R
\name{Neighborhood_voting_function}
\alias{Neighborhood_voting_function}
\title{Performs elections within tiled images with neighborhood information}
\usage{
Neighborhood_voting_function(
  N_cores = 1,
  Tiled_Images,
  Minimum_cell_no_per_tile = 1,
  Neighborhoods_included
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Tiled_Images}{A list containing tiled images obtained using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}} with neighborhood information.}

\item{Minimum_cell_no_per_tile}{An integer indicating the minimum number of cells that a tile must contain. Tiles below the limit will not be included in the analysis.}

\item{Neighborhoods_included}{A character vector indicating the neighborhood labels that will be included in the analysis.}
}
\value{
A list containing image information with by-tile neighborhood counts and a tibble with a summary of the number of winner per image.
}
\description{
Given a tiled image, the function calculates neighborhood prevalence by tile. It then assigns neighborhood identity to each tile based on results.
}
\details{
The winner from each tile will be reported as ABSOLUTE if there is a single neighborhood represented in each tile, SUPER if the winner has more than 50\% of votes and simple if the winner has less than 50\% of votes.
If there are any ties, alphabetical order will break the tie.
}
\examples{
\dontrun{
#Tile images with neighborhood information-----------------------------------
Tiled_Images <-
 Image_tiling_processing_function(
    DATA = CSM_Neighborhoods_test,
    Tile_width = 125,
    Tile_height = 125,
    Variables_to_keep = "Neighborhood_assignment",
    N_cores = 1
)

#Celebrate elections in each tile---------------------------------------------
Neighborhood_voting_function(
    N_cores = 1,
    Tiled_Images = Tiled_Images,
    Minimum_cell_no_per_tile = 2,
    Neighborhoods_included = unique(CSM_Neighborhoods_test$Neighborhood_assignment)
)
}

}
