% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_function.R
\name{Normalization_function}
\alias{Normalization_function}
\title{Normalizes cell feature expression data}
\usage{
Normalization_function(DATA, Strategy, Parameters)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Strategy}{'mxnorm' or 'simpleSeg'.}

\item{Parameters}{List of parameters to be used in the normalization process (see details).}
}
\value{
Returns a tibble containing normalized cell feature data.
}
\description{
Normalization of cell feature data following two potential strategies included in mxnorm and simpleSeg packages.
}
\details{
For mxnorm normalization, the list must contain the following elements:
slide_id (The column name of the DATA that contains the slide ID. this can be the Subject_Names or a user supplied vector containing the slide identity of these cells),
image_id (The column name of the DATA that contains the image ID. Can be the same as slide_id),
marker_cols (The name of the columns to be normalized),
transform ("None", "log10", "mean_divide","log10_mean_divide"),
method ("None", "ComBat","Registration"),
method_override (OPTIONAL) ("specify a custom normalization method. See mxnorm package documentation for more details),
method_override_name (OPTIONAL) (specify the name of the user defined method).

For simpleSeg approach, the list must contain the following elements:
cells (The tibble containing the cell feature data),
markers (The name of the columns to be normalized),
imageID (The column name of the DATA that contains the image ID),
transformation (Choose between:  NULL, "asinh", "sqrt"),
method ("trim99", "minMax"),
cores (Irrelevant argument. If parallelization is required use N_cores argument and \code{\link[=Normalization_function_parallel]{Normalization_function_parallel()}} function).
}
\examples{
\dontrun{
#Set normalization parameters-----------
mxnorm_Parameters <-
list(
  slide_id = "Subject_Names",
  image_id = "Subject_Names",
  marker_cols = names(CSM_Arrangedcellfeaturedata_test[-c(1:4)]),
  transform = "None",
  method = "ComBat",
  method_override = NULL,
  method_override_name = NULL
)

mxnorm_Parameters_slide <-
list(
  slide_id = c(rep("Slide1", times = 229), rep("Slide2", times = 818)),
  image_id = "Subject_Names", #The column name of the DATA that contains the image ID
  marker_cols = names(CSM_Arrangedcellfeaturedata_test[-c(1:4)]),
  transform = "None",
  method = "ComBat",
  method_override = NULL,
  method_override_name = NULL
)

simpleSeg_Parameters <-
list(
  cells = CSM_Arrangedcellfeaturedata_test,
  markers = names(CSM_Arrangedcellfeaturedata_test[-c(1:4)]),
  imageID = "Subject_Names",
  transformation = "sqrt",
  method = c("trim99", "minMax"),
  cores = 1
)

#Run normalization----------------------
Normalization_function(
  CSM_Arrangedcellfeaturedata_test,
  Strategy = "mxnorm",
  Parameters = mxnorm_Parameters
)

Normalization_function(
  CSM_Arrangedcellfeaturedata_test,
  Strategy = "simpleSeg",
  Parameters = simpleSeg_Parameters
)
}

}
\seealso{
\code{\link[=Normalization_function_parallel]{Normalization_function_parallel()}}, \code{\link[=Normalization_diagnostics]{Normalization_diagnostics()}}
}
