% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotype_assigner_function.R
\name{Phenotype_assigner_function}
\alias{Phenotype_assigner_function}
\title{Assigns a phenotype to every cell according to user specified labels and thresholding results.}
\usage{
Phenotype_assigner_function(DATA = NULL, Phenotype_possibilities = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing thresholded cell feature data.}

\item{Phenotype_possibilities}{A tibble obtained using \code{\link[=Marker_combinator_generator]{Marker_combinator_generator()}} containing a column named 'Phenotype'.}
}
\value{
Returns a cell feature tibble containing a new column named 'Phenotype' where cell labels are present.
}
\description{
The function takes a thresholded cell feature matrix and a tibble obtained using \code{\link[=Marker_combinator_generator]{Marker_combinator_generator()}}. The resulting tibble obtained using \code{\link[=Marker_combinator_generator]{Marker_combinator_generator()}}
must have a new column named 'Phenotype' where the user specifies the name of the phenotype according to the feature positivity pattern. The function uses this information to assign cell phenotypes to all cells in DATA.
}
\examples{
#Threshold data-------------------------------------
DATA_thresholded <- Thresholding_function(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Strategy = "EBI_Otsu",
  Local_thresholding = FALSE,
  Method_autothreshold = "Otsu",
  number_iterations_TriClass = 20,
  Percentile = 0.5,
  Defined_threshold = 0.1,
  Levels = 3
)
#Find unique feature positivity combinations----------
Phenotype_possibilities <- Marker_combinator_generator(
  DATA = DATA_thresholded,
  Markers = names(DATA_thresholded)[-c(1:4)]
)
#Assign a phenotype to each combination------------
Phenotype_possibilities$Phenotype <- c('TUMOR', 'OTHER', 'CD8_GZMBneg', 'CD8_GZMBneg', 'OTHER', 'CD8_GZMBpos', 'CD8_GZMBpos')

#Perform phenotyping-------------------------------
Phenotype_assigner_function(
  DATA = DATA_thresholded,
  Phenotype_possibilities = Phenotype_possibilities
)


}
\seealso{
\code{\link[=Thresholding_function]{Thresholding_function()}}, \code{\link[=Thresholding_function_tailored]{Thresholding_function_tailored()}}, \code{\link[=Marker_combinator_generator]{Marker_combinator_generator()}}
}
