% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping_evaluator_shiny_app_launcher.R
\name{Phenotyping_evaluator_shiny_app_launcher}
\alias{Phenotyping_evaluator_shiny_app_launcher}
\title{Launches a shiny App to explore cell phenotyping results}
\usage{
Phenotyping_evaluator_shiny_app_launcher(
  DATA = NULL,
  Directory = NULL,
  Ordered_Channels = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data and a column named 'Phenotype' containing cell phenotype labels.}

\item{Directory}{Character specifying the path to the folder where images are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}
}
\description{
After phenotyping, cell phenotypes can be explored interactively proyected on original images.
}
