% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pixel_Multilevel_thresholder.R
\name{Pixel_Multilevel_thresholder}
\alias{Pixel_Multilevel_thresholder}
\title{Thresholds the foreground pixels of an image according to user parameters into several levels}
\usage{
Pixel_Multilevel_thresholder(
  Target = NULL,
  Tissue_mask = NULL,
  Threshold_values = NULL,
  Blurr = NULL,
  Sigma = NULL
)
}
\arguments{
\item{Target}{An EBImage type object with a single channel to be thresholded}

\item{Tissue_mask}{An EBImage type object of the binary tissue mask}

\item{Threshold_values}{The value applied to threshold the image}

\item{Blurr}{A logical value indicating if blurring should be applied before thresholding}

\item{Sigma}{Sigma applied to the Gaussian kernel for blurring}
}
\value{
An EBImage object containing the multi-level thresholded image
}
\description{
Internal use only
}
\details{
Used in \code{\link[=Image_thresholding_app_launcher]{Image_thresholding_app_launcher()}}, \code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}}
}
\keyword{Internal}
