% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pixel_Threshold_calculator.R
\name{Pixel_Threshold_calculator}
\alias{Pixel_Threshold_calculator}
\title{Performs image pixel thresholding according to user preferences}
\usage{
Pixel_Threshold_calculator(
  N_cores = NULL,
  Directory = NULL,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL,
  Target_channel = NULL,
  Save_processed_images = NULL,
  Output_Directory = NULL,
  Local_thresholding = NULL,
  Threshold_type = NULL,
  Threshold_value = NULL,
  Levels = NULL,
  Threshold_type_tissueMask = NULL,
  Threshold_value_tissueMask = NULL,
  Blurr_tissueMask = NULL,
  Sigma_tissueMask = NULL,
  Blurr_target = NULL,
  Sigma_target = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Directory}{Character string specifying the path to the folder where images are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{Channels_to_keep}{Character vector indicating channels used in tissue mask generation.}

\item{Target_channel}{Character value indicating the target channel to be thresholded.}

\item{Save_processed_images}{Logical value indicating if thresholded images should be written (see details).}

\item{Local_thresholding}{A logical value indicating if local (per image) threshold should be calculated (see details).}

\item{Threshold_type}{Type of threshold should be one of the following: 'Arbitrary', 'Otsu' or 'Multilevel' (see details).}

\item{Threshold_value}{Numeric value to be used as threshold cut-off point. Must be provided for Arbitrary and CAN be provided for Multilevel (as a numeric vector).}

\item{Levels}{A integer indicating the number of desired levels to be calculated for multilevel thresholding. Only used if Threshold_value is NULL.}

\item{Threshold_type_tissueMask}{Type of threshold to performe tissue mask. Either 'Otsu', 'Arbitrary' or 'Absolute'.}

\item{Threshold_value_tissueMask}{Numeric value used if Arbitrary is the threshold type of choice}

\item{Blurr_tissueMask}{Logical value indicating if image blurring be performed before tissue mask generation}

\item{Sigma_tissueMask}{Numeric value indicating the sigma value to perform Gaussian blurring}

\item{Blurr_target}{Logical value indicating if image blurring be performed before target thresholding}

\item{Sigma_target}{Numeric value indicating the sigma value to perform Gaussian blurring}

\item{Output_directory}{Character specifying the path to the folder where  output images are written. It must be an empty folder.}
}
\value{
Returns a tibble with the total foreground pixels and pixels above threshold per image
}
\description{
Generates binary or multi-level images based on pixel thresholding.
}
\details{
If processed images are saved, these can be further combined with cell position data using \code{\link[=DATA_Phenotypes_distance]{DATA_Phenotypes_distance()}}.
Local thresholding calculates a threshold value for every image in the experiment. If set to FALSE, global thresholding will be applied, and a unique threshold value will be calculated for all images.
Otsu threshold is calculated using the EBImage::otsu function. Automatic multilevel threshold is calculated using the imagerExtra::ThresholdML function.
}
\seealso{
\code{\link[=Image_thresholding_app_launcher]{Image_thresholding_app_launcher()}}, \code{\link[=Binary_threshold_image_combinator]{Binary_threshold_image_combinator()}}, \code{\link[=MFI_Experimet_Calculator]{MFI_Experimet_Calculator()}}
}
