% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_Tumor_Stroma_identifier.R
\name{Quick_Tumor_Stroma_identifier}
\alias{Quick_Tumor_Stroma_identifier}
\title{Calculates tumor and stromal compartment based on tiles}
\usage{
Quick_Tumor_Stroma_identifier(
  DATA_Phenotypes,
  Index_phenotype,
  Accuracy,
  Min_cell_no = 1,
  Image_preview = NULL,
  N_cores = 1
)
}
\arguments{
\item{DATA_Phenotypes}{A dataframe or tibble containing a column named 'Phenotype' containing cell phenotype labels.}

\item{Index_phenotype}{A character value indicating the cell phenotype to be used to calculate the tissue compartment.}

\item{Accuracy}{A numeric value indicating the size of the tile. Smaller sizes give more accurate result.}

\item{Min_cell_no}{A integer value indicating the minimum number of cells within a tile to consider the tile positive for a compartment.}

\item{Image_preview}{A character value indicating the name of the image to be used in the preview. If Null a random sample will be selected.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}
}
\value{
Returns a tibble with cell features and a column named 'Compartment' containing cell location.
}
\description{
The function calculates tissue compartments (usually tumor and stroma) based on tile occupation by tumor cells.
}
\examples{
\dontrun{
Quick_Tumor_Stroma_identifier(
    DATA_Phenotypes = CSM_Phenotypecell_test,
    Index_phenotype = "TUMOR",
    Accuracy = 80,
    Min_cell_no = 3,
    Image_preview = NULL,
    N_cores = 1
)
}

}
