% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReClustering_function.R
\name{ReClustering_function}
\alias{ReClustering_function}
\title{Performs sub-clustering of specific cell populations}
\usage{
ReClustering_function(
  DATA = NULL,
  DATA_aside = NULL,
  Phenotype_variable = NULL,
  Phenotype_to_recluster = NULL,
  Variables_to_recluster = NULL,
  Strategy = NULL,
  N_Clusters = NULL,
  Levels = NULL,
  Arbitrary_cutoff = NULL,
  Batch_size = NULL,
  N_initiations = NULL,
  Max_iterations = NULL,
  Quality_metric = NULL,
  Max_iterations_km = NULL,
  Max_iterations_em = NULL,
  GMM_Distance = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{DATA_aside}{(OPTIONAL) A dataframe or tibble containing cell feature data that can be used in the re-clustering process.}

\item{Phenotype_variable}{A character value indicating the name of the column containing cell phenotype labels.}

\item{Phenotype_to_recluster}{A character value indicating the name of cell phenotype label to be re-clustered.}

\item{Variables_to_recluster}{A character vector indicating the variables contained in DATA or DATA_aside to be used in the re-clustering process.}

\item{Strategy}{One of the following: "Arbitrary", "Multilevel", "SOM", "Batch_K_means", "GMM" (see details).}

\item{N_Clusters}{For SOM, Batched K-means and GMM, an integer indicating the number of clusters to generate.}

\item{Levels}{For Multilevel, an integer indicating the number of levels to divide each feature.}

\item{Arbitrary_cutoff}{For arbitrary, a numeric value or numeric vector indicating the cut-off values to divide each feature.}

\item{Batch_size}{If strategy is Batch_K_means: Number of cells to be included in each random batch.}

\item{N_initiations}{If strategy is Batch_K_means: Number of times the algorithm is going to be tried to find the best clustering result.}

\item{Max_iterations}{If strategy is Batch_K_means: Max number of iterations in each try.}

\item{Quality_metric}{If strategy is GMM:T he quality measure used to test the number of clusters ("AIC" or "BIC").}

\item{Max_iterations_km}{If strategy is GMM: Number of max iterations in the K means clustering performed.}

\item{Max_iterations_em}{If strategy is GMM: Number of max iterations in the Expectation Maximization algorithm.}

\item{GMM_Distance}{If strategy is GMM: Distance metric used in the model ("eucl_dist" or "maha_dist").}
}
\value{
A tibble containing cell features and the new column with labels after re-clusterization process.
}
\description{
After phenotyping, specific cell phenotypes can be further subclassified according to previously un-used features. Using this function users may cluster cells in a step wise manner.
}
\details{
Multi_level thresholds are calculated using imagerExtra::ThresholdML function.

Self Organizing Maps clustering is performed using the FlowSOM::FlowSOM function.

Batch K-means and Gaussian Mixture Models are all based on the ClusterR package.
}
