% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentator_tester_app.R
\name{Segmentator_tester_app}
\alias{Segmentator_tester_app}
\title{Launches a shiny APP to explore cell segmentation parameters}
\usage{
Segmentator_tester_app(Directory = NULL, Ordered_Channels = NULL)
}
\arguments{
\item{Directory}{Character specifying the path to the folder where images to be segmented are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}
}
\description{
\code{Segmentator_tester_app()} launches an APP to interactively explore cell segmentation parameters interactively
Parameters can then be used to feed the \code{\link[=Cell_segmentator_quantificator]{Cell_segmentator_quantificator()}}.
}
