% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentator_tester_app.R
\name{Segmentator_tester_app}
\alias{Segmentator_tester_app}
\title{Launches a shiny APP to explore cell segmentation parameters}
\usage{
Segmentator_tester_app(Directory, Ordered_Channels)
}
\arguments{
\item{Directory}{Character specifying the path to the folder where images to be segmented are stored.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}
}
\description{
\code{Segmentator_tester_app()} launches an APP to interactively explore cell segmentation parameters. Parameters can then be used to feed the \code{\link[=Cell_segmentator_quantificator]{Cell_segmentator_quantificator()}} function.
}
\details{
Control panel controls the image display settings. If Pre process is active (default) image pre-processing will be applied to the nuclear channels before running segmentation (this can enhance performance in some scenarios).

Relevant buttons:
\itemize{
\item{GO!: Runs the cell segmentation and feature extraction algorithm with the current setting using the current image.}
\item{Download Parameters: Saves parameters in the Global environment.}
}

The lower panel of the control panel shows the current active parameters.

4 main panels:
\itemize{
\item{Upper left: Displays the image (use it to zoom in and out).}
\item{Upper right: Cell boundaries are displayed.}
\item{Lower left: Nucleus are depicted}
\item{Lower right: Heatmap of marker expression by cell. If image is zoomed in, only cells in the field of view are shown.}
}
}
\examples{
\dontrun{
#Create temporary input directory----------------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#Launch the app------------------------------------------------------------
Segmentator_tester_app(
    Directory = Input_Dir,
    Ordered_Channels = c("DAPI", "PDL1", "GZMB", "PD1", "CK-EPCAM", "CD8a", "FOXP3")
    )

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
}
}
\seealso{
\code{\link[=Cell_segmentator_quantificator]{Cell_segmentator_quantificator()}}
}
