% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smart_image_importer.R
\name{Smart_image_importer}
\alias{Smart_image_importer}
\title{Imports large images to the environment}
\usage{
Smart_image_importer(
  N_cores = 1,
  Image_directory,
  Log10_pixel_output = 6,
  X_crop_coordinates = NULL,
  Y_crop_coordinates = NULL,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL,
  Save_processed_images = FALSE,
  Output_Directory = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Image_directory}{Character specifying the path to the image file that needs to be imported.}

\item{Log10_pixel_output}{A single numeric value larger than 1. It indicates the size (in pixels) of the processed image. For example, if set to 6 (the default) the output image will have a size of 10^6 pixels.}

\item{X_crop_coordinates}{(OPTIONAL) If only a subset of the image needs to be imported, a numeric vector of length = 2, indicating the min and max X coordinates to perform image subsetting.}

\item{Y_crop_coordinates}{(OPTIONAL) If only a subset of the image needs to be imported, a numeric vector of length = 2, indicating the min and max Y coordinates to perform image subsetting.}

\item{Ordered_Channels}{(OPTIONAL) If image has multiple channel, the user can provide channel names. These names will be used with Channels_to_keep (see below) to filter the resulting image.}

\item{Channels_to_keep}{(OPTIONAL) If image has multiple channel, the user can provide channel names. These names will be used with Channels_to_keep to filter the resulting image.}

\item{Save_processed_images}{A logical value indicating if the resulting image should be written in disk.}

\item{Output_Directory}{If Save_processed_images is TRUE, the path to the folder where the resulting image should be written.}
}
\value{
A list containing the resulting image, the native image pixel dimensions, the processed image pixel dimensions and the subsetting coordinates (if required).
}
\description{
The function is aimed to import large images or image subsets to the environment for graphication purposes (generate images that can be easily plotted).
}
\details{
The function is able to read tif files. It can also work with pyramidal tif images. The \code{Smart_image_importer()} is designed to import any type of image irrespective of it's size. In order to
work with large images, the user needs to set a reasonable pixel size limit. Large tif images will be imported in tiles, downsized and then reconstructed to give rise to the output image. For pyramidal
tif files the resolution layer demonstrating the closest pixel size to the desired pixel-limit will be returned. If images are not exceedingly large (less than 1.5Gb), they will be imported and directly
downsized.

Tile importing is powered by the RBioFormats Bioconductor package that parses the Bioformats Java library to R. Sometimes even if the user#' has installed the RBioFormats package from
bioconductor the function will still return an error. If this happens, the issue can be solved checking the installed Java version in the computer and manually installing the CRAN rJava package.
}
\examples{
\dontrun{
#Create temporary input directory------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
function(Image){
   EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
   file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
})

#reduce the pixel content of the image
Smart_image_importer(Image_directory = dir(Input_Dir, full.names = TRUE)[1], 
                     Log10_pixel_output = 5
) 

}
}
