% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT_Heterogeneity_Analyzer.R
\name{SPIAT_Heterogeneity_Analyzer}
\alias{SPIAT_Heterogeneity_Analyzer}
\title{Analyzes the spatial heterogeneity following the pipeline in the SPIAT package}
\usage{
SPIAT_Heterogeneity_Analyzer(
  N_cores = NULL,
  DATA_SPIAT = NULL,
  DATA_Phenotypes = NULL,
  Tile_size = NULL,
  Phenotypes_included = NULL,
  Autocorrelation_metric = NULL,
  Entropy_threshold = NULL
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA_SPIAT}{A list containing SPIAT objects generated using \code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.}

\item{DATA_Phenotypes}{A dataframe or tibble containing cell feature data with a column named 'Phenotype' containing cell phenotype labels.}

\item{Tile_size}{A numeric value indicating the size of the tile.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the analysis.}

\item{Autocorrelation_metric}{Either "globalmoran" or "GearyC".}

\item{Entropy_threshold}{A numeric value indicating the the arbitrary threshold to be used.}
}
\value{
A tibble containing a summary of the analysis result by image.
}
\description{
The function divides the images into tiles, calculates entropy within tile and then counts the number of tiles above a user defined threshold.
Results may be similar to those obtained using \code{\link[=Tiled_image_heterogeneity_analyzer]{Tiled_image_heterogeneity_analyzer()}}.
}
\examples{
\dontrun{
#Generate SPIAT object list----------------------------
DATA_SPIAT <-
SPIAT_object_generator(
    DATA_Intensities = CSM_Arrangedcellfeaturedata_test,
    DATA_Phenotypes = CSM_Phenotypecell_test
)

#Analyze heterogeneity--------------------------------
SPIAT_Heterogeneity_Analyzer(
    N_cores = 2,
    DATA_SPIAT = DATA_SPIAT,
    DATA_Phenotypes = CSM_Phenotypecell_test,
    Tile_size = 125,
    Phenotypes_included = c("TUMOR", "CD8_GZMBneg", "CD8_GZMBpos", "OTHER"),
    Entropy_threshold = 0.5,
    Autocorrelation_metric = "globalmoran"
)
}


}
\seealso{
\code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}, \code{\link[=Tiled_image_heterogeneity_analyzer]{Tiled_image_heterogeneity_analyzer()}}.
}
