% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT_neighborhood_identifier.R
\name{SPIAT_neighborhood_identifier}
\alias{SPIAT_neighborhood_identifier}
\title{Identifies cell neighborhoods using the pipeline present in the SPIAT package}
\usage{
SPIAT_neighborhood_identifier(
  N_cores = 1,
  DATA_SPIAT,
  Strategy,
  Cell_types_of_interest,
  Radius,
  Min_neighborhood_size,
  K_phenograph = NULL,
  No_Phenotype_name
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{DATA_SPIAT}{A list containing SPIAT objects generated using \code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.}

\item{Strategy}{A character value indicating the method to find the clusters. One of "hierarchical" or "dbscan".}

\item{Cell_types_of_interest}{A character vector indicating the cell phenotype labels included in the analysis.}

\item{Radius}{A numeric value indicating the distance radius used to calculate neighborhoods.}

\item{Min_neighborhood_size}{A integer value indicating the minimum number of cells to constitute a neighborhood.}

\item{K_phenograph}{Currently NULL by lack of SPIAT package usage. Will be employed when the phenograph strategy is available in SPIAT.}

\item{No_Phenotype_name}{(OPTIONAL) A character value indicating the cell phenotype label corresponding to non-phenotyped cells.}
}
\value{
A tibble containing a summary of the analysis result by image.
}
\description{
The function identifies cell neighborhoods using the SPIAT::identify_neighborhoods function.
}
\examples{
\dontrun{
#Generate SPIAT object list----------------------------
DATA_SPIAT <-
SPIAT_object_generator(
    DATA_Intensities = CSM_Arrangedcellfeaturedata_test,
    DATA_Phenotypes = CSM_Phenotypecell_test
)

#Find neighborhoods in each sample--------------------
SPIAT_neighborhood_identifier(
    N_cores = 2,
    DATA_SPIAT = DATA_SPIAT,
    Strategy = "hierarchical",
    Cell_types_of_interest = c("TUMOR", "CD8_GZMBneg", "CD8_GZMBpos", "OTHER"),
    Radius = 50,
    Min_neighborhood_size = 10,
    No_Phenotype_name = NULL
)
}

}
\seealso{
\code{\link[=SPIAT_object_generator]{SPIAT_object_generator()}}.
}
