% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_morphometry_analyzer.R
\name{Structure_morphometry_analyzer}
\alias{Structure_morphometry_analyzer}
\title{Generates by-sample summary metrics for structure morphometry analysis}
\usage{
Structure_morphometry_analyzer(DATA_Morphometry)
}
\arguments{
\item{DATA_Morphometry}{A list containing structure morphometry data (generated using the \code{\link[=Structure_morphometry_calculator]{Structure_morphometry_calculator()}}).}
}
\value{
A list containing a tibble with the summary results.
}
\description{
Given a list containing structure morphometry metrics (generated using the \code{\link[=Structure_morphometry_calculator]{Structure_morphometry_calculator()}}) the function will compute by-sample summary metrics.
}
\examples{
\dontrun{
#Divide cells into discrete tiles
Tiled_Images <-
Image_tiling_processing_function(
 N_cores = 1,
 DATA = CSM_Phenotypecell_test,
 Tile_width = 25,
 Tile_height = 25,
 Variables_to_keep = "Phenotype")

#Try to identify morphological structures for OTHER cells according to tile distribution
Structures <- 
Structure_morphometry_calculator(N_cores = 1,
                                
                                Tiled_Images = Tiled_Images,
                                Variable_name = "Phenotype",
                                Feature_to_analyze = "OTHER",
                                
                                Min_tile_number = 1, 
                                Minimum_cell_no_per_tile = 2)

#Analyze the results
Structure_morphometry_analyzer(DATA = Structures)

#Graph the results
Structure_morphometry_graph_maker(Tiled_Images = Tiled_Images,
                                  Variable_name = "Phenotype",
                                  DATA_Morphometry = Structures,
                                  Color_by = "Structure_ID",
                                  Image_name = "ABCM22001_B09_MiniCrop.tif")
}


}
\seealso{
\code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}, \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}, \code{\link[=Structure_morphometry_calculator]{Structure_morphometry_calculator()}}, \code{\link[=Structure_morphometry_graph_maker]{Structure_morphometry_graph_maker()}}
}
