% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_morphometry_calculator.R
\name{Structure_morphometry_calculator}
\alias{Structure_morphometry_calculator}
\title{Computes morphology metrics from tiled image cell data}
\usage{
Structure_morphometry_calculator(
  N_cores = 1,
  Tiled_Images,
  Variable_name,
  Feature_to_analyze,
  Min_tile_number = 4,
  Smooth_amount = 1e-07,
  Tolerance_value = 1e-07,
  Neighborhood_distance = 1,
  Minimum_cell_no_per_tile = 1
)
}
\arguments{
\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Tiled_Images}{A list containing tiled images obtained using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}} or the \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}.}

\item{Variable_name}{A character value indicating the column name containing the cell labels.}

\item{Feature_to_analyze}{A character value indicating the cell type or neighborhood type that will be morphologically analyzed}

\item{Min_tile_number}{A integer value specifying the minimum number of tiles any morphological structure must contain to be considered real.}

\item{Smooth_amount}{A numeric value specifying the amount of Gaussian smoothing to be applied to the tile information (higher values yield larger and more smoothened structures).}

\item{Tolerance_value}{A numeric value specifying the minimum height of the structure between its highest point (seed) and the point where it contacts another object (MAY BE NULL).}

\item{Neighborhood_distance}{A integer value indicating the radius of the neighborhood in pixels for the detection of neighboring objects. Higher value smooths out small objects.}

\item{Minimum_cell_no_per_tile}{(If Tiled_Images are generated using the \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}) A integer value indicating the minimum number of cells of interest a tile must contain to be considered positive (default is 1).}
}
\value{
A list containing a tibble with the structure morphology data and the structure mask.
}
\description{
Given a list containing cell data from tiled images (generated using the \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}} or the \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}) the function
identifies discrete structures and calculates a variety of morphology metrics (see details).
}
\details{
The function will identify discrete structures using the watershed algorithm implemented in the simpleSeg Bioconductor package. To this end, the image tiles will be assigned a
positive (1) or negative (0) value according to the presence or absence of the desired cell feature. Afterwards, these positive/negative tile pattern is transformed
into a pseudo-image and fed to the watershed algorithm to compute the structure mask image. This image is then analyzed to obtain the morphological features of the structures idenfified
using the cytomapper Bioconductor package.
}
\examples{
\dontrun{
#Divide cells into discrete tiles
Tiled_Images <-
Image_tiling_processing_function(
 N_cores = 1,
 DATA = CSM_Phenotypecell_test,
 Tile_width = 25,
 Tile_height = 25,
 Variables_to_keep = "Phenotype")

#Try to identify morphological structures for OTHER cells according to tile distribution
Structures <- 
Structure_morphometry_calculator(N_cores = 1,
                                
                                Tiled_Images = Tiled_Images,
                                Variable_name = "Phenotype",
                                Feature_to_analyze = "OTHER",
                                
                                Min_tile_number = 1, 
                                Smooth_amount = 0.01, 
                                Tolerance_value = 1, 
                                Neighborhood_distance = 1,
                                Minimum_cell_no_per_tile = 2)

#Analyze the results
Structure_morphometry_analyzer(DATA = Structures)

#Graph the results
Structure_morphometry_graph_maker(Tiled_Images = Tiled_Images,
                                  Variable_name = "Phenotype",
                                  DATA_Morphometry = Structures,
                                  Color_by = "Structure_ID",
                                  Image_name = "ABCM22001_B09_MiniCrop.tif")
}


}
\seealso{
\code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}, \code{\link[=Tiled_Image_Clustering_function]{Tiled_Image_Clustering_function()}}, \code{\link[=Structure_morphometry_analyzer]{Structure_morphometry_analyzer()}}, \code{\link[=Structure_morphometry_graph_maker]{Structure_morphometry_graph_maker()}}
}
