% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texture_features_calculator.R
\name{Texture_features_calculator}
\alias{Texture_features_calculator}
\title{Calculates the average texture features of a single cell type}
\usage{
Texture_features_calculator(Tiled_images, Phenotype_included)
}
\arguments{
\item{Tiled_images}{A list containing tiled images obtained using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}.}

\item{Phenotype_included}{A character value indicating which cell phenotype to analyze.}
}
\value{
A tibble containing averaged texture features by image.
}
\description{
The function calculates the average texture features of a single cell type for every image. Image must have been previously tiled using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}.
The tiled image is first transformed into a raster like object, then the grey level co-ocurrence matrix is calculated and texture features are obtained and averaged.
}
\details{
Metrics are calculated using the glcm R package.
}
\examples{
\dontrun{
#Divide cells into tiles-------------------------------------
Tiled_Images <-
Image_tiling_processing_function(
   N_cores = 2,
   DATA = CSM_Phenotypecell_test,
   Tile_width = 125,
   Tile_height = 125,
   Variables_to_keep = "Phenotype"
)

#Calculate average texture features for a given cell phenotype---
Texture_features_calculator(
    Tiled_images = Tiled_Images,
    Phenotype_included = "CD8_GZMBneg"
)
}

}
\seealso{
\code{\link[=Tiled_images_graphicator]{Tiled_images_graphicator()}}
}
