% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_exploration_function.R
\name{Thresholding_exploration_function}
\alias{Thresholding_exploration_function}
\title{Summarizes thresholded data by image}
\usage{
Thresholding_exploration_function(
  DATA = NULL,
  Calculate_Density = NULL,
  DATA_Area = NULL
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data that has been thresholded.}

\item{Calculate_Density}{A logical value indicating if density should be calculated.}

\item{DATA_Area}{If Calculate_Density is TRUE, a dataframe or tibble containing image names and tissue area information.}
}
\value{
Returns a tibble with cell counts, pertentages and optionally cell densities per image.
}
\description{
The function generates positive cell counts, percentages and optionally cell densities according to thresholded feature data.
}
\examples{
#Calculate tissue area-------------------
DATA_AREA <- Image_size_calculator(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Strategy = "Tiling",
  Image_to_plot = NULL,
  Tile_accuracy = 60
)

#Threshold data---------------------------
DATA_thresholded <- Thresholding_function(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Strategy = "EBI_Otsu",
  Local_thresholding = FALSE,
  Method_autothreshold = "Otsu",
  number_iterations_TriClass = 20,
  Percentile = 0.5,
  Defined_threshold = 0.1,
  Levels = 3
)

#Generate the summary tibble-------------
Thresholding_exploration_function(
  DATA = DATA_thresholded,
  Calculate_Density = TRUE,
  DATA_Area = DATA_AREA
)


}
