% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_graphicator_function.R
\name{Thresholding_graphicator_function}
\alias{Thresholding_graphicator_function}
\title{Generates thresholding results summary plots}
\usage{
Thresholding_graphicator_function(
  DATA,
  DATA_thresholded,
  Marker_names,
  Image_name
)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{DATA_thresholded}{A dataframe or tibble containing cell feature data that has been thresholded.}

\item{Marker_names}{A character vector indicating features that need to be plotted.}

\item{Image_name}{A character value indicating the image to be plotted.}
}
\value{
Plots individual cells feature expression values and cells that are above threshold.
}
\description{
The function generates plots comparing feature expression values and thresholding results for individual images.
}
\examples{
\dontrun{
#Threshold data
DATA_thresholded <- Thresholding_function(
  DATA = CSM_Arrangedcellfeaturedata_test,
  Strategy = "EBI_Otsu",
  Local_thresholding = FALSE,
  Method_autothreshold = "Otsu",
  number_iterations_TriClass = 20,
  Percentile = 0.5,
  Defined_threshold = 0.1,
  Levels = 3
)

#Plot results
Thresholding_graphicator_function(
  DATA = CSM_Arrangedcellfeaturedata_test,
  DATA_thresholded = DATA_thresholded,
  Marker_names = names(DATA_thresholded)[-c(1:4)],
  Image_name = "ABCM22001_B09_MiniCrop.tif"
)
}


}
