% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_summary_function.R
\name{Thresholding_summary_function}
\alias{Thresholding_summary_function}
\title{Calculates cut-off points of thresholded data and generates summary plots}
\usage{
Thresholding_summary_function(DATA = NULL, DATA_thresholded = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{DATA_thresholded}{A dataframe or tibble containing cell feature data that has been thresholded.}
}
\value{
Returns a tibble with cut-off values and generates summary plots.
}
\description{
Given the original and thresholded dataset, the function calculates cut-off values and generates summary graphs.
}
