% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholding_tester_app.R
\name{Thresholding_tester_app}
\alias{Thresholding_tester_app}
\title{Launches a shiny APP to explore feature thresholding methods}
\usage{
Thresholding_tester_app(DATA = NULL, Directory = NULL, Ordered_Channels = NULL)
}
\arguments{
\item{DATA}{A dataframe or tibble containing cell feature data.}

\item{Directory}{Character specifying the path to the folder where images are present.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}
}
\description{
\code{Thresholding_tester_app()} launches an APP to interactively explore thresholding methods interactively. Parameters can then be used in the \code{\link[=Thresholding_function]{Thresholding_function()}} or \code{\link[=Thresholding_function_tailored]{Thresholding_function_tailored()}} functions.
}
\seealso{
\code{\link[=Thresholding_function]{Thresholding_function()}}, \code{\link[=Thresholding_function_tailored]{Thresholding_function_tailored()}}
}
