% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile_to_image_cell_arrange_function.R
\name{Tile_to_image_cell_arrange_function}
\alias{Tile_to_image_cell_arrange_function}
\title{Arrange cell coordinates according to position in whole using tile information}
\usage{
Tile_to_image_cell_arrange_function(DATA = NULL, Dist_to_edge = NULL)
}
\arguments{
\item{DATA}{Dataframe or tibble with cell features. This data should have been processed at some point using \code{\link[=DATA_arrange_function]{DATA_arrange_function()}}}

\item{Dist_to_edge}{For overlapping tiles, cells closer to tile edge than Dist_to_edge will be removed}
}
\value{
A tibble containing cell features with modified X and Y coordinates
}
\description{
\code{Tile_to_image_cell_arrange_function()} modifies cell X and Y coordinates to reflect cell position in the entire image.
}
\seealso{
\code{\link[=Image_tile_deconstruction_function]{Image_tile_deconstruction_function()}}, \code{\link[=Image_from_tile_rebuilder]{Image_from_tile_rebuilder()}}
}
