% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiled_image_heterogeneity_analyzer.R
\name{Tiled_image_heterogeneity_analyzer}
\alias{Tiled_image_heterogeneity_analyzer}
\title{Generates a summary of the heterogeneity by tile.}
\usage{
Tiled_image_heterogeneity_analyzer(
  Tiled_heterogeneity_DATA = NULL,
  Strategy = NULL,
  Metric = NULL,
  Threshold = NULL
)
}
\arguments{
\item{Tiled_heterogeneity_DATA}{A list containing tiled images obtained using \code{\link[=Tiled_image_heterogeneity_calculator]{Tiled_image_heterogeneity_calculator()}}.}

\item{Strategy}{A character value indicating the analytical approach. Either 'Quantify_by_Threshold' or 'Overall_Summary' (see details).}

\item{Metric}{A character value indicating the heterogeneity metric to be plotted.}

\item{Threshold}{If Strategy is Quantify_by_Threshold, the arbitrary threshold to be used.}
}
\value{
A tibble containing the summary of the analysis
}
\description{
Generates a summary of the heterogeneity by tile.
}
\details{
Quantify_by_Threshold counts the number of tiles above the arbitrary threshold and spatial autocorrelation (Moran's I) of the positive tiles.

Overall_summary quantifies the central tendency and dispersion metrics of the tile heterogeneity metrics and their spatial autocorrelation.
}
