% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiled_images_graphicator.R
\name{Tiled_images_graphicator}
\alias{Tiled_images_graphicator}
\title{Generates plots of cell counts by tile}
\usage{
Tiled_images_graphicator(
  Tiled_images = NULL,
  Image_name = NULL,
  Phenotypes_included = NULL
)
}
\arguments{
\item{Tiled_images}{A list containing tiled images obtained using \code{\link[=Image_tiling_processing_function]{Image_tiling_processing_function()}}.}

\item{Image_name}{A character value indicating the image to be plotted.}

\item{Phenotypes_included}{A character vector indicating the phenotype labels that will be included in the plots.}
}
\value{
A list containing plots (one element per phenotype).
}
\description{
The function generates plots of the number of target cells present within each tile for a given image.
}
\examples{
\dontrun{
#Divide cells into tiles---------
Tiled_Images <-
Image_tiling_processing_function(
   N_cores = 2,
   DATA = CSM_Phenotypecell_test,
   Tile_width = 125,
   Tile_height = 125,
   Variables_to_keep = "Phenotype"
)

#Graph target cells
Tiled_images_graphicator(
   Tiled_images = Tiled_Images,
   Image_name = "ABCM22001_B14_MiniCrop.tif",
   Phenotypes_included = c("TUMOR", "CD8_GZMBneg", "CD8_GZMBpos", "OTHER")
)
}


}
