% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiled_neighborhoods_graphicator.R
\name{Tiled_neighborhoods_graphicator}
\alias{Tiled_neighborhoods_graphicator}
\title{Generates a plot illustrating neighborhoods by tile}
\usage{
Tiled_neighborhoods_graphicator(
  DATA_elections,
  Image_name,
  Graph_only_winner_neighborhood = FALSE
)
}
\arguments{
\item{DATA_elections}{An object created using the \code{\link[=Neighborhood_voting_function]{Neighborhood_voting_function()}} function containing neighborhood election results.}

\item{Image_name}{A character value indicating the image to be plotted.}

\item{Graph_only_winner_neighborhood}{A logical value indicating if only the winner neighborhood by tile should be plotted.}
}
\value{
Returns a plot with the tiles and the neighborhood elections results.
}
\description{
Given a dataset of neighborhoods by tile generated using the \code{\link[=Neighborhood_voting_function]{Neighborhood_voting_function()}} function, a graph of the
elections results will be generated.
}
\examples{
\dontrun{
#Tile images with neighborhood information-----------------------------------
Tiled_Images <-
 Image_tiling_processing_function(
    DATA = CSM_Neighborhoods_test,
    Tile_width = 125,
    Tile_height = 125,
    Variables_to_keep = "Neighborhood_assignment",
    N_cores = 1
)

#Celebrate elections in each tile---------------------------------------------
DATA_neighborhood_elections <-
Neighborhood_voting_function(
    N_cores = 1,
    Tiled_Images = Tiled_Images,
    Minimum_cell_no_per_tile = 2,
    Neighborhoods_included = unique(CSM_Neighborhoods_test$Neighborhood_assignment)
)

#Graph results----------------------------------------------------------------
Tiled_neighborhoods_graphicator(
DATA_elections = DATA_neighborhood_elections,
Image_name = "ABCM22001_B14_MiniCrop.tif",
Graph_only_winner_neighborhood = FALSE
)
}

}
