% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tissue_mask_generator.R
\name{Tissue_mask_generator}
\alias{Tissue_mask_generator}
\title{Generates a tissue mask based on all the channels of an image and thresholding parameters}
\usage{
Tissue_mask_generator(
  Image = NULL,
  Threshold_type = NULL,
  Threshold_value = NULL,
  Blurr = NULL,
  Sigma = NULL
)
}
\arguments{
\item{Image}{An EBImage type object}

\item{Threshold_type}{Type of threshold method to be performed: 'Absolute', 'Otsu' or 'Arbitrary'}

\item{Threshold_value}{If Threshold_type = 'Arbitrary', the value applied to threshold the image}

\item{Blurr}{A logical value indicating if blurring should be applied before thresholding}

\item{Sigma}{Sigma applied to the Gaussian kernel for blurring}
}
\value{
An EBImage object containing the binary tissue mask of foreground pixels
}
\description{
Internal use only
}
\details{
Used in \code{\link[=Image_thresholding_app_launcher]{Image_thresholding_app_launcher()}}, \code{\link[=Pixel_Threshold_calculator]{Pixel_Threshold_calculator()}}, \code{\link[=MFI_Experimet_Calculator]{MFI_Experimet_Calculator()}}
}
\keyword{Internal}
