% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMA_dearrayer_function.R
\name{TMA_dearrayer_function}
\alias{TMA_dearrayer_function}
\title{Generates individual core images from TMA whole slide images}
\usage{
TMA_dearrayer_function(
  Original_Image_Directory,
  Downsized_Image,
  Output_Directory,
  RGB_Color_images = FALSE,
  N_cores = 1,
  Label_text_size = 7,
  Ordered_Channels = NULL,
  Channels_to_keep = NULL,
  Threshold_type_tissueMask = "Arbitrary",
  Threshold_value_tissueMask = 0.1,
  Blurr_tissueMask = TRUE,
  Sigma_tissueMask = 5,
  Erode_kernel_size = 3,
  Erode_rounds = 0,
  Dilate_kernel_size = 3,
  Dilate_rounds = 0,
  Min_size = 100,
  Watershed_tolerance = 1,
  Watershed_ext = 1,
  Tolerance_column = 100,
  Tolerance_row = 100,
  Core_names_matrix = NULL
)
}
\arguments{
\item{Original_Image_Directory}{A character value containing the path to the original TMA image.}

\item{Downsized_Image}{A list created using the \code{\link[=Smart_image_importer]{Smart_image_importer()}} function. It contains the downsized image representation of the TMA and the original pixel size.}

\item{Output_Directory}{A character value containing the path to the folder where core images will be saved.}

\item{RGB_Color_images}{A logical value indicating if TMA contains RGB color images (these will be processed in a different manner to grayscale multi-channel tif images).}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Label_text_size}{Numeric value indicating the size of the text of the labels (used to generate adequately sized images).}

\item{Ordered_Channels}{(OPTIONAL) If image has multiple channel, the user can provide channel names. These names will be used with Channels_to_keep (see below) to filter the resulting image.}

\item{Channels_to_keep}{(OPTIONAL) If image has multiple channel, the user can provide channel names. These names will be used with Channels_to_keep to filter the resulting image.}

\item{Threshold_type_tissueMask}{Type of threshold to performe tissue mask. Either 'Otsu', 'Arbitrary' or 'Absolute'.}

\item{Threshold_value_tissueMask}{Numeric value used if Arbitrary is the threshold type of choice.}

\item{Blurr_tissueMask}{Logical value indicating if image blurring be performed before tissue mask generation.}

\item{Sigma_tissueMask}{Numeric value indicating the sigma value to perform Gaussian blurring.}

\item{Erode_kernel_size}{Erosion kernel size. Used if Erode_rounds is >= 1.}

\item{Erode_rounds}{Number of erosion rounds to be performed.}

\item{Dilate_kernel_size}{Dilation kernel size. Used if Dilate_rounds is >= 1.}

\item{Dilate_rounds}{Number of dilation rounds to be performed.}

\item{Min_size}{A integer value indicating the minimum size (in pixels) a TMA core must have to be considered a valid core.}

\item{Watershed_tolerance}{A numeric value > 0 indicating the tolerance of the watershed algorithm. Higher values reduce the number of individual TMA cores identified.}

\item{Watershed_ext}{A numeric value indicating the extension watershed parameter. Higher values reduce the number of individual TMA cores identified.}

\item{Tolerance_column}{A numeric value indicating the tolerance from perfect alignment to consider a new column being identified (see details).}

\item{Tolerance_row}{A numeric value indicating the tolerance from perfect alignment to consider a new row being identified (see details).}

\item{Core_names_matrix}{(OPTIONAL BUT HIGHLY RECOMMENDED) A dataframe or matrix containing TMA core names. These names will be used to identify output core images.}
}
\value{
The function exports individual core images identified to the indicated output directory.
}
\description{
The function uses downsized representations of whole slide TMA images to find core position and extract them from the original image (see details)
}
\details{
The function works with downsized whole slide images generated using the \code{\link[=Smart_image_importer]{Smart_image_importer()}}. First a tissue mask is generated to identify the foreground and background
pixels. Afterwards this mask is processed (hole filling, erosion and finally dilation). Then the mask is transformed to a distmap of the distance of every pixel to the closest
background pixel. Afterwards, this distmap y segmented using a watershed algorithm to identify individual tumor cores. The bounding box of these cores is used to extract
the core images from the originial TMA whole slide image.

If the image provided is an RGB image, in order to generate the mask the image is first inverted.

After identifying the position of the TMA cores, the number of rows and columns present in the TMA is computed by clustering the X and Y centroid coordinates of the cores.
If the tolerance values are high, higher rates of missalignment will be tolerated to consider new rows/columns.

After identifying the position of TMA cores and the row and column they belong to, cores are labelled according to the user provided matrix.

If the TMA cores are large images (due to tissue size or image resolution), the function may hit the maximum RAM usage capacity of any JAVA-based functions. In order to avoid this,
the user may need to open a fresh R session and before running any other code run \code{options(java.parameters = "-Xmx200g")} or \code{rJava::.jinit(parameters="-Xmx200g")} to increase max RAM
use to an absurdly high memory use (200Gb in this case).
}
\examples{
\dontrun{
TMA_image_downsized <-
    Smart_image_importer(Image_directory = "Path/to/original/TMA/Image",
                         Log10_pixel_output = 6
  )

TMA_dearrayer_function(Original_Image_Directory = "Path/to/original/TMA/Image",
                      Downsized_Image = TMA_image_downsized,
                      Output_Directory = "Path/to/ouput/directory")

}
}
