% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watershed_cell_mask_generator.R
\name{Watershed_cell_mask_generator}
\alias{Watershed_cell_mask_generator}
\title{Generates cell segmentation mask using the watershed algorithm}
\usage{
Watershed_cell_mask_generator(
  Directory,
  Parameter_list = NULL,
  Ordered_Channels = NULL,
  N_cores = 1,
  Output_directory = NULL,
  Nuclear_marker = NULL,
  Cell_body_method = NULL,
  Min_pixel = NULL,
  Smooth_amount = NULL,
  Normalization = NULL,
  Watershed_type = NULL,
  Tolerance_value = NULL,
  Neighborhood_distance = NULL,
  Disc_size = NULL,
  Tissue_mask_markers = NULL,
  Perform_PCA = FALSE,
  Perform_nuclear_channel_processing = NULL,
  Black_level = NULL,
  White_level = NULL,
  Gamma_level = NULL,
  Equalize = NULL,
  Opening_kernel_size = NULL,
  Closing_kernel_size = NULL
)
}
\arguments{
\item{Directory}{Character specifying the path to the folder where images to be segmented are present.}

\item{Parameter_list}{List obtained using the \code{\link[=Segmentator_tester_app]{Segmentator_tester_app()}} containing segmentation parameters.}

\item{Ordered_Channels}{Character vector specifying image channels in their exact order.}

\item{N_cores}{Integer. Number of cores to parallelize your computation.}

\item{Output_directory}{A character value indicating the path to the empty folder where output images should be written.}

\item{Nuclear_marker}{(Used if Parameters_list is NULL) Character vector of markers corresponding to the nuclei.}

\item{Cell_body_method}{(Used if Parameters_list is NULL) Method of cytoplasm identification. Can be 'none', 'dilate' or 'discModel'.}

\item{Min_pixel}{(Used if Parameters_list is NULL) Integer value specifying the minimum pixels for an object to be recognized as a cell and not noise.}

\item{Smooth_amount}{(Used if Parameters_list is NULL) Numeric value specifying the amount of Gaussian smoothing to be applied to the image.}

\item{Normalization}{(Used if Parameters_list is NULL) Single value or vector specifying the transformations from "sqrt", "asinh", "norm99", "maxThresh" and "tissueMask.}

\item{Watershed_type}{(Used if Parameters_list is NULL) Method used to perform watersheding. Accepted values: "intensity", "distance" or "combine".}

\item{Tolerance_value}{(Used if Parameters_list is NULL) Numeric value specifying the minimum height of the object in the units of image intensity between its highest point (seed) and the point where it contacts another object (MAY BE NULL).}

\item{Neighborhood_distance}{(Used if Parameters_list is NULL) Radius of the neighborhood in pixels for the detection of neighboring objects. Higher value smooths out small objects.}

\item{Disc_size}{(Used if Parameters_list is NULL) The size of dilation around nuclei to create cell disc or capture cytoplasm.}

\item{Tissue_mask_markers}{(Used if Parameters_list is NULL) A vector specifying the channels to be used to create the tissue mask if specified in transformation.}

\item{Perform_PCA}{(Used if Parameters_list is NULL) A logical value specifying wether to run PCA on aggregated nucleus markers in order to detect the cellular nucclei.}

\item{Perform_nuclear_channel_processing}{(Used if Parameters_list is NULL) A logical value specifying if nuclear channel pre-processing should be performed.}

\item{Black_level}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Numeric indicating the \% below the max intensity level to be set to absolute black.}

\item{White_level}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Numeric indicating the \% above the max intensity level to be set to absolute white.}

\item{Gamma_level}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Numeric value between -3 and +3 to indicate channel gamma.}

\item{Equalize}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) A logical value specifying if channel should be equalized.}

\item{Opening_kernel_size}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Opening kernel size (set to 1 if no opening is required).}

\item{Closing_kernel_size}{(Used if Parameters_list is NULL)(Used if Perform_nuclear_channel_processing is TRUE) Closing kernel size (set to 1 if no closing is required).}
}
\value{
Writes the cell mask images in the output directory
}
\description{
Performs watershed cell segmentation using the simpleSeg Bioconductor R package.
Parameters can then be obtained using \code{\link[=Segmentator_tester_app]{Segmentator_tester_app()}}.
}
\examples{
\dontrun{
#Create temporary input directory----------------------------------------
Input_Dir <- tempfile(pattern = "tempdir1_Input")
dir.create(Input_Dir, recursive = TRUE)

Output_Dir <- tempfile(pattern = "tempdir2_Output")
dir.create(Output_Dir, recursive = TRUE)

#Save images in Input directory
purrr::map(1:2,
          function(Image){
            EBImage::writeImage(CSM_MiniMultiTiff_test[[Image]],
                                file.path(Input_Dir, names(CSM_MiniMultiTiff_test)[Image]))
         })

#Check a segmentation parameters list obtained using the dedicated function----------
print(CSM_SegmentParams_test)


#Generate the whole cell masks using watershed---------------------------------------
Watershed_cell_mask_generator(
 Directory = Input_Dir,
 Output_directory = Output_Dir,
 Parameter_list = CSM_SegmentParams_test
)

#Prepare parameters to obtain subcellular nuclear and cytoplasmic compartments---
Subcellular_compartment_parameters <- 
 list(Nuclear = list(Channel_name = "DAPI", 
                     Type = "Positive",
                     Threshold_type = "Arbitrary", 
                     Threshold_value = 0.05,
                     Blurr = TRUE, 
                     Sigma = 0.4),
      Cytoplasmic = list(Channel_name = "DAPI", 
                         Type = "Negative",
                         Threshold_type = "Arbitrary", 
                         Threshold_value = 0.05,
                         Blurr = TRUE, 
                         Sigma = 0.4)
 )

#Obtain the cell feature matrix--------------------------------------------
Cell_feature_extractor(Image_directory = Input_Dir,
                      Mask_directory = Output_Dir,
                      Ordered_Channels = CSM_SegmentParams_test[["Channels_to_keep"]],
                      Subcellular_compartment_parameter_list = Subcellular_compartment_parameters)

#Remove directories---------------------------------------------------------
unlink(Input_Dir, recursive = TRUE)
unlink(Output_Dir, recursive = TRUE)
}


}
\seealso{
\code{\link[=Segmentator_tester_app]{Segmentator_tester_app()}}
}
